new_dynarec: Code cleanup
