+++ src/device/dd/disk.c
@@ -19,6 +19,7 @@
 *   51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.          *
 * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
+#include <stdlib.h>
 #include "disk.h"
 
 #define M64P_CORE_PROTOTYPES 1
