+++ src/device/r4300/new_dynarec/arm/arm_cpu_features.c
@@ -34,11 +34,11 @@ static unsigned char check_arm_cpu_feature(const char* feature)
 {
     unsigned char status = 0;
     FILE *pFile = osal_file_open(procfile, "r");
-    
+
     if (pFile != NULL)
     {
         char line[1024];
-        
+
         while (fgets(line , sizeof(line) , pFile) != NULL)
         {
             if (strncmp(line, "Features\t: ", 11))
@@ -46,7 +46,7 @@ static unsigned char check_arm_cpu_feature(const char* feature)
 
             if (strstr(line + 11, feature) != NULL)
                 status = 1;
-            
+
             break;
         }
         fclose(pFile);
@@ -58,18 +58,18 @@ static unsigned char get_arm_cpu_implementer(void)
 {
     unsigned char implementer = 0;
     FILE *pFile = osal_file_open(procfile, "r");
-    
+
     if (pFile != NULL)
     {
         char line[1024];
-        
+
         while (fgets(line , sizeof(line) , pFile) != NULL)
         {
             if (strncmp(line, "CPU implementer\t: ", 18))
                 continue;
-            
+
             sscanf(line+18, "0x%02hhx", &implementer);
-            
+
             break;
         }
         fclose(pFile);
@@ -81,18 +81,18 @@ static unsigned short get_arm_cpu_part(void)
 {
     unsigned short part = 0;
     FILE *pFile = osal_file_open(procfile, "r");
-    
+
     if (pFile != NULL)
     {
         char line[1024];
-        
+
         while (fgets(line , sizeof(line) , pFile) != NULL)
         {
             if (strncmp(line, "CPU part\t: ", 11))
                 continue;
-            
+
             sscanf(line+11, "0x%03hx", &part);
-            
+
             break;
         }
         fclose(pFile);
@@ -115,7 +115,7 @@ void detect_arm_cpu_features(void)
     arm_cpu_features.VFPv4    = check_arm_cpu_feature("vfpv4");
     arm_cpu_features.IDIVa    = check_arm_cpu_feature("idiva");
     arm_cpu_features.IDIVt    = check_arm_cpu_feature("idivt");
-    
+
     // Qualcomm Krait supports IDIVa but it doesn't report it. Check for krait.
     if (get_arm_cpu_implementer() == 0x51 && get_arm_cpu_part() == 0x6F)
         arm_cpu_features.IDIVa = arm_cpu_features.IDIVt = 1;
