+++ src/device/r4300/new_dynarec/new_dynarec.c
@@ -61,10 +61,8 @@ void recomp_dbg_block(int addr);
 #elif NEW_DYNAREC == NEW_DYNAREC_ARM
 #include "arm/arm_cpu_features.h"
 #include "arm/assem_arm.h"
-#define EAX 0 /* ??? required for syscall_assemble and do_ccstub even in ARM mode */
 #elif  NEW_DYNAREC == NEW_DYNAREC_ARM64
 #include "arm64/assem_arm64.h"
-#define EAX 0 /* ??? required for syscall_assemble and do_ccstub even in ARM mode */
 #else
 #error Unsupported dynarec architecture
 #endif
@@ -134,7 +132,7 @@ void recomp_dbg_block(int addr);
 #define STORE 2   // Store
 #define LOADLR 3  // Unaligned load
 #define STORELR 4 // Unaligned store
-#define MOV 5     // Move 
+#define MOV 5     // Move
 #define ALU 6     // Arithmetic/logic
 #define MULTDIV 7 // Multiply/divide
 #define SHIFT 8   // Shift by register
@@ -227,40 +225,13 @@ void dyna_linker(void);
 void dyna_linker_ds(void);
 void breakpoint(void);
 
-/* interpreted opcode */
-static void TLBWI_new(int pcaddr, int count);
-static void TLBWR_new(int pcaddr, int count);
-static void MFC0_new(int copr, int count);
-static void MTC0_new(int copr, int count, int pcaddr);
-static void read_byte_new(int pcaddr, int count);
-static void read_hword_new(int pcaddr, int count);
-static void read_word_new(int pcaddr, int count);
-static void read_dword_new(int pcaddr, int count);
-static void write_byte_new(int pcaddr, int count);
-static void write_hword_new(int pcaddr, int count);
-static void write_word_new(int pcaddr, int count);
-static void write_dword_new(int pcaddr, int count);
-static void LWL_new(int pcaddr, int count);
-static void LWR_new(int pcaddr, int count);
-static void LDL_new(int pcaddr, int count);
-static void LDR_new(int pcaddr, int count);
-static void SWL_new(int pcaddr, int count);
-static void SWR_new(int pcaddr, int count);
-static void SDL_new(int pcaddr, int count);
-static void SDR_new(int pcaddr, int count);
-
 int new_recompile_block(int addr);
 void invalidate_block(u_int block);
 void *get_addr_ht(u_int vaddr);
+void *get_addr_32(u_int vaddr,u_int flags);
 
-static void wb_register(signed char r,signed char regmap[],uint64_t dirty,uint64_t is32);
-static void wb_dirtys(signed char i_regmap[],uint64_t i_is32,uint64_t i_dirty);
 static void load_regs_entry(int t);
-static void load_all_consts(signed char regmap[],int is32,u_int dirty,u_int isconst,int i);
-static void do_readstub(int n);
 static void inline_readstub(int type,int i,u_int addr_const,char addr,struct regstat *i_regs,int target,int adj,u_int reglist);
-static void do_writestub(int n);
-static void inline_writestub(int type,int i,u_int addr_const,char addr,struct regstat *i_regs,int target,int adj,u_int reglist);
 
 void *base_addr;
 void *base_addr_rx;
@@ -390,7 +361,7 @@ static void flush_dirty_uppers(struct regstat *cur)
   for (hr=0;hr<HOST_REGS;hr++) {
     if((cur->dirty>>hr)&1) {
       reg=cur->regmap[hr];
-      if(reg>=64) 
+      if(reg>=64)
         if((cur->is32>>(reg&63))&1) cur->regmap[hr]=-1;
     }
   }
@@ -547,7 +518,7 @@ static int needed_again(int r, int i)
   int j;
   /*int b=-1;*/
   int rn=10;
-  
+
   if(i>0&&(itype[i-1]==UJUMP||itype[i-1]==RJUMP||(source[i-1]>>16)==0x1000))
   {
     if(ba[i-1]<start || ba[i-1]>start+slen*4-4)
@@ -666,13 +637,13 @@ static void unneeded_registers(int istart,int iend,int r)
     {
       // If subroutine call, flag return address as a possible branch target
       if(rt1[i]==31 && i<slen-2) bt[i+2]=1;
-      
+
       if(ba[i]<start || ba[i]>=(start+slen*4))
       {
         // Branch out of this block, flush all regs
         u=1;
         uu=1;
-        /* Hexagon hack 
+        /* Hexagon hack
         if(itype[i]==UJUMP&&rt1[i]==31)
         {
           uu=u=0x300C00F; // Discard at, v0-v1, t6-t9
@@ -900,7 +871,7 @@ static void provisional_32bit(void)
   int i,j;
   uint64_t is32=1;
   uint64_t lastbranch=1;
-  
+
   for(i=0;i<slen;i++)
   {
     if(i>0) {
@@ -937,13 +908,13 @@ static void provisional_32bit(void)
       uint64_t temp_is32=is32;
       for(j=i-1;j>=0;j--)
       {
-        if(ba[j]==start+i*4) 
+        if(ba[j]==start+i*4)
           //temp_is32&=branch_regs[j].is32;
           temp_is32&=p32[j];
       }
       for(j=i;j<slen;j++)
       {
-        if(ba[j]==start+i*4) 
+        if(ba[j]==start+i*4)
           temp_is32=1;
       }
       is32=temp_is32;
@@ -1126,7 +1097,7 @@ static void provisional_r32(void)
 {
   u_int r32=0;
   int i;
-  
+
   for (i=slen-1;i>=0;i--)
   {
     int hr;
@@ -1224,7 +1195,7 @@ static void provisional_r32(void)
     }
     //requires_32bit[i]=r32;
     pr32[i]=r32;
-    
+
     // Dirty registers which are 32-bit, require 32-bit input
     // as they will be written as 32-bit values
     for(hr=0;hr<HOST_REGS;hr++)
@@ -1577,7 +1548,7 @@ static void clean_registers(int istart,int iend,int wr)
         if((regs[i].regmap[r]&63)==rt2[i]) wont_dirty_i|=1<<r;
         if(regs[i].regmap[r]==CCREG) wont_dirty_i|=1<<r;
         if(i>istart) {
-          if(itype[i]!=RJUMP&&itype[i]!=UJUMP&&itype[i]!=CJUMP&&itype[i]!=SJUMP&&itype[i]!=FJUMP) 
+          if(itype[i]!=RJUMP&&itype[i]!=UJUMP&&itype[i]!=CJUMP&&itype[i]!=SJUMP&&itype[i]!=FJUMP)
           {
             // Don't store a register immediately after writing it,
             // may prevent dual-issue.
@@ -1673,11 +1644,142 @@ static void clean_registers(int istart,int iend,int wr)
   }
 }
 
+// Is the branch target a valid internal jump?
+static int internal_branch(uint64_t i_is32,int addr)
+{
+  if(addr&1) return 0; // Indirect (register) jump
+  if((u_int)addr>=start && (u_int)addr<start+slen*4-4)
+  {
+    int t=(addr-start)>>2;
+    // Delay slots are not valid branch targets
+    //if(t>0&&(itype[t-1]==RJUMP||itype[t-1]==UJUMP||itype[t-1]==CJUMP||itype[t-1]==SJUMP||itype[t-1]==FJUMP)) return 0;
+    // 64 -> 32 bit transition requires a recompile
+    /*if(is32[t]&~unneeded_reg_upper[t]&~i_is32)
+    {
+      if(requires_32bit[t]&~i_is32) DebugMessage(M64MSG_VERBOSE, "optimizable: no");
+      else DebugMessage(M64MSG_VERBOSE, "optimizable: yes");
+    }*/
+    //if(is32[t]&~unneeded_reg_upper[t]&~i_is32) return 0;
+    if(requires_32bit[t]&~i_is32) return 0;
+    else return 1;
+  }
+  return 0;
+}
+
+static int get_final_value(int hr, int i, int *value)
+{
+  int reg=regs[i].regmap[hr];
+  while(i<slen-1) {
+    if(regs[i+1].regmap[hr]!=reg) break;
+    if(!((regs[i+1].isconst>>hr)&1)) break;
+    if(bt[i+1]) break;
+    i++;
+  }
+  if(i<slen-1) {
+    if(itype[i]==UJUMP||itype[i]==RJUMP||itype[i]==CJUMP||itype[i]==SJUMP) {
+      *value=constmap[i][hr];
+      return 1;
+    }
+  }
+  *value=constmap[i][hr];
+  //DebugMessage(M64MSG_VERBOSE, "c=%x",(int)constmap[i][hr]);
+  if(i==slen-1) return 1;
+  if(reg<64) {
+    return !((unneeded_reg[i+1]>>reg)&1);
+  }else{
+    return !((unneeded_reg_upper[i+1]>>reg)&1);
+  }
+}
+
+static int match_bt(signed char i_regmap[],uint64_t i_is32,uint64_t i_dirty,int addr)
+{
+  if((u_int)addr>=start && (u_int)addr<start+slen*4-4)
+  {
+    int t=(addr-start)>>2;
+    int hr;
+    if(regs[t].regmap_entry[HOST_CCREG]!=CCREG) return 0;
+    for(hr=0;hr<HOST_REGS;hr++)
+    {
+      if(hr!=EXCLUDE_REG)
+      {
+        if(i_regmap[hr]!=regs[t].regmap_entry[hr])
+        {
+          if(regs[t].regmap_entry[hr]>=0&&(regs[t].regmap_entry[hr]|64)<TEMPREG+64)
+          {
+            return 0;
+          }
+          else
+          if((i_dirty>>hr)&1)
+          {
+            if(i_regmap[hr]<TEMPREG)
+            {
+              if(!((unneeded_reg[t]>>i_regmap[hr])&1))
+                return 0;
+            }
+            else if(i_regmap[hr]>=64&&i_regmap[hr]<TEMPREG+64)
+            {
+              if(!((unneeded_reg_upper[t]>>(i_regmap[hr]&63))&1))
+                return 0;
+            }
+          }
+        }
+        else // Same register but is it 32-bit or dirty?
+        if(i_regmap[hr]>=0)
+        {
+          if(!((regs[t].dirty>>hr)&1))
+          {
+            if((i_dirty>>hr)&1)
+            {
+              if(!((unneeded_reg[t]>>i_regmap[hr])&1))
+              {
+                //DebugMessage(M64MSG_VERBOSE, "%x: dirty no match",addr);
+                return 0;
+              }
+            }
+          }
+          if((((regs[t].was32^i_is32)&~unneeded_reg_upper[t])>>(i_regmap[hr]&63))&1)
+          {
+            //DebugMessage(M64MSG_VERBOSE, "%x: is32 no match",addr);
+            return 0;
+          }
+        }
+      }
+    }
+    //if(is32[t]&~unneeded_reg_upper[t]&~i_is32) return 0;
+    if(requires_32bit[t]&~i_is32) return 0;
+    // Delay slots are not valid branch targets
+    //if(t>0&&(itype[t-1]==RJUMP||itype[t-1]==UJUMP||itype[t-1]==CJUMP||itype[t-1]==SJUMP||itype[t-1]==FJUMP)) return 0;
+    // Delay slots require additional processing, so do not match
+    if(is_ds[t]) return 0;
+  }
+  else
+  {
+    int hr;
+    for(hr=0;hr<HOST_REGS;hr++)
+    {
+      if(hr!=EXCLUDE_REG)
+      {
+        if(i_regmap[hr]>=0)
+        {
+          if(hr!=HOST_CCREG||i_regmap[hr]!=CCREG)
+          {
+            if((i_dirty>>hr)&1)
+            {
+              return 0;
+            }
+          }
+        }
+      }
+    }
+  }
+  return 1;
+}
+
 // Allocate every register, preserving source/target regs
 static void alloc_all(struct regstat *cur,int i)
 {
   int hr;
-  
+
   for(hr=0;hr<HOST_REGS;hr++) {
     if(hr!=EXCLUDE_REG) {
       if(((cur->regmap[hr]&63)!=rs1[i])&&((cur->regmap[hr]&63)!=rs2[i])&&
@@ -1701,7 +1803,7 @@ static void add_to_linker(intptr_t addr,u_int target,int ext)
   assert(linkcount<MAXBLOCK);
   link_addr[linkcount][0]=addr;
   link_addr[linkcount][1]=target;
-  link_addr[linkcount][2]=ext;  
+  link_addr[linkcount][2]=ext;
   linkcount++;
 }
 
@@ -1732,6 +1834,449 @@ static void remove_hash(u_int vaddr)
   }
 }
 
+/**** Interpreted opcodes ****/
+#define UPDATE_COUNT_IN \
+  struct r4300_core* r4300 = &g_dev.r4300; \
+  struct new_dynarec_hot_state* state = &r4300->new_dynarec_hot_state; \
+  state->cycle_count += count; \
+  state->pending_exception = 0;
+
+#define UPDATE_COUNT_OUT \
+  state->cycle_count -= !state->pending_exception * count;
+
+void* SYSCALL_new(void)
+{
+    struct r4300_core* r4300 = &g_dev.r4300;
+    struct new_dynarec_hot_state* state = &r4300->new_dynarec_hot_state;
+    r4300->delay_slot = 0;
+    cached_interp_SYSCALL();
+    return get_addr_ht(state->pcaddr);
+}
+
+void* ERET_new(void)
+{
+    struct r4300_core* r4300 = &g_dev.r4300;
+    struct new_dynarec_hot_state* state = &r4300->new_dynarec_hot_state;
+
+    cp0_update_count(r4300);
+    if (state->cp0_regs[CP0_STATUS_REG] & CP0_STATUS_ERL)
+    {
+        DebugMessage(M64MSG_ERROR, "error in ERET");
+        state->stop = 1;
+    }
+    else
+    {
+        state->cp0_regs[CP0_STATUS_REG] &= ~CP0_STATUS_EXL;
+        state->pcaddr = state->cp0_regs[CP0_EPC_REG];
+    }
+    r4300->llbit = 0;
+    r4300->delay_slot = 0;
+    r4300_check_interrupt(r4300, CP0_CAUSE_IP2, r4300->mi->regs[MI_INTR_REG] & r4300->mi->regs[MI_INTR_MASK_REG]); // ???
+    r4300->cp0.last_addr = state->pcaddr;
+    state->pending_exception = 0;
+    if (state->cycle_count >= 0) { gen_interrupt(r4300); }
+
+    if(state->stop)
+        return NULL;
+
+    if(state->pending_exception)
+        return get_addr_ht(state->pcaddr);
+    else
+    {
+        uint32_t is64 = 0;
+        for(int i=0;i<32;i++)
+            is64 = (((int)(state->regs[i]>>32)^((int)state->regs[i]>>31))!=0)<<i;
+
+        is64 |= (((int)(state->hi>>32)^((int)state->hi>>31))!=0);
+        is64 |= (((int)(state->lo>>32)^((int)state->lo>>31))!=0);
+        return get_addr_32(state->pcaddr, is64);
+    }
+}
+
+static void TLBWI_new(int pcaddr, int count)
+{
+  unsigned int i;
+  UPDATE_COUNT_IN
+  state->pcaddr = pcaddr;
+
+  /* Remove old entries */
+  unsigned int old_start_even=r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].start_even;
+  unsigned int old_end_even=r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].end_even;
+  unsigned int old_start_odd=r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].start_odd;
+  unsigned int old_end_odd=r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].end_odd;
+  for (i=old_start_even>>12; i<=old_end_even>>12; i++)
+  {
+    if(i<0x80000||i>0xBFFFF)
+    {
+      invalidate_block(i);
+      state->memory_map[i]=(uintptr_t)-1;
+    }
+  }
+  for (i=old_start_odd>>12; i<=old_end_odd>>12; i++)
+  {
+    if(i<0x80000||i>0xBFFFF)
+    {
+      invalidate_block(i);
+      state->memory_map[i]=(uintptr_t)-1;
+    }
+  }
+  cached_interp_TLBWI();
+  //DebugMessage(M64MSG_VERBOSE, "TLBWI: index=%d",state->cp0_regs[CP0_INDEX_REG]);
+  //DebugMessage(M64MSG_VERBOSE, "TLBWI: start_even=%x end_even=%x phys_even=%x v=%d d=%d",r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].start_even,r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].end_even,r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].phys_even,r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].v_even,r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].d_even);
+  //DebugMessage(M64MSG_VERBOSE, "TLBWI: start_odd=%x end_odd=%x phys_odd=%x v=%d d=%d",r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].start_odd,r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].end_odd,r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].phys_odd,r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].v_odd,r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].d_odd);
+  /* Combine r4300->cp0.tlb.LUT_r, r4300->cp0.tlb.LUT_w, and invalid_code into a single table
+     for fast look up. */
+  for (i=r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].start_even>>12; i<=r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].end_even>>12; i++)
+  {
+    //DebugMessage(M64MSG_VERBOSE, "%x: r:%8x w:%8x",i,r4300->cp0.tlb.LUT_r[i],r4300->cp0.tlb.LUT_w[i]);
+    if(i<0x80000||i>0xBFFFF)
+    {
+      if(r4300->cp0.tlb.LUT_r[i]) {
+        state->memory_map[i]=((uintptr_t)g_dev.rdram.dram+(uintptr_t)((r4300->cp0.tlb.LUT_r[i]&0xFFFFF000)-0x80000000)-(i<<12))>>2;
+        // FIXME: should make sure the physical page is invalid too
+        if(!r4300->cp0.tlb.LUT_w[i]||!r4300->cached_interp.invalid_code[i]) {
+          state->memory_map[i]|=WRITE_PROTECT; // Write protect
+        }else{
+          assert(r4300->cp0.tlb.LUT_r[i]==r4300->cp0.tlb.LUT_w[i]);
+        }
+        if(!using_tlb) DebugMessage(M64MSG_VERBOSE, "Enabled TLB");
+        // Tell the dynamic recompiler to generate tlb lookup code
+        using_tlb=1;
+      }
+      else state->memory_map[i]=(uintptr_t)-1;
+    }
+    //DebugMessage(M64MSG_VERBOSE, "memory_map[%x]: %8x (+%8x)",i,state->memory_map[i],state->memory_map[i]<<2);
+  }
+  for (i=r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].start_odd>>12; i<=r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].end_odd>>12; i++)
+  {
+    //DebugMessage(M64MSG_VERBOSE, "%x: r:%8x w:%8x",i,r4300->cp0.tlb.LUT_r[i],r4300->cp0.tlb.LUT_w[i]);
+    if(i<0x80000||i>0xBFFFF)
+    {
+      if(r4300->cp0.tlb.LUT_r[i]) {
+        state->memory_map[i]=((uintptr_t)g_dev.rdram.dram+(uintptr_t)((r4300->cp0.tlb.LUT_r[i]&0xFFFFF000)-0x80000000)-(i<<12))>>2;
+        // FIXME: should make sure the physical page is invalid too
+        if(!r4300->cp0.tlb.LUT_w[i]||!r4300->cached_interp.invalid_code[i]) {
+          state->memory_map[i]|=WRITE_PROTECT; // Write protect
+        }else{
+          assert(r4300->cp0.tlb.LUT_r[i]==r4300->cp0.tlb.LUT_w[i]);
+        }
+        if(!using_tlb) DebugMessage(M64MSG_VERBOSE, "Enabled TLB");
+        // Tell the dynamic recompiler to generate tlb lookup code
+        using_tlb=1;
+      }
+      else state->memory_map[i]=(uintptr_t)-1;
+    }
+    //DebugMessage(M64MSG_VERBOSE, "memory_map[%x]: %8x (+%8x)",i,state->memory_map[i],state->memory_map[i]<<2);
+  }
+  UPDATE_COUNT_OUT
+}
+
+static void TLBWR_new(int pcaddr, int count)
+{
+  unsigned int i;
+  UPDATE_COUNT_IN
+  state->pcaddr = pcaddr;
+  cp0_update_count(r4300);
+  state->cp0_regs[CP0_RANDOM_REG] = (state->cp0_regs[CP0_COUNT_REG]/r4300->cp0.count_per_op % (32 - state->cp0_regs[CP0_WIRED_REG])) + state->cp0_regs[CP0_WIRED_REG];
+  /* Remove old entries */
+  unsigned int old_start_even=r4300->cp0.tlb.entries[state->cp0_regs[CP0_RANDOM_REG]&0x3F].start_even;
+  unsigned int old_end_even=r4300->cp0.tlb.entries[state->cp0_regs[CP0_RANDOM_REG]&0x3F].end_even;
+  unsigned int old_start_odd=r4300->cp0.tlb.entries[state->cp0_regs[CP0_RANDOM_REG]&0x3F].start_odd;
+  unsigned int old_end_odd=r4300->cp0.tlb.entries[state->cp0_regs[CP0_RANDOM_REG]&0x3F].end_odd;
+  for (i=old_start_even>>12; i<=old_end_even>>12; i++)
+  {
+    if(i<0x80000||i>0xBFFFF)
+    {
+      invalidate_block(i);
+      state->memory_map[i]=(uintptr_t)-1;
+    }
+  }
+  for (i=old_start_odd>>12; i<=old_end_odd>>12; i++)
+  {
+    if(i<0x80000||i>0xBFFFF)
+    {
+      invalidate_block(i);
+      state->memory_map[i]=(uintptr_t)-1;
+    }
+  }
+  cached_interp_TLBWR();
+  /* Combine r4300->cp0.tlb.LUT_r, r4300->cp0.tlb.LUT_w, and invalid_code into a single table
+     for fast look up. */
+  for (i=r4300->cp0.tlb.entries[state->cp0_regs[CP0_RANDOM_REG]&0x3F].start_even>>12; i<=r4300->cp0.tlb.entries[state->cp0_regs[CP0_RANDOM_REG]&0x3F].end_even>>12; i++)
+  {
+    //DebugMessage(M64MSG_VERBOSE, "%x: r:%8x w:%8x",i,r4300->cp0.tlb.LUT_r[i],r4300->cp0.tlb.LUT_w[i]);
+    if(i<0x80000||i>0xBFFFF)
+    {
+      if(r4300->cp0.tlb.LUT_r[i]) {
+        state->memory_map[i]=((uintptr_t)g_dev.rdram.dram+(uintptr_t)((r4300->cp0.tlb.LUT_r[i]&0xFFFFF000)-0x80000000)-(i<<12))>>2;
+        // FIXME: should make sure the physical page is invalid too
+        if(!r4300->cp0.tlb.LUT_w[i]||!r4300->cached_interp.invalid_code[i]) {
+          state->memory_map[i]|=WRITE_PROTECT; // Write protect
+        }else{
+          assert(r4300->cp0.tlb.LUT_r[i]==r4300->cp0.tlb.LUT_w[i]);
+        }
+        if(!using_tlb) DebugMessage(M64MSG_VERBOSE, "Enabled TLB");
+        // Tell the dynamic recompiler to generate tlb lookup code
+        using_tlb=1;
+      }
+      else state->memory_map[i]=(uintptr_t)-1;
+    }
+    //DebugMessage(M64MSG_VERBOSE, "memory_map[%x]: %8x (+%8x)",i,state->memory_map[i],state->memory_map[i]<<2);
+  }
+  for (i=r4300->cp0.tlb.entries[state->cp0_regs[CP0_RANDOM_REG]&0x3F].start_odd>>12; i<=r4300->cp0.tlb.entries[state->cp0_regs[CP0_RANDOM_REG]&0x3F].end_odd>>12; i++)
+  {
+    //DebugMessage(M64MSG_VERBOSE, "%x: r:%8x w:%8x",i,r4300->cp0.tlb.LUT_r[i],r4300->cp0.tlb.LUT_w[i]);
+    if(i<0x80000||i>0xBFFFF)
+    {
+      if(r4300->cp0.tlb.LUT_r[i]) {
+        state->memory_map[i]=((uintptr_t)g_dev.rdram.dram+(uintptr_t)((r4300->cp0.tlb.LUT_r[i]&0xFFFFF000)-0x80000000)-(i<<12))>>2;
+        // FIXME: should make sure the physical page is invalid too
+        if(!r4300->cp0.tlb.LUT_w[i]||!r4300->cached_interp.invalid_code[i]) {
+          state->memory_map[i]|=WRITE_PROTECT; // Write protect
+        }else{
+          assert(r4300->cp0.tlb.LUT_r[i]==r4300->cp0.tlb.LUT_w[i]);
+        }
+        if(!using_tlb) DebugMessage(M64MSG_VERBOSE, "Enabled TLB");
+        // Tell the dynamic recompiler to generate tlb lookup code
+        using_tlb=1;
+      }
+      else state->memory_map[i]=(uintptr_t)-1;
+    }
+    //DebugMessage(M64MSG_VERBOSE, "memory_map[%x]: %8x (+%8x)",i,state->memory_map[i],state->memory_map[i]<<2);
+  }
+  UPDATE_COUNT_OUT
+}
+
+static void MFC0_new(int copr, int count)
+{
+  UPDATE_COUNT_IN
+  state->fake_pc.f.r.nrd = copr;
+  cached_interp_MFC0();
+  UPDATE_COUNT_OUT
+}
+
+static void MTC0_new(int copr, int count, int pcaddr)
+{
+  UPDATE_COUNT_IN
+  state->pcaddr = pcaddr;
+  r4300->delay_slot = 0;
+  state->fake_pc.f.r.nrd = copr;
+  cached_interp_MTC0();
+  UPDATE_COUNT_OUT
+}
+
+#define BITS_BELOW_MASK32(x) ((UINT32_C(1) << (x)) - 1)
+#define BITS_ABOVE_MASK32(x) (~(BITS_BELOW_MASK32((x))))
+
+#define BITS_BELOW_MASK64(x) ((UINT64_C(1) << (x)) - 1)
+#define BITS_ABOVE_MASK64(x) (~(BITS_BELOW_MASK64((x))))
+
+static unsigned int bshift(uint32_t address)
+{
+    return ((address & 3) ^ 3) << 3;
+}
+
+static unsigned int hshift(uint32_t address)
+{
+    return ((address & 2) ^ 2) << 3;
+}
+
+static void read_byte_new(int pcaddr, int count)
+{
+  uint32_t value;
+  UPDATE_COUNT_IN
+  state->pcaddr = pcaddr&~1;
+  r4300->delay_slot = pcaddr & 1;
+  unsigned int shift = bshift(state->address);
+  if (r4300_read_aligned_word(r4300, state->address, &value)) {
+    state->rdword = (uint64_t)((value >> shift) & 0xff);
+  }
+  UPDATE_COUNT_OUT
+}
+
+static void read_hword_new(int pcaddr, int count)
+{
+  uint32_t value;
+  UPDATE_COUNT_IN
+  state->pcaddr = pcaddr&~1;
+  r4300->delay_slot = pcaddr & 1;
+  unsigned int shift = hshift(state->address);
+  if (r4300_read_aligned_word(r4300, state->address, &value)) {
+    state->rdword = (uint64_t)((value >> shift) & 0xffff);
+  }
+  UPDATE_COUNT_OUT
+}
+
+static void read_word_new(int pcaddr, int count)
+{
+  uint32_t value;
+  UPDATE_COUNT_IN
+  state->pcaddr = pcaddr&~1;
+  r4300->delay_slot = pcaddr & 1;
+  if (r4300_read_aligned_word(r4300, state->address, &value)) {
+    state->rdword = (uint64_t)(value);
+  }
+  UPDATE_COUNT_OUT
+}
+
+static void read_dword_new(int pcaddr, int count)
+{
+  UPDATE_COUNT_IN
+  state->pcaddr = pcaddr&~1;
+  r4300->delay_slot = pcaddr & 1;
+  r4300_read_aligned_dword(r4300, state->address, (uint64_t*)&state->rdword);
+  UPDATE_COUNT_OUT
+}
+
+static void write_byte_new(int pcaddr, int count)
+{
+  UPDATE_COUNT_IN
+  state->pcaddr = pcaddr&~1;
+  r4300->delay_slot = pcaddr & 1;
+  unsigned int shift = bshift(state->address);
+  state->wword <<= shift;
+  r4300_write_aligned_word(r4300, state->address, state->wword, UINT32_C(0xff) << shift);
+  UPDATE_COUNT_OUT
+}
+
+static void write_hword_new(int pcaddr, int count)
+{
+  UPDATE_COUNT_IN
+  state->pcaddr = pcaddr&~1;
+  r4300->delay_slot = pcaddr & 1;
+  unsigned int shift = hshift(state->address);
+  state->wword <<= shift;
+  r4300_write_aligned_word(r4300, state->address, state->wword, UINT32_C(0xffff) << shift);
+  UPDATE_COUNT_OUT
+}
+
+static void write_word_new(int pcaddr, int count)
+{
+  UPDATE_COUNT_IN
+  state->pcaddr = pcaddr&~1;
+  r4300->delay_slot = pcaddr & 1;
+  r4300_write_aligned_word(r4300, state->address, state->wword, UINT32_C(0xffffffff));
+  UPDATE_COUNT_OUT
+}
+
+static void write_dword_new(int pcaddr, int count)
+{
+  UPDATE_COUNT_IN
+  state->pcaddr = pcaddr&~1;
+  r4300->delay_slot = pcaddr & 1;
+  /* NOTE: in dynarec, we only need an all-one mask */
+  r4300_write_aligned_dword(r4300, state->address, state->wdword, ~UINT64_C(0));
+  UPDATE_COUNT_OUT
+}
+
+static void LWL_new(int pcaddr, int count)
+{
+  UPDATE_COUNT_IN
+  state->pcaddr = pcaddr&~1;
+  r4300->delay_slot = pcaddr & 1;
+  uint32_t value;
+  unsigned int n = (state->address & 3);
+  unsigned int shift = 8 * n;
+  uint32_t mask = BITS_BELOW_MASK32(8 * n);
+  if (r4300_read_aligned_word(r4300, state->address & ~UINT32_C(0x3), &value)) {
+    state->rdword = (uint64_t)((state->wword & mask) | (value << shift));
+  }
+  UPDATE_COUNT_OUT
+}
+
+static void LWR_new(int pcaddr, int count)
+{
+  UPDATE_COUNT_IN
+  state->pcaddr = pcaddr&~1;
+  r4300->delay_slot = pcaddr & 1;
+  uint32_t value;
+  unsigned int n = (state->address & 3);
+  unsigned int shift = 8 * (3 - n);
+  uint32_t mask = (n == 3) ? UINT32_C(0) : BITS_ABOVE_MASK32(8 * (n + 1));
+  if (r4300_read_aligned_word(r4300, state->address & ~UINT32_C(0x3), &value)) {
+    state->rdword = (uint64_t)((state->wword & mask) | (value >> shift));
+  }
+  UPDATE_COUNT_OUT
+}
+
+static void LDL_new(int pcaddr, int count)
+{
+  UPDATE_COUNT_IN
+  state->pcaddr = pcaddr&~1;
+  r4300->delay_slot = pcaddr & 1;
+  uint64_t value;
+  unsigned int n = (state->address & 7);
+  unsigned int shift = 8 * n;
+  uint64_t mask = BITS_BELOW_MASK64(8 * n);
+  if(r4300_read_aligned_dword(r4300, state->address & ~UINT32_C(7), &value)) {
+      state->rdword = (state->wdword & mask) | (value << shift);
+  }
+  UPDATE_COUNT_OUT
+}
+
+static void LDR_new(int pcaddr, int count)
+{
+  UPDATE_COUNT_IN
+  state->pcaddr = pcaddr&~1;
+  r4300->delay_slot = pcaddr & 1;
+  uint64_t value;
+  unsigned int n = (state->address & 7);
+  unsigned int shift = 8 * (7 - n);
+  uint64_t mask = (n == 7) ? UINT64_C(0) : BITS_ABOVE_MASK64(8 * (n + 1));
+  if(r4300_read_aligned_dword(r4300, state->address & ~UINT32_C(7), &value)){
+      state->rdword = (state->wdword & mask) | (value >> shift);
+  }
+  UPDATE_COUNT_OUT
+}
+
+static void SWL_new(int pcaddr, int count)
+{
+  UPDATE_COUNT_IN
+  state->pcaddr = pcaddr&~1;
+  r4300->delay_slot = pcaddr & 1;
+  unsigned int n = (state->address & 3);
+  unsigned int shift = 8 * n;
+  uint32_t mask = (n == 0) ? ~UINT32_C(0) : BITS_BELOW_MASK32(8 * (4 - n));
+  r4300_write_aligned_word(r4300, state->address & ~UINT32_C(0x3), state->wword >> shift, mask);
+  UPDATE_COUNT_OUT
+}
+
+static void SWR_new(int pcaddr, int count)
+{
+  UPDATE_COUNT_IN
+  state->pcaddr = pcaddr&~1;
+  r4300->delay_slot = pcaddr & 1;
+  unsigned int n = (state->address & 3);
+  unsigned int shift = 8 * (3 - n);
+  uint32_t mask = BITS_ABOVE_MASK32(8 * (3 - n));
+  r4300_write_aligned_word(r4300, state->address & ~UINT32_C(0x3), state->wword << shift, mask);
+  UPDATE_COUNT_OUT
+}
+
+static void SDL_new(int pcaddr, int count)
+{
+  UPDATE_COUNT_IN
+  state->pcaddr = pcaddr&~1;
+  r4300->delay_slot = pcaddr & 1;
+  unsigned int n = (state->address & 7);
+  unsigned int shift = 8 * n;
+  uint64_t mask = (n == 0) ? ~UINT64_C(0) : BITS_BELOW_MASK64(8 * (8 - n));
+  r4300_write_aligned_dword(r4300, state->address & ~UINT32_C(0x7), state->wdword >> shift, mask);
+  UPDATE_COUNT_OUT
+}
+
+static void SDR_new(int pcaddr, int count)
+{
+  UPDATE_COUNT_IN
+  state->pcaddr = pcaddr&~1;
+  r4300->delay_slot = pcaddr & 1;
+  unsigned int n = (state->address & 7);
+  unsigned int shift = 8 * (7 - n);
+  uint64_t mask = BITS_ABOVE_MASK64(8 * (7 - n));
+  r4300_write_aligned_dword(r4300, state->address & ~UINT32_C(0x7), state->wdword << shift, mask);
+  UPDATE_COUNT_OUT
+}
+
 #if NEW_DYNAREC == NEW_DYNAREC_X86
 #include "x86/assem_x86.c"
 #elif NEW_DYNAREC == NEW_DYNAREC_X64
@@ -1751,18 +2296,18 @@ static void tlb_speed_hacks()
   {
     u_int addr;
     int n;
-    switch (ROM_HEADER.Country_code&0xFF) 
+    switch (ROM_HEADER.Country_code&0xFF)
     {
       case 0x45: // U
         addr=0x34b30;
-        break;                   
-      case 0x4A: // J 
-        addr=0x34b70;    
-        break;    
-      case 0x50: // E 
+        break;
+      case 0x4A: // J
+        addr=0x34b70;
+        break;
+      case 0x50: // E
         addr=0x329f0;
-        break;                        
-      default: 
+        break;
+      default:
         // Unknown country code
         addr=0;
         break;
@@ -1789,6 +2334,7 @@ static void tlb_speed_hacks()
   }
 }
 
+/**** Linker ****/
 u_int verify_dirty(struct ll_entry * head)
 {
   void *source;
@@ -1801,7 +2347,7 @@ u_int verify_dirty(struct ll_entry * head)
     unsigned int page=head->start>>12;
     uintptr_t map_value=g_dev.r4300.new_dynarec_hot_state.memory_map[page];
 
-    if((intptr_t)map_value<(intptr_t)0) 
+    if((intptr_t)map_value<(intptr_t)0)
       return head->vaddr;
 
     while(page<((head->start+head->length-1)>>12)) {
@@ -2262,6 +2808,7 @@ static void invalidate_page(u_int page)
     head=next;
   }
 }
+
 void invalidate_block(u_int block)
 {
   u_int page;
@@ -2320,7 +2867,7 @@ void invalidate_block(u_int block)
   #if NEW_DYNAREC >= NEW_DYNAREC_ARM
     do_clear_cache();
   #endif
-  
+
   // Don't trap writes
   g_dev.r4300.cached_interp.invalid_code[block]=1;
   // If there is a valid TLB entry for this page, remove write protect
@@ -2459,15 +3006,41 @@ void clean_blocks(u_int page)
   }
 }
 
-static void emit_extjump(intptr_t addr, int target)
+void* cop1_unusable(void)
 {
-  emit_extjump2(addr, target, (intptr_t)dyna_linker);
-}
-static void emit_extjump_ds(intptr_t addr, int target)
-{
-  emit_extjump2(addr, target, (intptr_t)dyna_linker_ds);
+    struct r4300_core* r4300 = &g_dev.r4300;
+    struct new_dynarec_hot_state* state = &r4300->new_dynarec_hot_state;
+    r4300->delay_slot = state->pcaddr & 1;
+    state->pcaddr &= ~1;
+    state->cp0_regs[CP0_CAUSE_REG] = CP0_CAUSE_EXCCODE_CPU | CP0_CAUSE_CE1;
+    exception_general(r4300);
+    return get_addr_ht(state->pcaddr);
 }
 
+void dynarec_gen_interrupt(void)
+{
+    struct r4300_core* r4300 = &g_dev.r4300;
+    struct new_dynarec_hot_state* state = &r4300->new_dynarec_hot_state;
+    cp0_update_count(r4300);
+    uint32_t page = ((state->cp0_regs[CP0_COUNT_REG]>>19)&0x1fc);
+    unsigned int *candidate = (unsigned int *)&restore_candidate[page];
+    page <<= 3;
+    r4300->delay_slot = 0;
+
+    if(*candidate)
+    {
+        for(int i=0;i<32;i++)
+        {
+            if((*candidate>>i)&1)
+                clean_blocks(page+i);
+        }
+        *candidate = 0;
+    }
+
+    gen_interrupt(r4300);
+}
+
+/**** Register allocation ****/
 static void mov_alloc(struct regstat *current,int i)
 {
   // Note: Don't need to actually alloc the source registers
@@ -2999,18 +3572,21 @@ static void cop1_alloc(struct regstat *current,int i)
   }
   minimum_free_regs[i]=1;
 }
+
 static void fconv_alloc(struct regstat *current,int i)
 {
   alloc_reg(current,i,CSREG); // Load status
   alloc_reg_temp(current,i,-1);
   minimum_free_regs[i]=1;
 }
+
 static void float_alloc(struct regstat *current,int i)
 {
   alloc_reg(current,i,CSREG); // Load status
   alloc_reg_temp(current,i,-1);
   minimum_free_regs[i]=1;
 }
+
 static void fcomp_alloc(struct regstat *current,int i)
 {
   alloc_reg(current,i,CSREG); // Load status
@@ -3142,6 +3718,170 @@ static void pagespan_alloc(struct regstat *current,int i)
   //else ...
 }
 
+/**** Assembler utils ****/
+static void wb_invalidate(signed char pre[],signed char entry[],uint64_t dirty,uint64_t is32,uint64_t u,uint64_t uu)
+{
+  int hr;
+  for(hr=0;hr<HOST_REGS;hr++) {
+    if(hr!=EXCLUDE_REG) {
+      if(pre[hr]!=entry[hr]) {
+        if(pre[hr]>=0) {
+          if((dirty>>hr)&1) {
+            if(get_reg(entry,pre[hr])<0) {
+              if(pre[hr]<64) {
+                if(!((u>>pre[hr])&1)) {
+                  emit_storereg(pre[hr],hr);
+                  if( ((is32>>pre[hr])&1) && !((uu>>pre[hr])&1) ) {
+                    emit_sarimm(hr,31,hr);
+                    emit_storereg(pre[hr]|64,hr);
+                  }
+                }
+              }else{
+                if(!((uu>>(pre[hr]&63))&1) && !((is32>>(pre[hr]&63))&1)) {
+                  emit_storereg(pre[hr],hr);
+                }
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+  // Move from one register to another (no writeback)
+  for(hr=0;hr<HOST_REGS;hr++) {
+    if(hr!=EXCLUDE_REG) {
+      if(pre[hr]!=entry[hr]) {
+        if(pre[hr]>=0&&(pre[hr]&63)<TEMPREG) {
+          int nr;
+          if((nr=get_reg(entry,pre[hr]))>=0) {
+            #ifdef NATIVE_64
+            if(pre[hr]>=INVCP) emit_mov64(hr,nr);
+            else
+            #endif
+            emit_mov(hr,nr);
+          }
+        }
+      }
+    }
+  }
+}
+
+// Write out all dirty registers (except cycle count)
+static void wb_dirtys(signed char i_regmap[],uint64_t i_is32,uint64_t i_dirty)
+{
+  int hr;
+  for(hr=0;hr<HOST_REGS;hr++) {
+    if(hr!=EXCLUDE_REG) {
+      if(((i_regmap[hr]&63)>0)&&((i_regmap[hr]&63)<CSREG)) {
+        if((i_dirty>>hr)&1) {
+          if(i_regmap[hr]<64) {
+            emit_storereg(i_regmap[hr],hr);
+            if(((i_is32>>i_regmap[hr])&1)) {
+              #ifdef DESTRUCTIVE_WRITEBACK
+              emit_sarimm(hr,31,hr);
+              emit_storereg(i_regmap[hr]|64,hr);
+              #else
+              emit_sarimm(hr,31,HOST_TEMPREG);
+              emit_storereg(i_regmap[hr]|64,HOST_TEMPREG);
+              #endif
+            }
+          }else{
+            if(!((i_is32>>(i_regmap[hr]&63))&1)) {
+              emit_storereg(i_regmap[hr],hr);
+            }
+          }
+        }
+      }
+    }
+  }
+}
+
+// Write out dirty registers that we need to reload (pair with load_needed_regs)
+// This writes the registers not written by store_regs_bt
+static void wb_needed_dirtys(signed char i_regmap[],uint64_t i_is32,uint64_t i_dirty,int addr)
+{
+  int hr;
+  int t=(addr-start)>>2;
+  for(hr=0;hr<HOST_REGS;hr++) {
+    if(hr!=EXCLUDE_REG) {
+      if(((i_regmap[hr]&63)>0)&&((i_regmap[hr]&63)<CSREG)) {
+        if(i_regmap[hr]==regs[t].regmap_entry[hr] && ((regs[t].dirty>>hr)&1) && !(((i_is32&~regs[t].was32&~unneeded_reg_upper[t])>>(i_regmap[hr]&63))&1)) {
+          if((i_dirty>>hr)&1) {
+            if(i_regmap[hr]<64) {
+              emit_storereg(i_regmap[hr],hr);
+              if(((i_is32>>i_regmap[hr])&1)) {
+                #ifdef DESTRUCTIVE_WRITEBACK
+                emit_sarimm(hr,31,hr);
+                emit_storereg(i_regmap[hr]|64,hr);
+                #else
+                emit_sarimm(hr,31,HOST_TEMPREG);
+                emit_storereg(i_regmap[hr]|64,HOST_TEMPREG);
+                #endif
+              }
+            }else{
+              if(!((i_is32>>(i_regmap[hr]&63))&1)) {
+                emit_storereg(i_regmap[hr],hr);
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
+
+#ifndef DESTRUCTIVE_WRITEBACK
+// Sign-extend to 64 bits and write out upper half of a register
+// This is useful where we have a 32-bit value in a register, and want to
+// keep it in a 32-bit register, but can't guarantee that it won't be read
+// as a 64-bit value later.
+static void wb_sx(signed char pre[],signed char entry[],uint64_t dirty,uint64_t is32_pre,uint64_t is32,uint64_t u,uint64_t uu)
+{
+  if(is32_pre==is32) return;
+  int hr,reg;
+  for(hr=0;hr<HOST_REGS;hr++) {
+    if(hr!=EXCLUDE_REG) {
+      //if(pre[hr]==entry[hr]) {
+        if((reg=pre[hr])>=0) {
+          if((dirty>>hr)&1) {
+            if( ((is32_pre&~is32&~uu)>>reg)&1 ) {
+              emit_sarimm(hr,31,HOST_TEMPREG);
+              emit_storereg(reg|64,HOST_TEMPREG);
+            }
+          }
+        }
+      //}
+    }
+  }
+}
+
+static void wb_valid(signed char pre[],signed char entry[],u_int dirty_pre,u_int dirty,uint64_t is32_pre,uint64_t u,uint64_t uu)
+{
+  //if(dirty_pre==dirty) return;
+  int hr,reg;
+  for(hr=0;hr<HOST_REGS;hr++) {
+    if(hr!=EXCLUDE_REG) {
+      reg=pre[hr];
+      if(((~u)>>(reg&63))&1) {
+        if(((reg&63)>0)&&((reg&63)<CSREG)) {
+          if(((dirty_pre&~dirty)>>hr)&1) {
+            if(reg<64) {
+              emit_storereg(reg,hr);
+              if( ((is32_pre&~uu)>>reg)&1 ) {
+                emit_sarimm(hr,31,HOST_TEMPREG);
+                emit_storereg(reg|64,HOST_TEMPREG);
+              }
+            } else {
+              emit_storereg(reg,hr);
+            }
+          }
+        }
+      }
+    }
+  }
+}
+#endif
+
 // Write out a single register
 static void wb_register(signed char r,signed char regmap[],uint64_t dirty,uint64_t is32)
 {
@@ -3164,71 +3904,1295 @@ static void wb_register(signed char r,signed char regmap[],uint64_t dirty,uint64
     }
   }
 }
-#if 0
-static int mchecksum(void)
+
+// Store dirty registers prior to branch
+static void store_regs_bt(signed char i_regmap[],uint64_t i_is32,uint64_t i_dirty,int addr)
 {
-  int i;
-  int sum=0;
-  for(i=0;i<2097152;i++) {
-    unsigned int temp=sum;
-    sum<<=1;
-    sum|=(~temp)>>31;
-    sum^=((u_int *)g_dev.rdram.dram)[i];
+  if(internal_branch(i_is32,addr))
+  {
+    int t=(addr-start)>>2;
+    int hr;
+    for(hr=0;hr<HOST_REGS;hr++) {
+      if(hr!=EXCLUDE_REG) {
+        if(((i_regmap[hr]&63)>0)&&((i_regmap[hr]&63)<CSREG)) {
+          if(i_regmap[hr]!=regs[t].regmap_entry[hr] || !((regs[t].dirty>>hr)&1) || (((i_is32&~regs[t].was32&~unneeded_reg_upper[t])>>(i_regmap[hr]&63))&1)) {
+            if((i_dirty>>hr)&1) {
+              if(i_regmap[hr]<64) {
+                if(!((unneeded_reg[t]>>i_regmap[hr])&1)) {
+                  emit_storereg(i_regmap[hr],hr);
+                  if(((i_is32>>i_regmap[hr])&1) && !((unneeded_reg_upper[t]>>i_regmap[hr])&1)) {
+                    #ifdef DESTRUCTIVE_WRITEBACK
+                    emit_sarimm(hr,31,hr);
+                    emit_storereg(i_regmap[hr]|64,hr);
+                    #else
+                    emit_sarimm(hr,31,HOST_TEMPREG);
+                    emit_storereg(i_regmap[hr]|64,HOST_TEMPREG);
+                    #endif
+                  }
+                }
+              }else{
+                if(!((i_is32>>(i_regmap[hr]&63))&1) && !((unneeded_reg_upper[t]>>(i_regmap[hr]&63))&1)) {
+                  emit_storereg(i_regmap[hr],hr);
+                }
+              }
+            }
+          }
+        }
+      }
+    }
   }
-  return sum;
-}
-
-static int rchecksum(void)
-{
-  int i;
-  int sum=0;
-  for(i=0;i<64;i++)
-    sum^=((u_int *)reg)[i];
-  return sum;
-}
-
-static void rlist(void)
-{
-  int i;
-  DebugMessage(M64MSG_VERBOSE, "TRACE: ");
-  for(i=0;i<32;i++)
-    DebugMessage(M64MSG_VERBOSE, "r%d:%8x%8x ",i,((int *)(reg+i))[1],((int *)(reg+i))[0]);
-  DebugMessage(M64MSG_VERBOSE, "TRACE: ");
-  for(i=0;i<32;i++)
-    DebugMessage(M64MSG_VERBOSE, "f%d:%8x%8x ",i,((int*)g_dev.r4300.cp1.regs_simple[i])[1],*((int*)g_dev.r4300.cp1.regs_simple[i]));
-}
-
-static void memdebug(int i)
-{
-  //DebugMessage(M64MSG_VERBOSE, "TRACE: count=%d next=%d (checksum %x) lo=%8x%8x",r4300_cp0_regs(&g_dev.r4300.cp0)[CP0_COUNT_REG],g_dev.r4300.cp0.next_interrupt,mchecksum(),(int)(reg[LOREG]>>32),(int)reg[LOREG]);
-  //DebugMessage(M64MSG_VERBOSE, "TRACE: count=%d next=%d (rchecksum %x)",r4300_cp0_regs(&g_dev.r4300.cp0)[CP0_COUNT_REG],g_dev.r4300.cp0.next_interrupt,rchecksum());
-  //rlist();
-  if((signed int)r4300_cp0_regs(&g_dev.r4300.cp0)[CP0_COUNT_REG]>=-2084597794&&(signed int)r4300_cp0_regs(&g_dev.r4300.cp0)[CP0_COUNT_REG]<0) {
-    DebugMessage(M64MSG_VERBOSE, "TRACE: count=%d next=%d (checksum %x)",r4300_cp0_regs(&g_dev.r4300.cp0)[CP0_COUNT_REG],g_dev.r4300.cp0.next_interrupt,mchecksum());
-    //DebugMessage(M64MSG_VERBOSE, "TRACE: count=%d next=%d (checksum %x) Status=%x",r4300_cp0_regs(&g_dev.r4300.cp0)[CP0_COUNT_REG],g_dev.r4300.cp0.next_interrupt,mchecksum(),r4300_cp0_regs(&g_dev.r4300.cp0)[CP0_STATUS_REG]);
-    //DebugMessage(M64MSG_VERBOSE, "TRACE: count=%d next=%d (checksum %x) hi=%8x%8x",r4300_cp0_regs(&g_dev.r4300.cp0)[CP0_COUNT_REG],g_dev.r4300.cp0.next_interrupt,mchecksum(),(int)(reg[HIREG]>>32),(int)reg[HIREG]);
-    rlist();
-    #if NEW_DYNAREC == NEW_DYNAREC_X86
-    DebugMessage(M64MSG_VERBOSE, "TRACE: %x",(&i)[-1]);
-    #endif
-    #if NEW_DYNAREC == NEW_DYNAREC_ARM
-    int j;
-    DebugMessage(M64MSG_VERBOSE, "TRACE: %x ",(&j)[10]);
-    DebugMessage(M64MSG_VERBOSE, "TRACE: %x %x %x %x %x %x %x %x %x %x %x %x %x %x %x %x %x %x %x %x",(&j)[1],(&j)[2],(&j)[3],(&j)[4],(&j)[5],(&j)[6],(&j)[7],(&j)[8],(&j)[9],(&j)[10],(&j)[11],(&j)[12],(&j)[13],(&j)[14],(&j)[15],(&j)[16],(&j)[17],(&j)[18],(&j)[19],(&j)[20]);
-    #endif
-    //fflush(stdout);
+  else
+  {
+    // Branch out of this block, write out all dirty regs
+    wb_dirtys(i_regmap,i_is32,i_dirty);
   }
-  //DebugMessage(M64MSG_VERBOSE, "TRACE: %x",(&i)[-1]);
 }
+
+// Load all needed registers for branch target
+static void load_regs_bt(signed char i_regmap[],uint64_t i_is32,uint64_t i_dirty,int addr)
+{
+  //if(addr>=start && addr<(start+slen*4))
+  if(internal_branch(i_is32,addr))
+  {
+    int t=(addr-start)>>2;
+    int hr;
+    // Store the cycle count before loading something else
+    if(i_regmap[HOST_CCREG]!=CCREG) {
+      assert(i_regmap[HOST_CCREG]==-1);
+    }
+    if(regs[t].regmap_entry[HOST_CCREG]!=CCREG) {
+      emit_storereg(CCREG,HOST_CCREG);
+    }
+    // Load 32-bit regs
+    for(hr=0;hr<HOST_REGS;hr++) {
+      if(hr!=EXCLUDE_REG&&regs[t].regmap_entry[hr]>=0&&regs[t].regmap_entry[hr]<TEMPREG) {
+        #ifdef DESTRUCTIVE_WRITEBACK
+        if(i_regmap[hr]!=regs[t].regmap_entry[hr] || ( !((regs[t].dirty>>hr)&1) && ((i_dirty>>hr)&1) && (((i_is32&~unneeded_reg_upper[t])>>i_regmap[hr])&1) ) || (((i_is32&~regs[t].was32&~unneeded_reg_upper[t])>>(i_regmap[hr]&63))&1)) {
+        #else
+        if(i_regmap[hr]!=regs[t].regmap_entry[hr] ) {
+        #endif
+          if(regs[t].regmap_entry[hr]==0) {
+            emit_zeroreg(hr);
+          }
+          else if(regs[t].regmap_entry[hr]!=CCREG)
+          {
+            emit_loadreg(regs[t].regmap_entry[hr],hr);
+          }
+        }
+      }
+    }
+    //Load 64-bit regs
+    for(hr=0;hr<HOST_REGS;hr++) {
+      if(hr!=EXCLUDE_REG&&regs[t].regmap_entry[hr]>=64&&regs[t].regmap_entry[hr]<TEMPREG+64) {
+        if(i_regmap[hr]!=regs[t].regmap_entry[hr]) {
+          assert(regs[t].regmap_entry[hr]!=64);
+          if((i_is32>>(regs[t].regmap_entry[hr]&63))&1) {
+            int lr=get_reg(regs[t].regmap_entry,regs[t].regmap_entry[hr]-64);
+            if(lr<0) {
+              emit_loadreg(regs[t].regmap_entry[hr],hr);
+            }
+            else
+            {
+              emit_sarimm(lr,31,hr);
+            }
+          }
+          else
+          {
+            emit_loadreg(regs[t].regmap_entry[hr],hr);
+          }
+        }
+        else if((i_is32>>(regs[t].regmap_entry[hr]&63))&1) {
+          int lr=get_reg(regs[t].regmap_entry,regs[t].regmap_entry[hr]-64);
+          if(lr<0) {
+            emit_loadreg(regs[t].regmap_entry[hr],hr);
+          }
+          else
+          {
+            emit_sarimm(lr,31,hr);
+          }
+        }
+      }
+    }
+  }
+}
+
+// Load the specified registers
+// This only loads the registers given as arguments because
+// we don't want to load things that will be overwritten
+static void load_regs(signed char entry[],signed char regmap[],int is32,int rs1,int rs2)
+{
+  int hr;
+  // Load 32-bit regs
+  for(hr=0;hr<HOST_REGS;hr++) {
+    if(hr!=EXCLUDE_REG&&regmap[hr]>=0) {
+      if(entry[hr]!=regmap[hr]) {
+        if(regmap[hr]==rs1||regmap[hr]==rs2)
+        {
+          if(regmap[hr]==0) {
+            emit_zeroreg(hr);
+          }
+          else
+          {
+            emit_loadreg(regmap[hr],hr);
+          }
+        }
+      }
+    }
+  }
+  //Load 64-bit regs
+  for(hr=0;hr<HOST_REGS;hr++) {
+    if(hr!=EXCLUDE_REG&&regmap[hr]>=0) {
+      if(entry[hr]!=regmap[hr]) {
+        if(regmap[hr]-64==rs1||regmap[hr]-64==rs2)
+        {
+          assert(regmap[hr]!=64);
+          if((is32>>(regmap[hr]&63))&1) {
+            int lr=get_reg(regmap,regmap[hr]-64);
+            if(lr>=0)
+              emit_sarimm(lr,31,hr);
+            else
+              emit_loadreg(regmap[hr],hr);
+          }
+          else
+          {
+            emit_loadreg(regmap[hr],hr);
+          }
+        }
+      }
+    }
+  }
+}
+
+// Load registers with known constants
+static void load_consts(signed char pre[],signed char regmap[],int is32,int i)
+{
+  int hr;
+  // Load 32-bit regs
+  for(hr=0;hr<HOST_REGS;hr++) {
+    if(hr!=EXCLUDE_REG&&regmap[hr]>=0) {
+      //if(entry[hr]!=regmap[hr]) {
+      if(i==0||!((regs[i-1].isconst>>hr)&1)||pre[hr]!=regmap[hr]||bt[i]) {
+        if(((regs[i].isconst>>hr)&1)&&regmap[hr]<64&&regmap[hr]>0) {
+          int value;
+          if(get_final_value(hr,i,&value)) {
+            if(value==0) {
+              emit_zeroreg(hr);
+            }
+            else {
+              emit_movimm(value,hr);
+            }
+          }
+        }
+      }
+    }
+  }
+  // Load 64-bit regs
+  for(hr=0;hr<HOST_REGS;hr++) {
+    if(hr!=EXCLUDE_REG&&regmap[hr]>=0) {
+      //if(entry[hr]!=regmap[hr]) {
+      if(i==0||!((regs[i-1].isconst>>hr)&1)||pre[hr]!=regmap[hr]||bt[i]) {
+        if(((regs[i].isconst>>hr)&1)&&regmap[hr]>64) {
+          if((is32>>(regmap[hr]&63))&1) {
+            int lr=get_reg(regmap,regmap[hr]-64);
+            assert(lr>=0);
+            emit_sarimm(lr,31,hr);
+          }
+          else
+          {
+            int value;
+            if(get_final_value(hr,i,&value)) {
+              if(value==0) {
+                emit_zeroreg(hr);
+              }
+              else {
+                emit_movimm(value,hr);
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
+
+static void load_all_consts(signed char regmap[],int is32,u_int dirty,u_int isconst,int i)
+{
+  int hr;
+  // Load 32-bit regs
+  for(hr=0;hr<HOST_REGS;hr++) {
+    if(hr!=EXCLUDE_REG&&regmap[hr]>=0&&((dirty>>hr)&1)) {
+      if(((isconst>>hr)&1)&&regmap[hr]<64&&regmap[hr]>0) {
+        int value=constmap[i][hr];
+        if(value==0) {
+          emit_zeroreg(hr);
+        }
+        else {
+          emit_movimm(value,hr);
+        }
+      }
+    }
+  }
+  // Load 64-bit regs
+  for(hr=0;hr<HOST_REGS;hr++) {
+    if(hr!=EXCLUDE_REG&&regmap[hr]>=0&&((dirty>>hr)&1)) {
+      if(((isconst>>hr)&1)&&regmap[hr]>64) {
+        if((is32>>(regmap[hr]&63))&1) {
+          int lr=get_reg(regmap,regmap[hr]-64);
+          assert(lr>=0);
+          emit_sarimm(lr,31,hr);
+        }
+        else
+        {
+          int value=constmap[i][hr];
+          if(value==0) {
+            emit_zeroreg(hr);
+          }
+          else {
+            emit_movimm(value,hr);
+          }
+        }
+      }
+    }
+  }
+}
+
+// Load all registers (except cycle count)
+static void load_all_regs(signed char i_regmap[])
+{
+  int hr;
+  for(hr=0;hr<HOST_REGS;hr++) {
+    if(hr!=EXCLUDE_REG) {
+      if(i_regmap[hr]==0) {
+        emit_zeroreg(hr);
+      }
+      else
+      if(i_regmap[hr]>0 && (i_regmap[hr]&63)<TEMPREG && i_regmap[hr]!=CCREG)
+      {
+        emit_loadreg(i_regmap[hr],hr);
+      }
+    }
+  }
+}
+
+// Load all current registers also needed by next instruction
+static void load_needed_regs(signed char i_regmap[],signed char next_regmap[])
+{
+  int hr;
+  for(hr=0;hr<HOST_REGS;hr++) {
+    if(hr!=EXCLUDE_REG) {
+      if(get_reg(next_regmap,i_regmap[hr])>=0) {
+        if(i_regmap[hr]==0) {
+          emit_zeroreg(hr);
+        }
+        else
+        if(i_regmap[hr]>0 && (i_regmap[hr]&63)<TEMPREG && i_regmap[hr]!=CCREG)
+        {
+          emit_loadreg(i_regmap[hr],hr);
+        }
+      }
+    }
+  }
+}
+
+// Load all regs, storing cycle count if necessary
+static void load_regs_entry(int t)
+{
+  int hr;
+  if(is_ds[t]) emit_addimm(HOST_CCREG,CLOCK_DIVIDER,HOST_CCREG);
+  else if(ccadj[t]) emit_addimm(HOST_CCREG,-ccadj[t]*CLOCK_DIVIDER,HOST_CCREG);
+  if(regs[t].regmap_entry[HOST_CCREG]!=CCREG) {
+    emit_storereg(CCREG,HOST_CCREG);
+  }
+  // Load 32-bit regs
+  for(hr=0;hr<HOST_REGS;hr++) {
+    if(regs[t].regmap_entry[hr]>=0&&regs[t].regmap_entry[hr]<TEMPREG) {
+      if(regs[t].regmap_entry[hr]==0) {
+        emit_zeroreg(hr);
+      }
+      else if(regs[t].regmap_entry[hr]!=CCREG)
+      {
+        emit_loadreg(regs[t].regmap_entry[hr],hr);
+      }
+    }
+  }
+  // Load 64-bit regs
+  for(hr=0;hr<HOST_REGS;hr++) {
+    if(regs[t].regmap_entry[hr]>=64&&regs[t].regmap_entry[hr]<TEMPREG+64) {
+      assert(regs[t].regmap_entry[hr]!=64);
+      if((regs[t].was32>>(regs[t].regmap_entry[hr]&63))&1) {
+        int lr=get_reg(regs[t].regmap_entry,regs[t].regmap_entry[hr]-64);
+        if(lr<0) {
+          emit_loadreg(regs[t].regmap_entry[hr],hr);
+        }
+        else
+        {
+          emit_sarimm(lr,31,hr);
+        }
+      }
+      else
+      {
+        emit_loadreg(regs[t].regmap_entry[hr],hr);
+      }
+    }
+  }
+}
+
+// Load registers prior to the start of a loop
+// so that they are not loaded within the loop
+static void loop_preload(signed char pre[],signed char entry[])
+{
+  int hr;
+  for(hr=0;hr<HOST_REGS;hr++) {
+    if(hr!=EXCLUDE_REG) {
+      if(pre[hr]!=entry[hr]) {
+        if(entry[hr]>=0) {
+          if(get_reg(pre,entry[hr])<0) {
+            assem_debug("loop preload:");
+            //DebugMessage(M64MSG_VERBOSE, "loop preload: %d",hr);
+            if(entry[hr]==0) {
+              emit_zeroreg(hr);
+            }
+            else if(entry[hr]<TEMPREG)
+            {
+              emit_loadreg(entry[hr],hr);
+            }
+            else if(entry[hr]-64<TEMPREG)
+            {
+              emit_loadreg(entry[hr],hr);
+            }
+          }
+        }
+      }
+    }
+  }
+}
+
+// Generate address for load/store instruction
+static void address_generation(int i,struct regstat *i_regs,signed char entry[])
+{
+  if(itype[i]==LOAD||itype[i]==LOADLR||itype[i]==STORE||itype[i]==STORELR||itype[i]==C1LS) {
+    int ra=0;
+    int agr=AGEN1+(i&1);
+    int mgr=MGEN1+(i&1);
+    if(itype[i]==LOAD) {
+      ra=get_reg(i_regs->regmap,agr);
+      if(ra<0) ra=get_reg(i_regs->regmap,-1);
+      assert(ra>=0);
+    }
+    if(itype[i]==LOADLR) {
+      ra=get_reg(i_regs->regmap,FTEMP);
+    }
+    if(itype[i]==STORE||itype[i]==STORELR) {
+      ra=get_reg(i_regs->regmap,agr);
+      if(ra<0) ra=get_reg(i_regs->regmap,-1);
+    }
+    if(itype[i]==C1LS) {
+      if (opcode[i]==0x31||opcode[i]==0x35) // LWC1/LDC1
+        ra=get_reg(i_regs->regmap,FTEMP);
+      else { // SWC1/SDC1
+        ra=get_reg(i_regs->regmap,agr);
+        if(ra<0) ra=get_reg(i_regs->regmap,-1);
+      }
+    }
+    int rs=get_reg(i_regs->regmap,rs1[i]);
+    int rm=get_reg(i_regs->regmap,TLREG);
+    if(ra>=0) {
+      int offset=imm[i];
+      int c=(i_regs->wasconst>>rs)&1;
+      if(rs1[i]==0) {
+        // Using r0 as a base address
+        /*if(rm>=0) {
+          if(!entry||entry[rm]!=mgr) {
+            generate_map_const(offset,rm);
+          } // else did it in the previous cycle
+        }*/
+        if(!entry||entry[ra]!=agr) {
+          emit_movimm(offset,ra);
+        } // else did it in the previous cycle
+      }
+      else if(rs<0) {
+        if(!entry||entry[ra]!=rs1[i])
+          emit_loadreg(rs1[i],ra);
+        //if(!entry||entry[ra]!=rs1[i])
+        //  DebugMessage(M64MSG_VERBOSE, "poor load scheduling!");
+      }
+      else if(c) {
+        if(rm>=0) {
+          if(!entry||entry[rm]!=mgr) {
+            if(itype[i]==STORE||itype[i]==STORELR||opcode[i]==0x39||opcode[i]==0x3D) {
+              // Stores to memory go thru the mapper to detect self-modifying
+              // code, loads don't.
+              if((unsigned int)(constmap[i][rs]+offset)>=0xC0000000 ||
+                 (unsigned int)(constmap[i][rs]+offset)<0x80800000 )
+                generate_map_const(constmap[i][rs]+offset,rm);
+            }else{
+              if((signed int)(constmap[i][rs]+offset)>=(signed int)0xC0000000)
+                generate_map_const(constmap[i][rs]+offset,rm);
+            }
+          }
+        }
+
+        if(!entry||entry[ra]!=agr) {
+          int load=0;
+          #ifndef INTERPRET_LOAD
+          load|=(itype[i]==LOAD);
+          #endif
+          #ifndef INTERPRET_C1LS
+          load|=(opcode[i]==0x31||opcode[i]==0x35);
+          #endif
+          #ifndef INTERPRET_LOADLR
+          load|=(opcode[i]==0x22||opcode[i]==0x26||opcode[i]==0x1a||opcode[i]==0x1b);
+          #endif
+
+          #ifdef HOST_IMM_ADDR32
+          // on x86 when source reg is constant
+          // LB/LBU/LH/LHU/LW/LWU/LD/LWC1/LDC1 doesn't need address generation when not using tlb
+          // SDL/SDR/SB/SH/SW/SD/SWC1/SDC1 always need address generation
+          // ROREG not required
+          if(!load||(using_tlb&&((signed int)constmap[i][rs]+offset)>=(signed int)0xC0000000))
+          #endif
+          #if defined(RAM_OFFSET) && !defined(NATIVE_64)
+          // on arm address generation is always required
+          // When LB/LBU/LH/LHU/LW/LWU/LD/LWC1/LDC1 loading from rdram, add rdram address to load address to avoid loading ROREG
+          // ROREG only required for SDL/SDR/SB/SH/SW/SD/SWC1/SDC1 and when not loading from rdram
+          if(load&&(signed int)constmap[i][rs]+offset<(signed int)0x80800000)
+            emit_movimm(constmap[i][rs]+offset+(intptr_t)g_dev.rdram.dram-(intptr_t)0x80000000,ra);
+          else
+          #endif
+          // on arm64 and x64 when source reg is constant
+          // address generation is always required
+          // ROREG is always required
+          emit_movimm(constmap[i][rs]+offset,ra);
+
+        } // else did it in the previous cycle
+      }
+      if(offset&&!c&&rs1[i]) {
+        if(rs>=0) {
+          emit_addimm(rs,offset,ra);
+        }else{
+          emit_addimm(ra,offset,ra);
+        }
+      }
+    }
+  }
+  // Preload constants for next instruction
+  if(itype[i+1]==LOAD||itype[i+1]==LOADLR||itype[i+1]==STORE||itype[i+1]==STORELR||itype[i+1]==C1LS) {
+    int agr,ra;
+    #if (NEW_DYNAREC!=NEW_DYNAREC_X86) && (NEW_DYNAREC!=NEW_DYNAREC_X64)
+    // Mapper entry
+    agr=MGEN1+((i+1)&1);
+    ra=get_reg(i_regs->regmap,agr);
+    if(ra>=0) {
+      int rs=get_reg(regs[i+1].regmap,rs1[i+1]);
+      int offset=imm[i+1];
+      int c=(regs[i+1].wasconst>>rs)&1;
+      if(c) {
+        if(itype[i+1]==STORE||itype[i+1]==STORELR||opcode[i+1]==0x39||opcode[i+1]==0x3D) {
+          // Stores to memory go thru the mapper to detect self-modifying
+          // code, loads don't.
+          if((unsigned int)(constmap[i+1][rs]+offset)>=0xC0000000 ||
+             (unsigned int)(constmap[i+1][rs]+offset)<0x80800000 )
+            generate_map_const(constmap[i+1][rs]+offset,ra);
+        }else{
+          if((signed int)(constmap[i+1][rs]+offset)>=(signed int)0xC0000000)
+            generate_map_const(constmap[i+1][rs]+offset,ra);
+        }
+      }
+      /*else if(rs1[i]==0) {
+        generate_map_const(offset,ra);
+      }*/
+    }
+    #endif
+    // Actual address
+    agr=AGEN1+((i+1)&1);
+    ra=get_reg(i_regs->regmap,agr);
+    if(ra>=0) {
+      int rs=get_reg(regs[i+1].regmap,rs1[i+1]);
+      int offset=imm[i+1];
+      int c=(regs[i+1].wasconst>>rs)&1;
+      if(c) {
+        int load=0;
+        #ifndef INTERPRET_LOAD
+        load|=(itype[i+1]==LOAD);
+        #endif
+        #ifndef INTERPRET_C1LS
+        load|=(opcode[i+1]==0x31||opcode[i+1]==0x35);
+        #endif
+        #ifndef INTERPRET_LOADLR
+        load|=(opcode[i+1]==0x22||opcode[i+1]==0x26||opcode[i+1]==0x1a||opcode[i+1]==0x1b);
+        #endif
+
+        #ifdef HOST_IMM_ADDR32
+        if(!load||(using_tlb&&((signed int)constmap[i+1][rs]+offset)>=(signed int)0xC0000000))
+        #endif
+        #if defined(RAM_OFFSET) && !defined(NATIVE_64)
+        if(load&&(signed int)constmap[i+1][rs]+offset<(signed int)0x80800000)
+          emit_movimm(constmap[i+1][rs]+offset+(intptr_t)g_dev.rdram.dram-(intptr_t)0x80000000,ra);
+        else
+        #endif
+        emit_movimm(constmap[i+1][rs]+offset,ra);
+      }
+      else if(rs1[i+1]==0) {
+        // Using r0 as a base address
+        emit_movimm(offset,ra);
+      }
+    }
+  }
+}
+
+static void emit_extjump(intptr_t addr, int target)
+{
+  emit_extjump2(addr, target, (intptr_t)dyna_linker);
+}
+
+static void emit_extjump_ds(intptr_t addr, int target)
+{
+  emit_extjump2(addr, target, (intptr_t)dyna_linker_ds);
+}
+
+static void do_cc(int i,signed char i_regmap[],int *adj,int addr,int taken,int invert)
+{
+  int count;
+  intptr_t jaddr;
+  intptr_t idle=0;
+  if(itype[i]==RJUMP)
+  {
+    *adj=0;
+  }
+  //if(ba[i]>=start && ba[i]<(start+slen*4))
+  if(internal_branch(branch_regs[i].is32,ba[i]))
+  {
+    int t=(ba[i]-start)>>2;
+    if(is_ds[t]) *adj=-1; // Branch into delay slot adds an extra cycle
+    else *adj=ccadj[t];
+  }
+  else
+  {
+    *adj=0;
+  }
+  count=ccadj[i];
+  if(taken==TAKEN && i==(ba[i]-start)>>2 && source[i+1]==0) {
+    // Idle loop
+    idle=(intptr_t)out;
+    emit_test(HOST_CCREG,HOST_CCREG);
+#if NEW_DYNAREC >= NEW_DYNAREC_ARM
+    emit_cmovs_imm(0,HOST_CCREG);
+#else
+    emit_cmovs(&const_zero,HOST_CCREG);
+#endif
+    emit_addimm(HOST_CCREG,CLOCK_DIVIDER*2,HOST_CCREG);
+    jaddr=(intptr_t)out;
+    emit_jmp(0);
+  }
+  else if(*adj==0||invert) {
+    emit_addimm_and_set_flags(CLOCK_DIVIDER*(count+2),HOST_CCREG);
+    jaddr=(intptr_t)out;
+    emit_jns(0);
+  }
+  else
+  {
+    emit_cmpimm(HOST_CCREG,-(int)CLOCK_DIVIDER*(count+2));
+    jaddr=(intptr_t)out;
+    emit_jns(0);
+  }
+  add_stub(CC_STUB,jaddr,idle?idle:(intptr_t)out,(*adj==0||invert||idle)?0:(count+2),i,addr,taken,0);
+}
+
+/**** Stubs ****/
+static void do_ccstub(int n)
+{
+  literal_pool(256);
+  assem_debug("do_ccstub %x",start+stubs[n][4]*4);
+  set_jump_target(stubs[n][1],(intptr_t)out);
+  int i=stubs[n][4];
+  if(stubs[n][6]==NULLDS) {
+    // Delay slot instruction is nullified ("likely" branch)
+    wb_dirtys(regs[i].regmap,regs[i].is32,regs[i].dirty);
+  }
+  else if(stubs[n][6]!=TAKEN) {
+    wb_dirtys(branch_regs[i].regmap,branch_regs[i].is32,branch_regs[i].dirty);
+  }
+  else {
+    if(internal_branch(branch_regs[i].is32,ba[i]))
+      wb_needed_dirtys(branch_regs[i].regmap,branch_regs[i].is32,branch_regs[i].dirty,ba[i]);
+  }
+  if(stubs[n][5]!=-1)
+  {
+    // Save PC as return address
+    emit_movimm(stubs[n][5],0);
+    emit_writeword(0,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.pcaddr);
+  }
+  else
+  {
+    // Return address depends on which way the branch goes
+    if(itype[i]==CJUMP||itype[i]==SJUMP||itype[i]==FJUMP)
+    {
+      int s1l=get_reg(branch_regs[i].regmap,rs1[i]);
+      int s1h=get_reg(branch_regs[i].regmap,rs1[i]|64);
+      int s2l=get_reg(branch_regs[i].regmap,rs2[i]);
+      int s2h=get_reg(branch_regs[i].regmap,rs2[i]|64);
+      if(rs1[i]==0)
+      {
+        s1l=s2l;s1h=s2h;
+        s2l=s2h=-1;
+      }
+      else if(rs2[i]==0)
+      {
+        s2l=s2h=-1;
+      }
+      if((branch_regs[i].is32>>rs1[i])&(branch_regs[i].is32>>rs2[i])&1) {
+        s1h=s2h=-1;
+      }
+      assert(s1l>=0);
+      #ifdef DESTRUCTIVE_WRITEBACK
+      if(rs1[i]) {
+        if((branch_regs[i].dirty>>s1l)&(branch_regs[i].is32>>rs1[i])&1)
+          emit_loadreg(rs1[i],s1l);
+      }
+      else {
+        if((branch_regs[i].dirty>>s1l)&(branch_regs[i].is32>>rs2[i])&1)
+          emit_loadreg(rs2[i],s1l);
+      }
+      if(s2l>=0)
+        if((branch_regs[i].dirty>>s2l)&(branch_regs[i].is32>>rs2[i])&1)
+          emit_loadreg(rs2[i],s2l);
+      #endif
+      int hr=0;
+      int addr,alt,ntaddr;
+      while(hr<HOST_REGS)
+      {
+        if(hr!=EXCLUDE_REG && hr!=HOST_CCREG &&
+           (branch_regs[i].regmap[hr]&63)!=rs1[i] &&
+           (branch_regs[i].regmap[hr]&63)!=rs2[i] )
+        {
+          addr=hr++;break;
+        }
+        hr++;
+      }
+      while(hr<HOST_REGS)
+      {
+        if(hr!=EXCLUDE_REG && hr!=HOST_CCREG &&
+           (branch_regs[i].regmap[hr]&63)!=rs1[i] &&
+           (branch_regs[i].regmap[hr]&63)!=rs2[i] )
+        {
+          alt=hr++;break;
+        }
+        hr++;
+      }
+      if((opcode[i]&0x2E)==6) // BLEZ/BGTZ needs another register
+      {
+        while(hr<HOST_REGS)
+        {
+          if(hr!=EXCLUDE_REG && hr!=HOST_CCREG &&
+             (branch_regs[i].regmap[hr]&63)!=rs1[i] &&
+             (branch_regs[i].regmap[hr]&63)!=rs2[i] )
+          {
+            ntaddr=hr;break;
+          }
+          hr++;
+        }
+        assert(hr<HOST_REGS);
+      }
+      if((opcode[i]&0x2f)==4) // BEQ
+      {
+        #ifdef HAVE_CMOV_IMM
+        if(s1h<0) {
+          if(s2l>=0) emit_cmp(s1l,s2l);
+          else emit_test(s1l,s1l);
+          emit_cmov2imm_e_ne_compact(ba[i],start+i*4+8,addr);
+        }
+        else
+        #endif
+        {
+          emit_mov2imm_compact(ba[i],addr,start+i*4+8,alt);
+          if(s1h>=0) {
+            if(s2h>=0) emit_cmp(s1h,s2h);
+            else emit_test(s1h,s1h);
+            emit_cmovne_reg(alt,addr);
+          }
+          if(s2l>=0) emit_cmp(s1l,s2l);
+          else emit_test(s1l,s1l);
+          emit_cmovne_reg(alt,addr);
+        }
+      }
+      if((opcode[i]&0x2f)==5) // BNE
+      {
+        #ifdef HAVE_CMOV_IMM
+        if(s1h<0) {
+          if(s2l>=0) emit_cmp(s1l,s2l);
+          else emit_test(s1l,s1l);
+          emit_cmov2imm_e_ne_compact(start+i*4+8,ba[i],addr);
+        }
+        else
+        #endif
+        {
+          emit_mov2imm_compact(start+i*4+8,addr,ba[i],alt);
+          if(s1h>=0) {
+            if(s2h>=0) emit_cmp(s1h,s2h);
+            else emit_test(s1h,s1h);
+            emit_cmovne_reg(alt,addr);
+          }
+          if(s2l>=0) emit_cmp(s1l,s2l);
+          else emit_test(s1l,s1l);
+          emit_cmovne_reg(alt,addr);
+        }
+      }
+      if((opcode[i]&0x2f)==6) // BLEZ
+      {
+        //emit_movimm(ba[i],alt);
+        //emit_movimm(start+i*4+8,addr);
+        emit_mov2imm_compact(ba[i],alt,start+i*4+8,addr);
+        emit_cmpimm(s1l,1);
+        if(s1h>=0) emit_mov(addr,ntaddr);
+        emit_cmovl_reg(alt,addr);
+        if(s1h>=0) {
+          emit_test(s1h,s1h);
+          emit_cmovne_reg(ntaddr,addr);
+          emit_cmovs_reg(alt,addr);
+        }
+      }
+      if((opcode[i]&0x2f)==7) // BGTZ
+      {
+        //emit_movimm(ba[i],addr);
+        //emit_movimm(start+i*4+8,ntaddr);
+        emit_mov2imm_compact(ba[i],addr,start+i*4+8,ntaddr);
+        emit_cmpimm(s1l,1);
+        if(s1h>=0) emit_mov(addr,alt);
+        emit_cmovl_reg(ntaddr,addr);
+        if(s1h>=0) {
+          emit_test(s1h,s1h);
+          emit_cmovne_reg(alt,addr);
+          emit_cmovs_reg(ntaddr,addr);
+        }
+      }
+      if((opcode[i]==1)&&(opcode2[i]&0x2D)==0) // BLTZ
+      {
+        //emit_movimm(ba[i],alt);
+        //emit_movimm(start+i*4+8,addr);
+        emit_mov2imm_compact(ba[i],alt,start+i*4+8,addr);
+        if(s1h>=0) emit_test(s1h,s1h);
+        else emit_test(s1l,s1l);
+        emit_cmovs_reg(alt,addr);
+      }
+      if((opcode[i]==1)&&(opcode2[i]&0x2D)==1) // BGEZ
+      {
+        //emit_movimm(ba[i],addr);
+        //emit_movimm(start+i*4+8,alt);
+        emit_mov2imm_compact(ba[i],addr,start+i*4+8,alt);
+        if(s1h>=0) emit_test(s1h,s1h);
+        else emit_test(s1l,s1l);
+        emit_cmovs_reg(alt,addr);
+      }
+      if(opcode[i]==0x11 && opcode2[i]==0x08 ) {
+        if(source[i]&0x10000) // BC1T
+        {
+          //emit_movimm(ba[i],alt);
+          //emit_movimm(start+i*4+8,addr);
+          emit_mov2imm_compact(ba[i],alt,start+i*4+8,addr);
+          emit_testimm(s1l,0x800000);
+          emit_cmovne_reg(alt,addr);
+        }
+        else // BC1F
+        {
+          //emit_movimm(ba[i],addr);
+          //emit_movimm(start+i*4+8,alt);
+          emit_mov2imm_compact(ba[i],addr,start+i*4+8,alt);
+          emit_testimm(s1l,0x800000);
+          emit_cmovne_reg(alt,addr);
+        }
+      }
+      emit_writeword(addr,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.pcaddr);
+    }
+    else
+    if(itype[i]==RJUMP)
+    {
+      int r=get_reg(branch_regs[i].regmap,rs1[i]);
+      if((rs1[i]==rt1[i+1]||rs1[i]==rt2[i+1])&&(rs1[i]!=0)) {
+        r=get_reg(branch_regs[i].regmap,RTEMP);
+      }
+      emit_writeword(r,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.pcaddr);
+    }
+    else {DebugMessage(M64MSG_ERROR, "Unknown branch type in do_ccstub");exit(1);}
+  }
+  // Update cycle count
+  assert(branch_regs[i].regmap[HOST_CCREG]==CCREG||branch_regs[i].regmap[HOST_CCREG]==-1);
+  if(stubs[n][3]) emit_addimm(HOST_CCREG,CLOCK_DIVIDER*stubs[n][3],HOST_CCREG);
+  emit_call((intptr_t)cc_interrupt);
+  if(stubs[n][3]) emit_addimm(HOST_CCREG,-(int)CLOCK_DIVIDER*stubs[n][3],HOST_CCREG);
+  if(stubs[n][6]==TAKEN) {
+    if(internal_branch(branch_regs[i].is32,ba[i]))
+      load_needed_regs(branch_regs[i].regmap,regs[(ba[i]-start)>>2].regmap_entry);
+    else if(itype[i]==RJUMP) {
+      int r=get_reg(branch_regs[i].regmap,rs1[i]);
+      if((rs1[i]==rt1[i+1]||rs1[i]==rt2[i+1])&&(rs1[i]!=0)) {
+        r=get_reg(branch_regs[i].regmap,RTEMP);
+      }
+#if NEW_DYNAREC==NEW_DYNAREC_ARM64
+      if(r==18) {
+        // x18 is used for trampoline jumps, move it to another register (x0)
+        emit_mov(r,0);
+        r=0;
+        stubs[n][2]=jump_vaddr_reg[0];
+      }
+#endif
+      emit_readword((intptr_t)&g_dev.r4300.new_dynarec_hot_state.pcaddr,r);
+    }
+  }else if(stubs[n][6]==NOTTAKEN) {
+    if(i<slen-2) load_needed_regs(branch_regs[i].regmap,regmap_pre[i+2]);
+    else load_all_regs(branch_regs[i].regmap);
+  }else if(stubs[n][6]==NULLDS) {
+    // Delay slot instruction is nullified ("likely" branch)
+    if(i<slen-2) load_needed_regs(regs[i].regmap,regmap_pre[i+2]);
+    else load_all_regs(regs[i].regmap);
+  }else{
+    load_all_regs(branch_regs[i].regmap);
+  }
+  emit_jmp(stubs[n][2]); // return address
+}
+
+static void do_cop1stub(int n)
+{
+  literal_pool(256);
+  assem_debug("do_cop1stub %x",start+stubs[n][3]*4);
+  set_jump_target(stubs[n][1],(intptr_t)out);
+  int i=stubs[n][3];
+  struct regstat *i_regs=(struct regstat *)stubs[n][5];
+  int ds=stubs[n][6];
+  if(!ds) {
+    load_all_consts(regs[i].regmap_entry,regs[i].was32,regs[i].wasdirty,regs[i].wasconst,i);
+    //if(i_regs!=&regs[i]) DebugMessage(M64MSG_VERBOSE, "oops: regs[i]=%x i_regs=%x",(int)&regs[i],(int)i_regs);
+  }
+  wb_dirtys(i_regs->regmap_entry,i_regs->was32,i_regs->wasdirty);
+
+  if(get_reg(i_regs->regmap,CCREG)<0) {
+    emit_loadreg(CCREG,HOST_CCREG);
+  }
+
+  emit_movimm(start+(i*4)+ds,0); // Get PC
+  emit_addimm(HOST_CCREG,CLOCK_DIVIDER*ccadj[i],HOST_CCREG); // CHECK: is this right?  There should probably be an extra cycle...
+  emit_jmp((intptr_t)fp_exception);
+}
+
+static void do_readstub(int n)
+{
+  assem_debug("do_readstub %x",start+stubs[n][3]*4);
+  literal_pool(256);
+  set_jump_target(stubs[n][1],(intptr_t)out);
+  int type=stubs[n][0];
+  int i=stubs[n][3];
+  int addr=stubs[n][4];
+  struct regstat *i_regs=(struct regstat *)stubs[n][5];
+  u_int reglist=stubs[n][7];
+  signed char *i_regmap=i_regs->regmap;
+  int rth,rt;
+
+  if(itype[i]==C1LS) {
+    rth=get_reg(i_regmap,FTEMP|64);
+    rt=get_reg(i_regmap,FTEMP);
+  }else{
+    rth=get_reg(i_regmap,rt1[i]|64);
+    rt=get_reg(i_regmap,rt1[i]);
+  }
+
+  assert(addr>=0);
+  emit_writeword(addr,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.address);
+
+  intptr_t ftable=0;
+  if(type==LOADB_STUB||type==LOADBU_STUB)
+    ftable=(intptr_t)read_byte_new;
+  else if(type==LOADH_STUB||type==LOADHU_STUB)
+    ftable=(intptr_t)read_hword_new;
+  else if(type==LOADW_STUB||type==LOADWU_STUB)
+    ftable=(intptr_t)read_word_new;
+  else if(type==LOADD_STUB)
+    ftable=(intptr_t)read_dword_new;
+  else if(type==LOADWL_STUB)
+  {
+    assert(rt>=0);
+    ftable=(intptr_t)LWL_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
+  }
+  else if(type==LOADWR_STUB)
+  {
+    assert(rt>=0);
+    ftable=(intptr_t)LWR_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
+  }
+  else if(type==LOADDL_STUB)
+  {
+    assert(rt>=0);
+    assert(rth>=0);
+    ftable=(intptr_t)LDL_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
+    emit_writeword(rth,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
+  }
+  else if(type==LOADDR_STUB)
+  {
+    assert(rt>=0);
+    assert(rth>=0);
+    ftable=(intptr_t)LDR_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
+    emit_writeword(rth,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
+  }
+
+  int cc=get_reg(i_regmap,CCREG);
+  if(cc>=0) {
+    emit_storereg(CCREG,cc);
+  }
+
+  save_regs(reglist);
+  int ds=i_regs!=&regs[i];
+
+#if NEW_DYNAREC == NEW_DYNAREC_X86
+  emit_pushimm(CLOCK_DIVIDER*(stubs[n][6]+1));
+  emit_pushimm((start+(i+1)*4)+ds);
+  emit_call((intptr_t)ftable);
+  emit_addimm(ESP,8,ESP);
+#else
+  emit_movimm((start+(i+1)*4)+ds,ARG1_REG);
+  emit_movimm(CLOCK_DIVIDER*(stubs[n][6]+1),ARG2_REG);
+  emit_call((intptr_t)ftable);
 #endif
 
-/* Debug:
-static void tlb_debug(u_int cause, u_int addr, u_int iaddr)
-{
-  DebugMessage(M64MSG_VERBOSE, "TLB Exception: instruction=%x addr=%x cause=%x",iaddr, addr, cause);
-}
-end debug */
+  restore_regs(reglist);
 
+  emit_cmpmem_imm((intptr_t)&g_dev.r4300.new_dynarec_hot_state.pending_exception,0);
+  intptr_t jaddr=(intptr_t)out;
+  emit_jeq(0);
+
+  if(!ds) load_all_consts(regs[i].regmap_entry,regs[i].was32,regs[i].wasdirty,regs[i].wasconst,i);
+  wb_dirtys(i_regs->regmap_entry,i_regs->was32,i_regs->wasdirty);
+
+  emit_jmp((intptr_t)&do_interrupt);
+  set_jump_target(jaddr,(intptr_t)out);
+
+  if(rt>=0) {
+    if(type==LOADB_STUB)
+      emit_movsbl((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
+    else if(type==LOADBU_STUB)
+      emit_movzbl((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
+    else if(type==LOADH_STUB)
+      emit_movswl((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
+    else if(type==LOADHU_STUB)
+      emit_movzwl((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
+    else if(type==LOADW_STUB||type==LOADWU_STUB||type==LOADWL_STUB||type==LOADWR_STUB) {
+      emit_readword((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
+      if(type==LOADWU_STUB) emit_zeroreg(rth);
+    }
+    else if(type==LOADD_STUB||type==LOADDL_STUB||type==LOADDR_STUB) {
+      emit_readword((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
+      if(rth>=0) emit_readword((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword+4,rth);
+    }
+  }
+  emit_jmp(stubs[n][2]); // return address
+}
+
+static void inline_readstub(int type, int i, u_int addr_const, char addr, struct regstat *i_regs, int target, int adj, u_int reglist)
+{
+  assem_debug("inline_readstub");
+  int rth=get_reg(i_regs->regmap,target|64);
+  int rt=get_reg(i_regs->regmap,target);
+
+#if NEW_DYNAREC <= NEW_DYNAREC_X64
+  if(addr_const)
+    emit_writeword_imm(addr_const,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.address);
+  else
+#endif
+  {
+    assert(addr>=0);
+    emit_writeword(addr,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.address);
+  }
+
+  intptr_t ftable=0;
+  if(type==LOADB_STUB||type==LOADBU_STUB)
+    ftable=(intptr_t)read_byte_new;
+  else if(type==LOADH_STUB||type==LOADHU_STUB)
+    ftable=(intptr_t)read_hword_new;
+  else if(type==LOADW_STUB||type==LOADWU_STUB)
+    ftable=(intptr_t)read_word_new;
+  else if(type==LOADD_STUB)
+    ftable=(intptr_t)read_dword_new;
+  else if(type==LOADWL_STUB)
+  {
+    assert(rt>=0);
+    ftable=(intptr_t)LWL_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
+  }
+  else if(type==LOADWR_STUB)
+  {
+    assert(rt>=0);
+    ftable=(intptr_t)LWR_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
+  }
+  else if(type==LOADDL_STUB)
+  {
+    assert(rt>=0);
+    assert(rth>=0);
+    ftable=(intptr_t)LDL_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
+    emit_writeword(rth,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
+  }
+  else if(type==LOADDR_STUB)
+  {
+    assert(rt>=0);
+    assert(rth>=0);
+    ftable=(intptr_t)LDR_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
+    emit_writeword(rth,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
+  }
+
+  int cc=get_reg(i_regs->regmap,CCREG);
+  if(cc>=0) {
+    emit_storereg(CCREG,cc);
+  }
+
+  save_regs(reglist);
+  int ds=i_regs!=&regs[i];
+
+#if NEW_DYNAREC == NEW_DYNAREC_X86
+  emit_pushimm(CLOCK_DIVIDER*(adj+1));
+  emit_pushimm((start+(i+1)*4)+ds);
+  emit_call((intptr_t)ftable);
+  emit_addimm(ESP,8,ESP);
+#else
+  emit_movimm((start+(i+1)*4)+ds,ARG1_REG);
+  emit_movimm(CLOCK_DIVIDER*(adj+1),ARG2_REG);
+  emit_call((intptr_t)ftable);
+#endif
+
+  restore_regs(reglist);
+
+  if((signed int)addr_const>=(signed int)0xC0000000) {
+    // Theoretically we can have a pagefault here, if the TLB has never
+    // been enabled and the address is outside the range 80000000..BFFFFFFF
+    // Write out the registers so the pagefault can be handled.  This is
+    // a very rare case and likely represents a bug.
+    emit_cmpmem_imm((intptr_t)&g_dev.r4300.new_dynarec_hot_state.pending_exception,0);
+    intptr_t jaddr=(intptr_t)out;
+    emit_jeq(0);
+
+    if(!ds) load_all_consts(regs[i].regmap_entry,regs[i].was32,regs[i].wasdirty,regs[i].wasconst,i);
+    wb_dirtys(i_regs->regmap_entry,i_regs->was32,i_regs->wasdirty);
+
+    emit_jmp((intptr_t)&do_interrupt);
+    set_jump_target(jaddr,(intptr_t)out);
+  }
+
+  if(rt>=0) {
+    if(type==LOADB_STUB)
+      emit_movsbl((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
+    else if(type==LOADBU_STUB)
+      emit_movzbl((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
+    else if(type==LOADH_STUB)
+      emit_movswl((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
+    else if(type==LOADHU_STUB)
+      emit_movzwl((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
+    else if(type==LOADW_STUB||type==LOADWU_STUB||type==LOADWL_STUB||type==LOADWR_STUB) {
+      emit_readword((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
+      if(type==LOADWU_STUB) emit_zeroreg(rth);
+    }
+    else if(type==LOADD_STUB||type==LOADDL_STUB||type==LOADDR_STUB) {
+      emit_readword((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
+      if(rth>=0) emit_readword((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword+4,rth);
+    }
+  }
+}
+
+static void do_writestub(int n)
+{
+  assem_debug("do_writestub %x",start+stubs[n][3]*4);
+  literal_pool(256);
+  set_jump_target(stubs[n][1],(intptr_t)out);
+  int type=stubs[n][0];
+  int i=stubs[n][3];
+  int addr=stubs[n][4];
+  struct regstat *i_regs=(struct regstat *)stubs[n][5];
+  u_int reglist=stubs[n][7];
+  signed char *i_regmap=i_regs->regmap;
+  int rth,rt,r;
+
+  if(itype[i]==C1LS) {
+    rth=get_reg(i_regmap,FTEMP|64);
+    rt=get_reg(i_regmap,r=FTEMP);
+  }else{
+    rth=get_reg(i_regmap,rs2[i]|64);
+    rt=get_reg(i_regmap,r=rs2[i]);
+  }
+  assert(addr>=0);
+  assert(rt>=0);
+  emit_writeword(addr,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.address);
+
+  intptr_t ftable=0;
+  if(type==STOREB_STUB){
+    ftable=(intptr_t)write_byte_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
+  }
+  else if(type==STOREH_STUB){
+    ftable=(intptr_t)write_hword_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
+  }
+  else if(type==STOREW_STUB){
+    ftable=(intptr_t)write_word_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
+  }
+  else if(type==STORED_STUB){
+    ftable=(intptr_t)write_dword_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
+    emit_writeword(r?rth:rt,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
+  }
+  else if(type==STOREWL_STUB){
+    ftable=(intptr_t)SWL_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
+  }
+  else if(type==STOREWR_STUB){
+    ftable=(intptr_t)SWR_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
+  }
+  else if(type==STOREDL_STUB){
+    ftable=(intptr_t)SDL_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
+    emit_writeword(r?rth:rt,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
+  }
+  else if(type==STOREDR_STUB){
+    ftable=(intptr_t)SDR_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
+    emit_writeword(r?rth:rt,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
+  }
+
+  int cc=get_reg(i_regmap,CCREG);
+  if(cc>=0) {
+    emit_storereg(CCREG,cc);
+  }
+
+  save_regs(reglist);
+  int ds=i_regs!=&regs[i];
+
+#if NEW_DYNAREC == NEW_DYNAREC_X86
+  emit_pushimm(CLOCK_DIVIDER*(stubs[n][6]+1));
+  emit_pushimm((start+(i+1)*4)+ds);
+  emit_call((intptr_t)ftable);
+  emit_addimm(ESP,8,ESP);
+#else
+  emit_movimm((start+(i+1)*4)+ds,ARG1_REG);
+  emit_movimm(CLOCK_DIVIDER*(stubs[n][6]+1),ARG2_REG);
+  emit_call((intptr_t)ftable);
+#endif
+
+  restore_regs(reglist);
+
+  emit_cmpmem_imm((intptr_t)&g_dev.r4300.new_dynarec_hot_state.pending_exception,0);
+  intptr_t jaddr=(intptr_t)out;
+  emit_jeq(0);
+
+  if(!ds) load_all_consts(regs[i].regmap_entry,regs[i].was32,regs[i].wasdirty,regs[i].wasconst,i);
+  wb_dirtys(i_regs->regmap_entry,i_regs->was32,i_regs->wasdirty);
+
+  emit_jmp((intptr_t)&do_interrupt);
+  set_jump_target(jaddr,(intptr_t)out);
+
+  if(cc>=0) {
+    emit_loadreg(CCREG,cc);
+  }
+  emit_jmp(stubs[n][2]); // return address
+}
+
+static void inline_writestub(int type, int i, u_int addr_const, char addr, struct regstat *i_regs, int target, int adj, u_int reglist)
+{
+  assem_debug("inline_writestub");
+  int rth=get_reg(i_regs->regmap,target|64);
+  int rt=get_reg(i_regs->regmap,target);
+  assert(rt>=0);
+
+#if NEW_DYNAREC <= NEW_DYNAREC_X64
+  if(addr_const)
+    emit_writeword_imm(addr_const,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.address);
+  else
+#endif
+  {
+    assert(addr>=0);
+    emit_writeword(addr,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.address);
+  }
+
+  intptr_t ftable=0;
+  if(type==STOREB_STUB){
+    ftable=(intptr_t)write_byte_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
+  }
+  else if(type==STOREH_STUB){
+    ftable=(intptr_t)write_hword_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
+  }
+  else if(type==STOREW_STUB){
+    ftable=(intptr_t)write_word_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
+  }
+  else if(type==STORED_STUB){
+    ftable=(intptr_t)write_dword_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
+    emit_writeword(target?rth:rt,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
+  }
+  else if(type==STOREWL_STUB){
+    ftable=(intptr_t)SWL_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
+  }
+  else if(type==STOREWR_STUB){
+    ftable=(intptr_t)SWR_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
+  }
+  else if(type==STOREDL_STUB){
+    ftable=(intptr_t)SDL_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
+    emit_writeword(target?rth:rt,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
+  }
+  else if(type==STOREDR_STUB){
+    ftable=(intptr_t)SDR_new;
+    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
+    emit_writeword(target?rth:rt,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
+  }
+
+  int cc=get_reg(i_regs->regmap,CCREG);
+  if(cc>=0) {
+    emit_storereg(CCREG,cc);
+  }
+
+  save_regs(reglist);
+  int ds=i_regs!=&regs[i];
+
+#if NEW_DYNAREC == NEW_DYNAREC_X86
+  emit_pushimm(CLOCK_DIVIDER*(adj+1));
+  emit_pushimm((start+(i+1)*4)+ds);
+  emit_call((intptr_t)ftable);
+  emit_addimm(ESP,8,ESP);
+#else
+  emit_movimm((start+(i+1)*4)+ds,ARG1_REG);
+  emit_movimm(CLOCK_DIVIDER*(adj+1),ARG2_REG);
+  emit_call((intptr_t)ftable);
+#endif
+
+  restore_regs(reglist);
+
+  if(((signed int)addr_const>=(signed int)0xC0000000)||((addr_const>>16)==0xa430)||((addr_const>>16)==0x8430)) {
+    // Theoretically we can have a pagefault here, if the TLB has never
+    // been enabled and the address is outside the range 80000000..BFFFFFFF
+    // Write out the registers so the pagefault can be handled.  This is
+    // a very rare case and likely represents a bug.
+    emit_cmpmem_imm((intptr_t)&g_dev.r4300.new_dynarec_hot_state.pending_exception,0);
+    intptr_t jaddr=(intptr_t)out;
+    emit_jeq(0);
+
+    if(!ds) load_all_consts(regs[i].regmap_entry,regs[i].was32,regs[i].wasdirty,regs[i].wasconst,i);
+    wb_dirtys(i_regs->regmap_entry,i_regs->was32,i_regs->wasdirty);
+
+    emit_jmp((intptr_t)&do_interrupt);
+    set_jump_target(jaddr,(intptr_t)out);
+  }
+
+  if(cc>=0) {
+    emit_loadreg(CCREG,cc);
+  }
+}
+
+/**** Assemble ****/
 static void cop0_assemble(int i,struct regstat *i_regs)
 {
   u_int hr,reglist=0;
@@ -4963,7 +6927,7 @@ static void syscall_assemble(int i,struct regstat *i_regs)
   signed char ccreg=get_reg(i_regs->regmap,CCREG);
   assert(ccreg==HOST_CCREG);
   assert(!is_delayslot);
-  emit_movimm(start+i*4,EAX); // Get PC
+  emit_movimm(start+i*4,0); // Get PC
   emit_addimm(HOST_CCREG,CLOCK_DIVIDER*ccadj[i],HOST_CCREG); // CHECK: is this right?  There should probably be an extra cycle...
   emit_jmp((intptr_t)jump_syscall);
 }
@@ -5014,784 +6978,6 @@ static void ds_assemble(int i,struct regstat *i_regs)
       DebugMessage(M64MSG_VERBOSE, "Jump in the delay slot.  This is probably a bug.");
   }
   is_delayslot=0;
-}
-
-// Is the branch target a valid internal jump?
-static int internal_branch(uint64_t i_is32,int addr)
-{
-  if(addr&1) return 0; // Indirect (register) jump
-  if((u_int)addr>=start && (u_int)addr<start+slen*4-4)
-  {
-    int t=(addr-start)>>2;
-    // Delay slots are not valid branch targets
-    //if(t>0&&(itype[t-1]==RJUMP||itype[t-1]==UJUMP||itype[t-1]==CJUMP||itype[t-1]==SJUMP||itype[t-1]==FJUMP)) return 0;
-    // 64 -> 32 bit transition requires a recompile
-    /*if(is32[t]&~unneeded_reg_upper[t]&~i_is32)
-    {
-      if(requires_32bit[t]&~i_is32) DebugMessage(M64MSG_VERBOSE, "optimizable: no");
-      else DebugMessage(M64MSG_VERBOSE, "optimizable: yes");
-    }*/
-    //if(is32[t]&~unneeded_reg_upper[t]&~i_is32) return 0;
-    if(requires_32bit[t]&~i_is32) return 0;
-    else return 1;
-  }
-  return 0;
-}
-
-#ifndef wb_invalidate
-static void wb_invalidate(signed char pre[],signed char entry[],uint64_t dirty,uint64_t is32,
-  uint64_t u,uint64_t uu)
-{
-  int hr;
-  for(hr=0;hr<HOST_REGS;hr++) {
-    if(hr!=EXCLUDE_REG) {
-      if(pre[hr]!=entry[hr]) {
-        if(pre[hr]>=0) {
-          if((dirty>>hr)&1) {
-            if(get_reg(entry,pre[hr])<0) {
-              if(pre[hr]<64) {
-                if(!((u>>pre[hr])&1)) {
-                  emit_storereg(pre[hr],hr);
-                  if( ((is32>>pre[hr])&1) && !((uu>>pre[hr])&1) ) {
-                    emit_sarimm(hr,31,hr);
-                    emit_storereg(pre[hr]|64,hr);
-                  }
-                }
-              }else{
-                if(!((uu>>(pre[hr]&63))&1) && !((is32>>(pre[hr]&63))&1)) {
-                  emit_storereg(pre[hr],hr);
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-  // Move from one register to another (no writeback)
-  for(hr=0;hr<HOST_REGS;hr++) {
-    if(hr!=EXCLUDE_REG) {
-      if(pre[hr]!=entry[hr]) {
-        if(pre[hr]>=0&&(pre[hr]&63)<TEMPREG) {
-          int nr;
-          if((nr=get_reg(entry,pre[hr]))>=0) {
-            #ifdef NATIVE_64
-            if(pre[hr]>=INVCP) emit_mov64(hr,nr);
-            else
-            #endif
-            emit_mov(hr,nr);
-          }
-        }
-      }
-    }
-  }
-}
-#endif
-
-// Load the specified registers
-// This only loads the registers given as arguments because
-// we don't want to load things that will be overwritten
-static void load_regs(signed char entry[],signed char regmap[],int is32,int rs1,int rs2)
-{
-  int hr;
-  // Load 32-bit regs
-  for(hr=0;hr<HOST_REGS;hr++) {
-    if(hr!=EXCLUDE_REG&&regmap[hr]>=0) {
-      if(entry[hr]!=regmap[hr]) {
-        if(regmap[hr]==rs1||regmap[hr]==rs2)
-        {
-          if(regmap[hr]==0) {
-            emit_zeroreg(hr);
-          }
-          else
-          {
-            emit_loadreg(regmap[hr],hr);
-          }
-        }
-      }
-    }
-  }
-  //Load 64-bit regs
-  for(hr=0;hr<HOST_REGS;hr++) {
-    if(hr!=EXCLUDE_REG&&regmap[hr]>=0) {
-      if(entry[hr]!=regmap[hr]) {
-        if(regmap[hr]-64==rs1||regmap[hr]-64==rs2)
-        {
-          assert(regmap[hr]!=64);
-          if((is32>>(regmap[hr]&63))&1) {
-            int lr=get_reg(regmap,regmap[hr]-64);
-            if(lr>=0)
-              emit_sarimm(lr,31,hr);
-            else
-              emit_loadreg(regmap[hr],hr);
-          }
-          else
-          {
-            emit_loadreg(regmap[hr],hr);
-          }
-        }
-      }
-    }
-  }
-}
-
-// Load registers prior to the start of a loop
-// so that they are not loaded within the loop
-static void loop_preload(signed char pre[],signed char entry[])
-{
-  int hr;
-  for(hr=0;hr<HOST_REGS;hr++) {
-    if(hr!=EXCLUDE_REG) {
-      if(pre[hr]!=entry[hr]) {
-        if(entry[hr]>=0) {
-          if(get_reg(pre,entry[hr])<0) {
-            assem_debug("loop preload:");
-            //DebugMessage(M64MSG_VERBOSE, "loop preload: %d",hr);
-            if(entry[hr]==0) {
-              emit_zeroreg(hr);
-            }
-            else if(entry[hr]<TEMPREG)
-            {
-              emit_loadreg(entry[hr],hr);
-            }
-            else if(entry[hr]-64<TEMPREG)
-            {
-              emit_loadreg(entry[hr],hr);
-            }
-          }
-        }
-      }
-    }
-  }
-}
-
-// Generate address for load/store instruction
-static void address_generation(int i,struct regstat *i_regs,signed char entry[])
-{
-  if(itype[i]==LOAD||itype[i]==LOADLR||itype[i]==STORE||itype[i]==STORELR||itype[i]==C1LS) {
-    int ra=0;
-    int agr=AGEN1+(i&1);
-    int mgr=MGEN1+(i&1);
-    if(itype[i]==LOAD) {
-      ra=get_reg(i_regs->regmap,agr);
-      if(ra<0) ra=get_reg(i_regs->regmap,-1);
-      assert(ra>=0);
-    }
-    if(itype[i]==LOADLR) {
-      ra=get_reg(i_regs->regmap,FTEMP);
-    }
-    if(itype[i]==STORE||itype[i]==STORELR) {
-      ra=get_reg(i_regs->regmap,agr);
-      if(ra<0) ra=get_reg(i_regs->regmap,-1);
-    }
-    if(itype[i]==C1LS) {
-      if (opcode[i]==0x31||opcode[i]==0x35) // LWC1/LDC1
-        ra=get_reg(i_regs->regmap,FTEMP);
-      else { // SWC1/SDC1
-        ra=get_reg(i_regs->regmap,agr);
-        if(ra<0) ra=get_reg(i_regs->regmap,-1);
-      }
-    }
-    int rs=get_reg(i_regs->regmap,rs1[i]);
-    int rm=get_reg(i_regs->regmap,TLREG);
-    if(ra>=0) {
-      int offset=imm[i];
-      int c=(i_regs->wasconst>>rs)&1;
-      if(rs1[i]==0) {
-        // Using r0 as a base address
-        /*if(rm>=0) {
-          if(!entry||entry[rm]!=mgr) {
-            generate_map_const(offset,rm);
-          } // else did it in the previous cycle
-        }*/
-        if(!entry||entry[ra]!=agr) {
-          emit_movimm(offset,ra);
-        } // else did it in the previous cycle
-      }
-      else if(rs<0) {
-        if(!entry||entry[ra]!=rs1[i])
-          emit_loadreg(rs1[i],ra);
-        //if(!entry||entry[ra]!=rs1[i])
-        //  DebugMessage(M64MSG_VERBOSE, "poor load scheduling!");
-      }
-      else if(c) {
-        if(rm>=0) {
-          if(!entry||entry[rm]!=mgr) {
-            if(itype[i]==STORE||itype[i]==STORELR||opcode[i]==0x39||opcode[i]==0x3D) {
-              // Stores to memory go thru the mapper to detect self-modifying
-              // code, loads don't.
-              if((unsigned int)(constmap[i][rs]+offset)>=0xC0000000 ||
-                 (unsigned int)(constmap[i][rs]+offset)<0x80800000 )
-                generate_map_const(constmap[i][rs]+offset,rm);
-            }else{
-              if((signed int)(constmap[i][rs]+offset)>=(signed int)0xC0000000)
-                generate_map_const(constmap[i][rs]+offset,rm);
-            }
-          }
-        }
-        
-        if(!entry||entry[ra]!=agr) {
-          int load=0;
-          #ifndef INTERPRET_LOAD
-          load|=(itype[i]==LOAD);
-          #endif
-          #ifndef INTERPRET_C1LS
-          load|=(opcode[i]==0x31||opcode[i]==0x35);
-          #endif
-          #ifndef INTERPRET_LOADLR
-          load|=(opcode[i]==0x22||opcode[i]==0x26||opcode[i]==0x1a||opcode[i]==0x1b);
-          #endif
-          
-          #ifdef HOST_IMM_ADDR32
-          // on x86 when source reg is constant
-          // LB/LBU/LH/LHU/LW/LWU/LD/LWC1/LDC1 doesn't need address generation when not using tlb
-          // SDL/SDR/SB/SH/SW/SD/SWC1/SDC1 always need address generation
-          // ROREG not required
-          if(!load||(using_tlb&&((signed int)constmap[i][rs]+offset)>=(signed int)0xC0000000))
-          #endif
-          #if defined(RAM_OFFSET) && !defined(NATIVE_64) 
-          // on arm address generation is always required
-          // When LB/LBU/LH/LHU/LW/LWU/LD/LWC1/LDC1 loading from rdram, add rdram address to load address to avoid loading ROREG
-          // ROREG only required for SDL/SDR/SB/SH/SW/SD/SWC1/SDC1 and when not loading from rdram
-          if(load&&(signed int)constmap[i][rs]+offset<(signed int)0x80800000)
-            emit_movimm(constmap[i][rs]+offset+(intptr_t)g_dev.rdram.dram-(intptr_t)0x80000000,ra);
-          else
-          #endif
-          // on arm64 and x64 when source reg is constant
-          // address generation is always required
-          // ROREG is always required 
-          emit_movimm(constmap[i][rs]+offset,ra);
-          
-        } // else did it in the previous cycle
-      }
-      if(offset&&!c&&rs1[i]) {
-        if(rs>=0) {
-          emit_addimm(rs,offset,ra);
-        }else{
-          emit_addimm(ra,offset,ra);
-        }
-      }
-    }
-  }
-  // Preload constants for next instruction
-  if(itype[i+1]==LOAD||itype[i+1]==LOADLR||itype[i+1]==STORE||itype[i+1]==STORELR||itype[i+1]==C1LS) {
-    int agr,ra;
-    #if (NEW_DYNAREC!=NEW_DYNAREC_X86) && (NEW_DYNAREC!=NEW_DYNAREC_X64)
-    // Mapper entry
-    agr=MGEN1+((i+1)&1);
-    ra=get_reg(i_regs->regmap,agr);
-    if(ra>=0) {
-      int rs=get_reg(regs[i+1].regmap,rs1[i+1]);
-      int offset=imm[i+1];
-      int c=(regs[i+1].wasconst>>rs)&1;
-      if(c) {
-        if(itype[i+1]==STORE||itype[i+1]==STORELR||opcode[i+1]==0x39||opcode[i+1]==0x3D) {
-          // Stores to memory go thru the mapper to detect self-modifying
-          // code, loads don't.
-          if((unsigned int)(constmap[i+1][rs]+offset)>=0xC0000000 ||
-             (unsigned int)(constmap[i+1][rs]+offset)<0x80800000 )
-            generate_map_const(constmap[i+1][rs]+offset,ra);
-        }else{
-          if((signed int)(constmap[i+1][rs]+offset)>=(signed int)0xC0000000)
-            generate_map_const(constmap[i+1][rs]+offset,ra);
-        }
-      }
-      /*else if(rs1[i]==0) {
-        generate_map_const(offset,ra);
-      }*/
-    }
-    #endif
-    // Actual address
-    agr=AGEN1+((i+1)&1);
-    ra=get_reg(i_regs->regmap,agr);
-    if(ra>=0) {
-      int rs=get_reg(regs[i+1].regmap,rs1[i+1]);
-      int offset=imm[i+1];
-      int c=(regs[i+1].wasconst>>rs)&1;
-      if(c) {
-        int load=0;
-        #ifndef INTERPRET_LOAD
-        load|=(itype[i+1]==LOAD);
-        #endif
-        #ifndef INTERPRET_C1LS
-        load|=(opcode[i+1]==0x31||opcode[i+1]==0x35);
-        #endif
-        #ifndef INTERPRET_LOADLR
-        load|=(opcode[i+1]==0x22||opcode[i+1]==0x26||opcode[i+1]==0x1a||opcode[i+1]==0x1b);
-        #endif
-        
-        #ifdef HOST_IMM_ADDR32
-        if(!load||(using_tlb&&((signed int)constmap[i+1][rs]+offset)>=(signed int)0xC0000000))
-        #endif
-        #if defined(RAM_OFFSET) && !defined(NATIVE_64)
-        if(load&&(signed int)constmap[i+1][rs]+offset<(signed int)0x80800000) 
-          emit_movimm(constmap[i+1][rs]+offset+(intptr_t)g_dev.rdram.dram-(intptr_t)0x80000000,ra);
-        else
-        #endif
-        emit_movimm(constmap[i+1][rs]+offset,ra);
-      }
-      else if(rs1[i+1]==0) {
-        // Using r0 as a base address
-        emit_movimm(offset,ra);
-      }
-    }
-  }
-}
-
-static int get_final_value(int hr, int i, int *value)
-{
-  int reg=regs[i].regmap[hr];
-  while(i<slen-1) {
-    if(regs[i+1].regmap[hr]!=reg) break;
-    if(!((regs[i+1].isconst>>hr)&1)) break;
-    if(bt[i+1]) break;
-    i++;
-  }
-  if(i<slen-1) {
-    if(itype[i]==UJUMP||itype[i]==RJUMP||itype[i]==CJUMP||itype[i]==SJUMP) {
-      *value=constmap[i][hr];
-      return 1;
-    }
-  }
-  *value=constmap[i][hr];
-  //DebugMessage(M64MSG_VERBOSE, "c=%x",(int)constmap[i][hr]);
-  if(i==slen-1) return 1;
-  if(reg<64) {
-    return !((unneeded_reg[i+1]>>reg)&1);
-  }else{
-    return !((unneeded_reg_upper[i+1]>>reg)&1);
-  }
-}
-
-// Load registers with known constants
-static void load_consts(signed char pre[],signed char regmap[],int is32,int i)
-{
-  int hr;
-  // Load 32-bit regs
-  for(hr=0;hr<HOST_REGS;hr++) {
-    if(hr!=EXCLUDE_REG&&regmap[hr]>=0) {
-      //if(entry[hr]!=regmap[hr]) {
-      if(i==0||!((regs[i-1].isconst>>hr)&1)||pre[hr]!=regmap[hr]||bt[i]) {
-        if(((regs[i].isconst>>hr)&1)&&regmap[hr]<64&&regmap[hr]>0) {
-          int value;
-          if(get_final_value(hr,i,&value)) {
-            if(value==0) {
-              emit_zeroreg(hr);
-            }
-            else {
-              emit_movimm(value,hr);
-            }
-          }
-        }
-      }
-    }
-  }
-  // Load 64-bit regs
-  for(hr=0;hr<HOST_REGS;hr++) {
-    if(hr!=EXCLUDE_REG&&regmap[hr]>=0) {
-      //if(entry[hr]!=regmap[hr]) {
-      if(i==0||!((regs[i-1].isconst>>hr)&1)||pre[hr]!=regmap[hr]||bt[i]) {
-        if(((regs[i].isconst>>hr)&1)&&regmap[hr]>64) {
-          if((is32>>(regmap[hr]&63))&1) {
-            int lr=get_reg(regmap,regmap[hr]-64);
-            assert(lr>=0);
-            emit_sarimm(lr,31,hr);
-          }
-          else
-          {
-            int value;
-            if(get_final_value(hr,i,&value)) {
-              if(value==0) {
-                emit_zeroreg(hr);
-              }
-              else {
-                emit_movimm(value,hr);
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-}
-static void load_all_consts(signed char regmap[],int is32,u_int dirty,u_int isconst,int i)
-{
-  int hr;
-  // Load 32-bit regs
-  for(hr=0;hr<HOST_REGS;hr++) {
-    if(hr!=EXCLUDE_REG&&regmap[hr]>=0&&((dirty>>hr)&1)) {
-      if(((isconst>>hr)&1)&&regmap[hr]<64&&regmap[hr]>0) {
-        int value=constmap[i][hr];
-        if(value==0) {
-          emit_zeroreg(hr);
-        }
-        else {
-          emit_movimm(value,hr);
-        }
-      }
-    }
-  }
-  // Load 64-bit regs
-  for(hr=0;hr<HOST_REGS;hr++) {
-    if(hr!=EXCLUDE_REG&&regmap[hr]>=0&&((dirty>>hr)&1)) {
-      if(((isconst>>hr)&1)&&regmap[hr]>64) {
-        if((is32>>(regmap[hr]&63))&1) {
-          int lr=get_reg(regmap,regmap[hr]-64);
-          assert(lr>=0);
-          emit_sarimm(lr,31,hr);
-        }
-        else
-        {
-          int value=constmap[i][hr];
-          if(value==0) {
-            emit_zeroreg(hr);
-          }
-          else {
-            emit_movimm(value,hr);
-          }
-        }
-      }
-    }
-  }
-}
-
-// Write out all dirty registers (except cycle count)
-static void wb_dirtys(signed char i_regmap[],uint64_t i_is32,uint64_t i_dirty)
-{
-  int hr;
-  for(hr=0;hr<HOST_REGS;hr++) {
-    if(hr!=EXCLUDE_REG) {
-      if(((i_regmap[hr]&63)>0)&&((i_regmap[hr]&63)<CSREG)) {
-        if((i_dirty>>hr)&1) {
-          if(i_regmap[hr]<64) {
-            emit_storereg(i_regmap[hr],hr);
-            if(((i_is32>>i_regmap[hr])&1)) {
-              #ifdef DESTRUCTIVE_WRITEBACK
-              emit_sarimm(hr,31,hr);
-              emit_storereg(i_regmap[hr]|64,hr);
-              #else
-              emit_sarimm(hr,31,HOST_TEMPREG);
-              emit_storereg(i_regmap[hr]|64,HOST_TEMPREG);
-              #endif
-            }
-          }else{
-            if(!((i_is32>>(i_regmap[hr]&63))&1)) {
-              emit_storereg(i_regmap[hr],hr);
-            }
-          }
-        }
-      }
-    }
-  }
-}
-// Write out dirty registers that we need to reload (pair with load_needed_regs)
-// This writes the registers not written by store_regs_bt
-static void wb_needed_dirtys(signed char i_regmap[],uint64_t i_is32,uint64_t i_dirty,int addr)
-{
-  int hr;
-  int t=(addr-start)>>2;
-  for(hr=0;hr<HOST_REGS;hr++) {
-    if(hr!=EXCLUDE_REG) {
-      if(((i_regmap[hr]&63)>0)&&((i_regmap[hr]&63)<CSREG)) {
-        if(i_regmap[hr]==regs[t].regmap_entry[hr] && ((regs[t].dirty>>hr)&1) && !(((i_is32&~regs[t].was32&~unneeded_reg_upper[t])>>(i_regmap[hr]&63))&1)) {
-          if((i_dirty>>hr)&1) {
-            if(i_regmap[hr]<64) {
-              emit_storereg(i_regmap[hr],hr);
-              if(((i_is32>>i_regmap[hr])&1)) {
-                #ifdef DESTRUCTIVE_WRITEBACK
-                emit_sarimm(hr,31,hr);
-                emit_storereg(i_regmap[hr]|64,hr);
-                #else
-                emit_sarimm(hr,31,HOST_TEMPREG);
-                emit_storereg(i_regmap[hr]|64,HOST_TEMPREG);
-                #endif
-              }
-            }else{
-              if(!((i_is32>>(i_regmap[hr]&63))&1)) {
-                emit_storereg(i_regmap[hr],hr);
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-}
-
-// Load all registers (except cycle count)
-static void load_all_regs(signed char i_regmap[])
-{
-  int hr;
-  for(hr=0;hr<HOST_REGS;hr++) {
-    if(hr!=EXCLUDE_REG) {
-      if(i_regmap[hr]==0) {
-        emit_zeroreg(hr);
-      }
-      else
-      if(i_regmap[hr]>0 && (i_regmap[hr]&63)<TEMPREG && i_regmap[hr]!=CCREG)
-      {
-        emit_loadreg(i_regmap[hr],hr);
-      }
-    }
-  }
-}
-
-// Load all current registers also needed by next instruction
-static void load_needed_regs(signed char i_regmap[],signed char next_regmap[])
-{
-  int hr;
-  for(hr=0;hr<HOST_REGS;hr++) {
-    if(hr!=EXCLUDE_REG) {
-      if(get_reg(next_regmap,i_regmap[hr])>=0) {
-        if(i_regmap[hr]==0) {
-          emit_zeroreg(hr);
-        }
-        else
-        if(i_regmap[hr]>0 && (i_regmap[hr]&63)<TEMPREG && i_regmap[hr]!=CCREG)
-        {
-          emit_loadreg(i_regmap[hr],hr);
-        }
-      }
-    }
-  }
-}
-
-// Load all regs, storing cycle count if necessary
-static void load_regs_entry(int t)
-{
-  int hr;
-  if(is_ds[t]) emit_addimm(HOST_CCREG,CLOCK_DIVIDER,HOST_CCREG);
-  else if(ccadj[t]) emit_addimm(HOST_CCREG,-ccadj[t]*CLOCK_DIVIDER,HOST_CCREG);
-  if(regs[t].regmap_entry[HOST_CCREG]!=CCREG) {
-    emit_storereg(CCREG,HOST_CCREG);
-  }
-  // Load 32-bit regs
-  for(hr=0;hr<HOST_REGS;hr++) {
-    if(regs[t].regmap_entry[hr]>=0&&regs[t].regmap_entry[hr]<TEMPREG) {
-      if(regs[t].regmap_entry[hr]==0) {
-        emit_zeroreg(hr);
-      }
-      else if(regs[t].regmap_entry[hr]!=CCREG)
-      {
-        emit_loadreg(regs[t].regmap_entry[hr],hr);
-      }
-    }
-  }
-  // Load 64-bit regs
-  for(hr=0;hr<HOST_REGS;hr++) {
-    if(regs[t].regmap_entry[hr]>=64&&regs[t].regmap_entry[hr]<TEMPREG+64) {
-      assert(regs[t].regmap_entry[hr]!=64);
-      if((regs[t].was32>>(regs[t].regmap_entry[hr]&63))&1) {
-        int lr=get_reg(regs[t].regmap_entry,regs[t].regmap_entry[hr]-64);
-        if(lr<0) {
-          emit_loadreg(regs[t].regmap_entry[hr],hr);
-        }
-        else
-        {
-          emit_sarimm(lr,31,hr);
-        }
-      }
-      else
-      {
-        emit_loadreg(regs[t].regmap_entry[hr],hr);
-      }
-    }
-  }
-}
-
-// Store dirty registers prior to branch
-static void store_regs_bt(signed char i_regmap[],uint64_t i_is32,uint64_t i_dirty,int addr)
-{
-  if(internal_branch(i_is32,addr))
-  {
-    int t=(addr-start)>>2;
-    int hr;
-    for(hr=0;hr<HOST_REGS;hr++) {
-      if(hr!=EXCLUDE_REG) {
-        if(((i_regmap[hr]&63)>0)&&((i_regmap[hr]&63)<CSREG)) {
-          if(i_regmap[hr]!=regs[t].regmap_entry[hr] || !((regs[t].dirty>>hr)&1) || (((i_is32&~regs[t].was32&~unneeded_reg_upper[t])>>(i_regmap[hr]&63))&1)) {
-            if((i_dirty>>hr)&1) {
-              if(i_regmap[hr]<64) {
-                if(!((unneeded_reg[t]>>i_regmap[hr])&1)) {
-                  emit_storereg(i_regmap[hr],hr);
-                  if(((i_is32>>i_regmap[hr])&1) && !((unneeded_reg_upper[t]>>i_regmap[hr])&1)) {
-                    #ifdef DESTRUCTIVE_WRITEBACK
-                    emit_sarimm(hr,31,hr);
-                    emit_storereg(i_regmap[hr]|64,hr);
-                    #else
-                    emit_sarimm(hr,31,HOST_TEMPREG);
-                    emit_storereg(i_regmap[hr]|64,HOST_TEMPREG);
-                    #endif
-                  }
-                }
-              }else{
-                if(!((i_is32>>(i_regmap[hr]&63))&1) && !((unneeded_reg_upper[t]>>(i_regmap[hr]&63))&1)) {
-                  emit_storereg(i_regmap[hr],hr);
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-  else
-  {
-    // Branch out of this block, write out all dirty regs
-    wb_dirtys(i_regmap,i_is32,i_dirty);
-  }
-}
-
-// Load all needed registers for branch target
-static void load_regs_bt(signed char i_regmap[],uint64_t i_is32,uint64_t i_dirty,int addr)
-{
-  //if(addr>=start && addr<(start+slen*4))
-  if(internal_branch(i_is32,addr))
-  {
-    int t=(addr-start)>>2;
-    int hr;
-    // Store the cycle count before loading something else
-    if(i_regmap[HOST_CCREG]!=CCREG) {
-      assert(i_regmap[HOST_CCREG]==-1);
-    }
-    if(regs[t].regmap_entry[HOST_CCREG]!=CCREG) {
-      emit_storereg(CCREG,HOST_CCREG);
-    }
-    // Load 32-bit regs
-    for(hr=0;hr<HOST_REGS;hr++) {
-      if(hr!=EXCLUDE_REG&&regs[t].regmap_entry[hr]>=0&&regs[t].regmap_entry[hr]<TEMPREG) {
-        #ifdef DESTRUCTIVE_WRITEBACK
-        if(i_regmap[hr]!=regs[t].regmap_entry[hr] || ( !((regs[t].dirty>>hr)&1) && ((i_dirty>>hr)&1) && (((i_is32&~unneeded_reg_upper[t])>>i_regmap[hr])&1) ) || (((i_is32&~regs[t].was32&~unneeded_reg_upper[t])>>(i_regmap[hr]&63))&1)) {
-        #else
-        if(i_regmap[hr]!=regs[t].regmap_entry[hr] ) {
-        #endif
-          if(regs[t].regmap_entry[hr]==0) {
-            emit_zeroreg(hr);
-          }
-          else if(regs[t].regmap_entry[hr]!=CCREG)
-          {
-            emit_loadreg(regs[t].regmap_entry[hr],hr);
-          }
-        }
-      }
-    }
-    //Load 64-bit regs
-    for(hr=0;hr<HOST_REGS;hr++) {
-      if(hr!=EXCLUDE_REG&&regs[t].regmap_entry[hr]>=64&&regs[t].regmap_entry[hr]<TEMPREG+64) {
-        if(i_regmap[hr]!=regs[t].regmap_entry[hr]) {
-          assert(regs[t].regmap_entry[hr]!=64);
-          if((i_is32>>(regs[t].regmap_entry[hr]&63))&1) {
-            int lr=get_reg(regs[t].regmap_entry,regs[t].regmap_entry[hr]-64);
-            if(lr<0) {
-              emit_loadreg(regs[t].regmap_entry[hr],hr);
-            }
-            else
-            {
-              emit_sarimm(lr,31,hr);
-            }
-          }
-          else
-          {
-            emit_loadreg(regs[t].regmap_entry[hr],hr);
-          }
-        }
-        else if((i_is32>>(regs[t].regmap_entry[hr]&63))&1) {
-          int lr=get_reg(regs[t].regmap_entry,regs[t].regmap_entry[hr]-64);
-          if(lr<0) {
-            emit_loadreg(regs[t].regmap_entry[hr],hr);
-          }
-          else
-          {
-            emit_sarimm(lr,31,hr);
-          }
-        }
-      }
-    }
-  }
-}
-
-static int match_bt(signed char i_regmap[],uint64_t i_is32,uint64_t i_dirty,int addr)
-{
-  if((u_int)addr>=start && (u_int)addr<start+slen*4-4)
-  {
-    int t=(addr-start)>>2;
-    int hr;
-    if(regs[t].regmap_entry[HOST_CCREG]!=CCREG) return 0;
-    for(hr=0;hr<HOST_REGS;hr++)
-    {
-      if(hr!=EXCLUDE_REG)
-      {
-        if(i_regmap[hr]!=regs[t].regmap_entry[hr])
-        {
-          if(regs[t].regmap_entry[hr]>=0&&(regs[t].regmap_entry[hr]|64)<TEMPREG+64)
-          {
-            return 0;
-          }
-          else 
-          if((i_dirty>>hr)&1)
-          {
-            if(i_regmap[hr]<TEMPREG)
-            {
-              if(!((unneeded_reg[t]>>i_regmap[hr])&1))
-                return 0;
-            }
-            else if(i_regmap[hr]>=64&&i_regmap[hr]<TEMPREG+64)
-            {
-              if(!((unneeded_reg_upper[t]>>(i_regmap[hr]&63))&1))
-                return 0;
-            }
-          }
-        }
-        else // Same register but is it 32-bit or dirty?
-        if(i_regmap[hr]>=0)
-        {
-          if(!((regs[t].dirty>>hr)&1))
-          {
-            if((i_dirty>>hr)&1)
-            {
-              if(!((unneeded_reg[t]>>i_regmap[hr])&1))
-              {
-                //DebugMessage(M64MSG_VERBOSE, "%x: dirty no match",addr);
-                return 0;
-              }
-            }
-          }
-          if((((regs[t].was32^i_is32)&~unneeded_reg_upper[t])>>(i_regmap[hr]&63))&1)
-          {
-            //DebugMessage(M64MSG_VERBOSE, "%x: is32 no match",addr);
-            return 0;
-          }
-        }
-      }
-    }
-    //if(is32[t]&~unneeded_reg_upper[t]&~i_is32) return 0;
-    if(requires_32bit[t]&~i_is32) return 0;
-    // Delay slots are not valid branch targets
-    //if(t>0&&(itype[t-1]==RJUMP||itype[t-1]==UJUMP||itype[t-1]==CJUMP||itype[t-1]==SJUMP||itype[t-1]==FJUMP)) return 0;
-    // Delay slots require additional processing, so do not match
-    if(is_ds[t]) return 0;
-  }
-  else
-  {
-    int hr;
-    for(hr=0;hr<HOST_REGS;hr++)
-    {
-      if(hr!=EXCLUDE_REG)
-      {
-        if(i_regmap[hr]>=0)
-        {
-          if(hr!=HOST_CCREG||i_regmap[hr]!=CCREG)
-          {
-            if((i_dirty>>hr)&1)
-            {
-              return 0;
-            }
-          }
-        }
-      }
-    }
-  }
-  return 1;
 }
 
 // Used when a branch jumps into the delay slot of another branch
@@ -5867,332 +7053,6 @@ static void ds_assemble_entry(int i)
   ccadj[t]=cc;
 }
 
-static void do_cc(int i,signed char i_regmap[],int *adj,int addr,int taken,int invert)
-{
-  int count;
-  intptr_t jaddr;
-  intptr_t idle=0;
-  if(itype[i]==RJUMP)
-  {
-    *adj=0;
-  }
-  //if(ba[i]>=start && ba[i]<(start+slen*4))
-  if(internal_branch(branch_regs[i].is32,ba[i]))
-  {
-    int t=(ba[i]-start)>>2;
-    if(is_ds[t]) *adj=-1; // Branch into delay slot adds an extra cycle
-    else *adj=ccadj[t];
-  }
-  else
-  {
-    *adj=0;
-  }
-  count=ccadj[i];
-  if(taken==TAKEN && i==(ba[i]-start)>>2 && source[i+1]==0) {
-    // Idle loop
-    idle=(intptr_t)out;
-    emit_test(HOST_CCREG,HOST_CCREG);
-#if NEW_DYNAREC >= NEW_DYNAREC_ARM
-    emit_cmovs_imm(0,HOST_CCREG);
-#else
-    emit_cmovs(&const_zero,HOST_CCREG);
-#endif
-    emit_addimm(HOST_CCREG,CLOCK_DIVIDER*2,HOST_CCREG);
-    jaddr=(intptr_t)out;
-    emit_jmp(0);
-  }
-  else if(*adj==0||invert) {
-    emit_addimm_and_set_flags(CLOCK_DIVIDER*(count+2),HOST_CCREG);
-    jaddr=(intptr_t)out;
-    emit_jns(0);
-  }
-  else
-  {
-    emit_cmpimm(HOST_CCREG,-(int)CLOCK_DIVIDER*(count+2));
-    jaddr=(intptr_t)out;
-    emit_jns(0);
-  }
-  add_stub(CC_STUB,jaddr,idle?idle:(intptr_t)out,(*adj==0||invert||idle)?0:(count+2),i,addr,taken,0);
-}
-
-static void do_ccstub(int n)
-{
-  literal_pool(256);
-  assem_debug("do_ccstub %x",start+stubs[n][4]*4);
-  set_jump_target(stubs[n][1],(intptr_t)out);
-  int i=stubs[n][4];
-  if(stubs[n][6]==NULLDS) {
-    // Delay slot instruction is nullified ("likely" branch)
-    wb_dirtys(regs[i].regmap,regs[i].is32,regs[i].dirty);
-  }
-  else if(stubs[n][6]!=TAKEN) {
-    wb_dirtys(branch_regs[i].regmap,branch_regs[i].is32,branch_regs[i].dirty);
-  }
-  else {
-    if(internal_branch(branch_regs[i].is32,ba[i]))
-      wb_needed_dirtys(branch_regs[i].regmap,branch_regs[i].is32,branch_regs[i].dirty,ba[i]);
-  }
-  if(stubs[n][5]!=-1)
-  {
-    // Save PC as return address
-    emit_movimm(stubs[n][5],EAX);
-    emit_writeword(EAX,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.pcaddr);
-  }
-  else
-  {
-    // Return address depends on which way the branch goes
-    if(itype[i]==CJUMP||itype[i]==SJUMP||itype[i]==FJUMP)
-    {
-      int s1l=get_reg(branch_regs[i].regmap,rs1[i]);
-      int s1h=get_reg(branch_regs[i].regmap,rs1[i]|64);
-      int s2l=get_reg(branch_regs[i].regmap,rs2[i]);
-      int s2h=get_reg(branch_regs[i].regmap,rs2[i]|64);
-      if(rs1[i]==0)
-      {
-        s1l=s2l;s1h=s2h;
-        s2l=s2h=-1;
-      }
-      else if(rs2[i]==0)
-      {
-        s2l=s2h=-1;
-      }
-      if((branch_regs[i].is32>>rs1[i])&(branch_regs[i].is32>>rs2[i])&1) {
-        s1h=s2h=-1;
-      }
-      assert(s1l>=0);
-      #ifdef DESTRUCTIVE_WRITEBACK
-      if(rs1[i]) {
-        if((branch_regs[i].dirty>>s1l)&(branch_regs[i].is32>>rs1[i])&1)
-          emit_loadreg(rs1[i],s1l);
-      } 
-      else {
-        if((branch_regs[i].dirty>>s1l)&(branch_regs[i].is32>>rs2[i])&1)
-          emit_loadreg(rs2[i],s1l);
-      }
-      if(s2l>=0)
-        if((branch_regs[i].dirty>>s2l)&(branch_regs[i].is32>>rs2[i])&1)
-          emit_loadreg(rs2[i],s2l);
-      #endif
-      int hr=0;
-      int addr,alt,ntaddr;
-      while(hr<HOST_REGS)
-      {
-        if(hr!=EXCLUDE_REG && hr!=HOST_CCREG &&
-           (branch_regs[i].regmap[hr]&63)!=rs1[i] &&
-           (branch_regs[i].regmap[hr]&63)!=rs2[i] )
-        {
-          addr=hr++;break;
-        }
-        hr++;
-      }
-      while(hr<HOST_REGS)
-      {
-        if(hr!=EXCLUDE_REG && hr!=HOST_CCREG &&
-           (branch_regs[i].regmap[hr]&63)!=rs1[i] &&
-           (branch_regs[i].regmap[hr]&63)!=rs2[i] )
-        {
-          alt=hr++;break;
-        }
-        hr++;
-      }
-      if((opcode[i]&0x2E)==6) // BLEZ/BGTZ needs another register
-      {
-        while(hr<HOST_REGS)
-        {
-          if(hr!=EXCLUDE_REG && hr!=HOST_CCREG &&
-             (branch_regs[i].regmap[hr]&63)!=rs1[i] &&
-             (branch_regs[i].regmap[hr]&63)!=rs2[i] )
-          {
-            ntaddr=hr;break;
-          }
-          hr++;
-        }
-        assert(hr<HOST_REGS);
-      }
-      if((opcode[i]&0x2f)==4) // BEQ
-      {
-        #ifdef HAVE_CMOV_IMM
-        if(s1h<0) {
-          if(s2l>=0) emit_cmp(s1l,s2l);
-          else emit_test(s1l,s1l);
-          emit_cmov2imm_e_ne_compact(ba[i],start+i*4+8,addr);
-        }
-        else
-        #endif
-        {
-          emit_mov2imm_compact(ba[i],addr,start+i*4+8,alt);
-          if(s1h>=0) {
-            if(s2h>=0) emit_cmp(s1h,s2h);
-            else emit_test(s1h,s1h);
-            emit_cmovne_reg(alt,addr);
-          }
-          if(s2l>=0) emit_cmp(s1l,s2l);
-          else emit_test(s1l,s1l);
-          emit_cmovne_reg(alt,addr);
-        }
-      }
-      if((opcode[i]&0x2f)==5) // BNE
-      {
-        #ifdef HAVE_CMOV_IMM
-        if(s1h<0) {
-          if(s2l>=0) emit_cmp(s1l,s2l);
-          else emit_test(s1l,s1l);
-          emit_cmov2imm_e_ne_compact(start+i*4+8,ba[i],addr);
-        }
-        else
-        #endif
-        {
-          emit_mov2imm_compact(start+i*4+8,addr,ba[i],alt);
-          if(s1h>=0) {
-            if(s2h>=0) emit_cmp(s1h,s2h);
-            else emit_test(s1h,s1h);
-            emit_cmovne_reg(alt,addr);
-          }
-          if(s2l>=0) emit_cmp(s1l,s2l);
-          else emit_test(s1l,s1l);
-          emit_cmovne_reg(alt,addr);
-        }
-      }
-      if((opcode[i]&0x2f)==6) // BLEZ
-      {
-        //emit_movimm(ba[i],alt);
-        //emit_movimm(start+i*4+8,addr);
-        emit_mov2imm_compact(ba[i],alt,start+i*4+8,addr);
-        emit_cmpimm(s1l,1);
-        if(s1h>=0) emit_mov(addr,ntaddr);
-        emit_cmovl_reg(alt,addr);
-        if(s1h>=0) {
-          emit_test(s1h,s1h);
-          emit_cmovne_reg(ntaddr,addr);
-          emit_cmovs_reg(alt,addr);
-        }
-      }
-      if((opcode[i]&0x2f)==7) // BGTZ
-      {
-        //emit_movimm(ba[i],addr);
-        //emit_movimm(start+i*4+8,ntaddr);
-        emit_mov2imm_compact(ba[i],addr,start+i*4+8,ntaddr);
-        emit_cmpimm(s1l,1);
-        if(s1h>=0) emit_mov(addr,alt);
-        emit_cmovl_reg(ntaddr,addr);
-        if(s1h>=0) {
-          emit_test(s1h,s1h);
-          emit_cmovne_reg(alt,addr);
-          emit_cmovs_reg(ntaddr,addr);
-        }
-      }
-      if((opcode[i]==1)&&(opcode2[i]&0x2D)==0) // BLTZ
-      {
-        //emit_movimm(ba[i],alt);
-        //emit_movimm(start+i*4+8,addr);
-        emit_mov2imm_compact(ba[i],alt,start+i*4+8,addr);
-        if(s1h>=0) emit_test(s1h,s1h);
-        else emit_test(s1l,s1l);
-        emit_cmovs_reg(alt,addr);
-      }
-      if((opcode[i]==1)&&(opcode2[i]&0x2D)==1) // BGEZ
-      {
-        //emit_movimm(ba[i],addr);
-        //emit_movimm(start+i*4+8,alt);
-        emit_mov2imm_compact(ba[i],addr,start+i*4+8,alt);
-        if(s1h>=0) emit_test(s1h,s1h);
-        else emit_test(s1l,s1l);
-        emit_cmovs_reg(alt,addr);
-      }
-      if(opcode[i]==0x11 && opcode2[i]==0x08 ) {
-        if(source[i]&0x10000) // BC1T
-        {
-          //emit_movimm(ba[i],alt);
-          //emit_movimm(start+i*4+8,addr);
-          emit_mov2imm_compact(ba[i],alt,start+i*4+8,addr);
-          emit_testimm(s1l,0x800000);
-          emit_cmovne_reg(alt,addr);
-        }
-        else // BC1F
-        {
-          //emit_movimm(ba[i],addr);
-          //emit_movimm(start+i*4+8,alt);
-          emit_mov2imm_compact(ba[i],addr,start+i*4+8,alt);
-          emit_testimm(s1l,0x800000);
-          emit_cmovne_reg(alt,addr);
-        }
-      }
-      emit_writeword(addr,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.pcaddr);
-    }
-    else
-    if(itype[i]==RJUMP)
-    {
-      int r=get_reg(branch_regs[i].regmap,rs1[i]);
-      if((rs1[i]==rt1[i+1]||rs1[i]==rt2[i+1])&&(rs1[i]!=0)) {
-        r=get_reg(branch_regs[i].regmap,RTEMP);
-      }
-      emit_writeword(r,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.pcaddr);
-    }
-    else {DebugMessage(M64MSG_ERROR, "Unknown branch type in do_ccstub");exit(1);}
-  }
-  // Update cycle count
-  assert(branch_regs[i].regmap[HOST_CCREG]==CCREG||branch_regs[i].regmap[HOST_CCREG]==-1);
-  if(stubs[n][3]) emit_addimm(HOST_CCREG,CLOCK_DIVIDER*stubs[n][3],HOST_CCREG);
-  emit_call((intptr_t)cc_interrupt);
-  if(stubs[n][3]) emit_addimm(HOST_CCREG,-(int)CLOCK_DIVIDER*stubs[n][3],HOST_CCREG);
-  if(stubs[n][6]==TAKEN) {
-    if(internal_branch(branch_regs[i].is32,ba[i]))
-      load_needed_regs(branch_regs[i].regmap,regs[(ba[i]-start)>>2].regmap_entry);
-    else if(itype[i]==RJUMP) {
-      int r=get_reg(branch_regs[i].regmap,rs1[i]);
-      if((rs1[i]==rt1[i+1]||rs1[i]==rt2[i+1])&&(rs1[i]!=0)) {
-        r=get_reg(branch_regs[i].regmap,RTEMP);
-      }
-#if NEW_DYNAREC==NEW_DYNAREC_ARM64
-      if(r==18) {
-        // x18 is used for trampoline jumps, move it to another register (x0)
-        emit_mov(r,0);
-        r=0;
-        stubs[n][2]=jump_vaddr_reg[0];
-      }
-#endif
-      emit_readword((intptr_t)&g_dev.r4300.new_dynarec_hot_state.pcaddr,r);
-    }
-  }else if(stubs[n][6]==NOTTAKEN) {
-    if(i<slen-2) load_needed_regs(branch_regs[i].regmap,regmap_pre[i+2]);
-    else load_all_regs(branch_regs[i].regmap);
-  }else if(stubs[n][6]==NULLDS) {
-    // Delay slot instruction is nullified ("likely" branch)
-    if(i<slen-2) load_needed_regs(regs[i].regmap,regmap_pre[i+2]);
-    else load_all_regs(regs[i].regmap);
-  }else{
-    load_all_regs(branch_regs[i].regmap);
-  }
-  emit_jmp(stubs[n][2]); // return address
-  
-  /* This works but uses a lot of memory...
-  emit_readword((intptr_t)&g_dev.r4300.cp0.next_interrupt,ECX);
-  emit_add(HOST_CCREG,ECX,EAX);
-  emit_writeword(EAX,(intptr_t)&r4300_cp0_regs(&g_dev.r4300.cp0)[CP0_COUNT_REG]);
-  emit_call((intptr_t)dynarec_gen_interrupt);
-  emit_readword((intptr_t)&r4300_cp0_regs(&g_dev.r4300.cp0)[CP0_COUNT_REG],HOST_CCREG);
-  emit_readword((intptr_t)&g_dev.r4300.cp0.next_interrupt,EAX);
-  emit_readword((intptr_t)&g_dev.r4300.new_dynarec_hot_state.pending_exception,EBX);
-  emit_sub(HOST_CCREG,EAX,HOST_CCREG);
-  emit_test(EBX,EBX);
-  intptr_t jne_instr=(intptr_t)out;
-  emit_jne(0);
-  if(stubs[n][3]) emit_addimm(HOST_CCREG,-2*stubs[n][3],HOST_CCREG);
-  load_all_regs(branch_regs[i].regmap);
-  emit_jmp(stubs[n][2]); // return address
-  set_jump_target(jne_instr,(intptr_t)out);
-  emit_readword((intptr_t)&g_dev.r4300.new_dynarec_hot_state.pcaddr,EAX);
-  // Call get_addr_ht instead of doing the hash table here.
-  // This code is executed infrequently and takes up a lot of space
-  // so smaller is better.
-  emit_storereg(CCREG,HOST_CCREG);
-  emit_pushreg(EAX);
-  emit_call((intptr_t)get_addr_ht);
-  emit_loadreg(CCREG,HOST_CCREG);
-  emit_addimm(ESP,4,ESP);
-  emit_jmpreg(EAX);*/
-}
-
 static void ujump_assemble(int i,struct regstat *i_regs)
 {
   #ifdef REG_PREFETCH
@@ -6212,10 +7072,10 @@ static void ujump_assemble(int i,struct regstat *i_regs)
 
   #ifdef REG_PREFETCH
   int temp=get_reg(branch_regs[i].regmap,PTEMP);
-  if(rt1[i]==31&&temp>=0) 
+  if(rt1[i]==31&&temp>=0)
   {
     int return_address=start+i*4+8;
-    if(get_reg(branch_regs[i].regmap,31)>0) 
+    if(get_reg(branch_regs[i].regmap,31)>0)
     if(i_regmap[temp]==PTEMP) emit_movimm((intptr_t)hash_table[((return_address>>16)^return_address)&0xFFFF],temp);
   }
   #endif
@@ -6232,7 +7092,6 @@ static void ujump_assemble(int i,struct regstat *i_regs)
     unsigned int return_address;
     assert(rt2[i+1]!=31);
     rt=get_reg(branch_regs[i].regmap,31);
-    assem_debug("branch(%d): eax=%d ecx=%d edx=%d ebx=%d ebp=%d esi=%d edi=%d",i,branch_regs[i].regmap[0],branch_regs[i].regmap[1],branch_regs[i].regmap[2],branch_regs[i].regmap[3],branch_regs[i].regmap[5],branch_regs[i].regmap[6],branch_regs[i].regmap[7]);
     //assert(rt>=0);
     return_address=start+i*4+8;
     if(rt>=0) {
@@ -6248,7 +7107,7 @@ static void ujump_assemble(int i,struct regstat *i_regs)
       #endif
       {
         #ifdef REG_PREFETCH
-        if(temp>=0) 
+        if(temp>=0)
         {
           if(i_regmap[temp]!=PTEMP) emit_movimm((intptr_t)hash_table[((return_address>>16)^return_address)&0xFFFF],temp);
         }
@@ -6304,7 +7163,7 @@ static void rjump_assemble(int i,struct regstat *i_regs)
   }
   address_generation(i+1,i_regs,regs[i].regmap_entry);
   #ifdef REG_PREFETCH
-  if(rt1[i]==31) 
+  if(rt1[i]==31)
   {
     if((temp=get_reg(branch_regs[i].regmap,PTEMP))>=0) {
       int return_address=start+i*4+8;
@@ -6332,11 +7191,10 @@ static void rjump_assemble(int i,struct regstat *i_regs)
     assert(rt1[i+1]!=rt1[i]);
     assert(rt2[i+1]!=rt1[i]);
     rt=get_reg(branch_regs[i].regmap,rt1[i]);
-    assem_debug("branch(%d): eax=%d ecx=%d edx=%d ebx=%d ebp=%d esi=%d edi=%d",i,branch_regs[i].regmap[0],branch_regs[i].regmap[1],branch_regs[i].regmap[2],branch_regs[i].regmap[3],branch_regs[i].regmap[5],branch_regs[i].regmap[6],branch_regs[i].regmap[7]);
     assert(rt>=0);
     return_address=start+i*4+8;
     #ifdef REG_PREFETCH
-    if(temp>=0) 
+    if(temp>=0)
     {
       if(i_regmap[temp]!=PTEMP) emit_movimm((intptr_t)hash_table[((return_address>>16)^return_address)&0xFFFF],temp);
     }
@@ -6419,7 +7277,7 @@ static void cjump_assemble(int i,struct regstat *i_regs)
   #ifdef CORTEX_A8_BRANCH_PREDICTION_HACK
   if(i>(ba[i]-start)>>2) invert=1;
   #endif
-  
+
   if(ooo[i]) {
     s1l=get_reg(branch_regs[i].regmap,rs1[i]);
     s1h=get_reg(branch_regs[i].regmap,rs1[i]|64);
@@ -6474,7 +7332,7 @@ static void cjump_assemble(int i,struct regstat *i_regs)
     load_regs(regs[i].regmap,branch_regs[i].regmap,regs[i].was32,CCREG,CCREG);
     cc=get_reg(branch_regs[i].regmap,CCREG);
     assert(cc==HOST_CCREG);
-    if(unconditional) 
+    if(unconditional)
       store_regs_bt(branch_regs[i].regmap,branch_regs[i].is32,branch_regs[i].dirty,ba[i]);
     //do_cc(i,branch_regs[i].regmap,&adj,unconditional?ba[i]:-1,unconditional);
     //assem_debug("cycle count (adj)");
@@ -6546,8 +7404,7 @@ static void cjump_assemble(int i,struct regstat *i_regs)
           emit_jne(0);
         }
       } // if(!only32)
-          
-      //DebugMessage(M64MSG_VERBOSE, "branch(%d): eax=%d ecx=%d edx=%d ebx=%d ebp=%d esi=%d edi=%d",i,branch_regs[i].regmap[0],branch_regs[i].regmap[1],branch_regs[i].regmap[2],branch_regs[i].regmap[3],branch_regs[i].regmap[5],branch_regs[i].regmap[6],branch_regs[i].regmap[7]);
+
       assert(s1l>=0);
       if(opcode[i]==4) // BEQ
       {
@@ -6583,7 +7440,7 @@ static void cjump_assemble(int i,struct regstat *i_regs)
         }else{
           add_to_linker((intptr_t)out,ba[i],branch_internal);
           if(only32) emit_jl(0);
-          else emit_jb(0); 
+          else emit_jb(0);
         }
       }
       if(opcode[i]==7) // BGTZ
@@ -6680,8 +7537,7 @@ static void cjump_assemble(int i,struct regstat *i_regs)
           emit_jne(1);
         }
       } // if(!only32)
-          
-      //DebugMessage(M64MSG_VERBOSE, "branch(%d): eax=%d ecx=%d edx=%d ebx=%d ebp=%d esi=%d edi=%d",i,branch_regs[i].regmap[0],branch_regs[i].regmap[1],branch_regs[i].regmap[2],branch_regs[i].regmap[3],branch_regs[i].regmap[5],branch_regs[i].regmap[6],branch_regs[i].regmap[7]);
+
       assert(s1l>=0);
       if((opcode[i]&0x2f)==4) // BEQ
       {
@@ -6813,7 +7669,7 @@ static void sjump_assemble(int i,struct regstat *i_regs)
 
   //if(opcode2[i]>=0x10) return; // FIXME (BxxZAL)
   assert(opcode2[i]<0x10||rs1[i]==0); // FIXME (BxxZAL)
-  
+
   if(ooo[i]) {
     s1l=get_reg(branch_regs[i].regmap,rs1[i]);
     s1h=get_reg(branch_regs[i].regmap,rs1[i]|64);
@@ -6857,7 +7713,6 @@ static void sjump_assemble(int i,struct regstat *i_regs)
       assert(rt1[i+1]!=31);
       assert(rt2[i+1]!=31);
       rt=get_reg(branch_regs[i].regmap,31);
-      assem_debug("branch(%d): eax=%d ecx=%d edx=%d ebx=%d ebp=%d esi=%d edi=%d",i,branch_regs[i].regmap[0],branch_regs[i].regmap[1],branch_regs[i].regmap[2],branch_regs[i].regmap[3],branch_regs[i].regmap[5],branch_regs[i].regmap[6],branch_regs[i].regmap[7]);
       if(rt>=0) {
         // Save the PC even if the branch is not taken
         return_address=start+i*4+8;
@@ -6869,7 +7724,7 @@ static void sjump_assemble(int i,struct regstat *i_regs)
     }
     cc=get_reg(branch_regs[i].regmap,CCREG);
     assert(cc==HOST_CCREG);
-    if(unconditional) 
+    if(unconditional)
       store_regs_bt(branch_regs[i].regmap,branch_regs[i].is32,branch_regs[i].dirty,ba[i]);
     //do_cc(i,branch_regs[i].regmap,&adj,unconditional?ba[i]:-1,unconditional);
     assem_debug("cycle count (adj)");
@@ -6956,7 +7811,7 @@ static void sjump_assemble(int i,struct regstat *i_regs)
           }
         }
       } // if(!only32)
-          
+
       if(invert) {
         #ifdef CORTEX_A8_BRANCH_PREDICTION_HACK
         if(match&&(!branch_internal||!is_ds[(ba[i]-start)>>2])) {
@@ -7000,7 +7855,6 @@ static void sjump_assemble(int i,struct regstat *i_regs)
     //DebugMessage(M64MSG_VERBOSE, "IOE");
     intptr_t nottaken=0;
     if(!unconditional) {
-      //DebugMessage(M64MSG_VERBOSE, "branch(%d): eax=%d ecx=%d edx=%d ebx=%d ebp=%d esi=%d edi=%d",i,branch_regs[i].regmap[0],branch_regs[i].regmap[1],branch_regs[i].regmap[2],branch_regs[i].regmap[3],branch_regs[i].regmap[5],branch_regs[i].regmap[6],branch_regs[i].regmap[7]);
       if(!only32)
       {
         assert(s1h>=0);
@@ -7194,7 +8048,7 @@ static void fjump_assemble(int i,struct regstat *i_regs)
         {
         }
       } // if(!only32)
-          
+
       if(invert) {
         if(adj) emit_addimm(cc,-(int)CLOCK_DIVIDER*adj,cc);
         #ifdef CORTEX_A8_BRANCH_PREDICTION_HACK
@@ -7227,8 +8081,7 @@ static void fjump_assemble(int i,struct regstat *i_regs)
     //DebugMessage(M64MSG_VERBOSE, "IOE");
     intptr_t nottaken=0;
     if(1) {
-      //DebugMessage(M64MSG_VERBOSE, "branch(%d): eax=%d ecx=%d edx=%d ebx=%d ebp=%d esi=%d edi=%d",i,branch_regs[i].regmap[0],branch_regs[i].regmap[1],branch_regs[i].regmap[2],branch_regs[i].regmap[3],branch_regs[i].regmap[5],branch_regs[i].regmap[6],branch_regs[i].regmap[7]);
-      if(1) {
+       if(1) {
         assert(fs>=0);
         emit_testimm(fs,0x800000);
         if(source[i]&0x10000) // BC1T
@@ -7766,528 +8619,11 @@ static void pagespan_ds(void)
   load_regs_bt(regs[0].regmap,regs[0].is32,regs[0].dirty,start+4);
 }
 
-static void do_cop1stub(int n)
-{
-  literal_pool(256);
-  assem_debug("do_cop1stub %x",start+stubs[n][3]*4);
-  set_jump_target(stubs[n][1],(intptr_t)out);
-  int i=stubs[n][3];
-  struct regstat *i_regs=(struct regstat *)stubs[n][5];
-  int ds=stubs[n][6];
-  if(!ds) {
-    load_all_consts(regs[i].regmap_entry,regs[i].was32,regs[i].wasdirty,regs[i].wasconst,i);
-    //if(i_regs!=&regs[i]) DebugMessage(M64MSG_VERBOSE, "oops: regs[i]=%x i_regs=%x",(int)&regs[i],(int)i_regs);
-  }
-  wb_dirtys(i_regs->regmap_entry,i_regs->was32,i_regs->wasdirty);
-
-  if(get_reg(i_regs->regmap,CCREG)<0) {
-    emit_loadreg(CCREG,HOST_CCREG);
-  }
-
-  emit_movimm(start+(i*4)+ds,0); // Get PC
-  emit_addimm(HOST_CCREG,CLOCK_DIVIDER*ccadj[i],HOST_CCREG); // CHECK: is this right?  There should probably be an extra cycle...
-  emit_jmp((intptr_t)fp_exception);
-}
-
-static void do_readstub(int n)
-{
-  assem_debug("do_readstub %x",start+stubs[n][3]*4);
-  literal_pool(256);
-  set_jump_target(stubs[n][1],(intptr_t)out);
-  int type=stubs[n][0];
-  int i=stubs[n][3];
-  int addr=stubs[n][4];
-  struct regstat *i_regs=(struct regstat *)stubs[n][5];
-  u_int reglist=stubs[n][7];
-  signed char *i_regmap=i_regs->regmap;
-  int rth,rt;
-
-  if(itype[i]==C1LS) {
-    rth=get_reg(i_regmap,FTEMP|64);
-    rt=get_reg(i_regmap,FTEMP);
-  }else{
-    rth=get_reg(i_regmap,rt1[i]|64);
-    rt=get_reg(i_regmap,rt1[i]);
-  }
-
-  assert(addr>=0);
-  emit_writeword(addr,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.address);
-
-  intptr_t ftable=0;
-  if(type==LOADB_STUB||type==LOADBU_STUB)
-    ftable=(intptr_t)read_byte_new;
-  else if(type==LOADH_STUB||type==LOADHU_STUB)
-    ftable=(intptr_t)read_hword_new;
-  else if(type==LOADW_STUB||type==LOADWU_STUB)
-    ftable=(intptr_t)read_word_new;
-  else if(type==LOADD_STUB)
-    ftable=(intptr_t)read_dword_new;
-  else if(type==LOADWL_STUB)
-  {
-    assert(rt>=0);
-    ftable=(intptr_t)LWL_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
-  }
-  else if(type==LOADWR_STUB)
-  {
-    assert(rt>=0);
-    ftable=(intptr_t)LWR_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
-  }
-  else if(type==LOADDL_STUB)
-  {
-    assert(rt>=0);
-    assert(rth>=0);
-    ftable=(intptr_t)LDL_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
-    emit_writeword(rth,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
-  }
-  else if(type==LOADDR_STUB)
-  {
-    assert(rt>=0);
-    assert(rth>=0);
-    ftable=(intptr_t)LDR_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
-    emit_writeword(rth,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
-  }
-
-  int cc=get_reg(i_regmap,CCREG);
-  if(cc>=0) {
-    emit_storereg(CCREG,cc);
-  }
-
-  save_regs(reglist);
-  int ds=i_regs!=&regs[i];
-
-#if NEW_DYNAREC == NEW_DYNAREC_X86
-  emit_pushimm(CLOCK_DIVIDER*(stubs[n][6]+1));
-  emit_pushimm((start+(i+1)*4)+ds);
-  emit_call((intptr_t)ftable);
-  emit_addimm(ESP,8,ESP);
-#else
-  emit_movimm((start+(i+1)*4)+ds,ARG1_REG);
-  emit_movimm(CLOCK_DIVIDER*(stubs[n][6]+1),ARG2_REG);
-  emit_call((intptr_t)ftable);
-#endif
-
-  restore_regs(reglist);
-
-  emit_cmpmem_imm((intptr_t)&g_dev.r4300.new_dynarec_hot_state.pending_exception,0);
-  intptr_t jaddr=(intptr_t)out;
-  emit_jeq(0);
-  
-  if(!ds) load_all_consts(regs[i].regmap_entry,regs[i].was32,regs[i].wasdirty,regs[i].wasconst,i);
-  wb_dirtys(i_regs->regmap_entry,i_regs->was32,i_regs->wasdirty);
-
-  emit_jmp((intptr_t)&do_interrupt);
-  set_jump_target(jaddr,(intptr_t)out);
-
-  if(rt>=0) {
-    if(type==LOADB_STUB)
-      emit_movsbl((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
-    else if(type==LOADBU_STUB)
-      emit_movzbl((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
-    else if(type==LOADH_STUB)
-      emit_movswl((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
-    else if(type==LOADHU_STUB)
-      emit_movzwl((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
-    else if(type==LOADW_STUB||type==LOADWU_STUB||type==LOADWL_STUB||type==LOADWR_STUB) {
-      emit_readword((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
-      if(type==LOADWU_STUB) emit_zeroreg(rth);
-    }
-    else if(type==LOADD_STUB||type==LOADDL_STUB||type==LOADDR_STUB) {
-      emit_readword((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
-      if(rth>=0) emit_readword((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword+4,rth);
-    }
-  }
-  emit_jmp(stubs[n][2]); // return address
-}
-
-static void inline_readstub(int type, int i, u_int addr_const, char addr, struct regstat *i_regs, int target, int adj, u_int reglist)
-{
-  assem_debug("inline_readstub");
-  int rth=get_reg(i_regs->regmap,target|64);
-  int rt=get_reg(i_regs->regmap,target);
-
-#if NEW_DYNAREC <= NEW_DYNAREC_X64
-  if(addr_const)
-    emit_writeword_imm(addr_const,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.address);
-  else
-#endif
-  {
-    assert(addr>=0);
-    emit_writeword(addr,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.address);
-  }
-
-  intptr_t ftable=0;
-  if(type==LOADB_STUB||type==LOADBU_STUB)
-    ftable=(intptr_t)read_byte_new;
-  else if(type==LOADH_STUB||type==LOADHU_STUB)
-    ftable=(intptr_t)read_hword_new;
-  else if(type==LOADW_STUB||type==LOADWU_STUB)
-    ftable=(intptr_t)read_word_new;
-  else if(type==LOADD_STUB)
-    ftable=(intptr_t)read_dword_new;
-  else if(type==LOADWL_STUB)
-  {
-    assert(rt>=0);
-    ftable=(intptr_t)LWL_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
-  }
-  else if(type==LOADWR_STUB)
-  {
-    assert(rt>=0);
-    ftable=(intptr_t)LWR_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
-  }
-  else if(type==LOADDL_STUB)
-  {
-    assert(rt>=0);
-    assert(rth>=0);
-    ftable=(intptr_t)LDL_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
-    emit_writeword(rth,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
-  }
-  else if(type==LOADDR_STUB)
-  {
-    assert(rt>=0);
-    assert(rth>=0);
-    ftable=(intptr_t)LDR_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
-    emit_writeword(rth,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
-  }
-
-  int cc=get_reg(i_regs->regmap,CCREG);
-  if(cc>=0) {
-    emit_storereg(CCREG,cc);
-  }
-
-  save_regs(reglist);
-  int ds=i_regs!=&regs[i];
-
-#if NEW_DYNAREC == NEW_DYNAREC_X86
-  emit_pushimm(CLOCK_DIVIDER*(adj+1));
-  emit_pushimm((start+(i+1)*4)+ds);
-  emit_call((intptr_t)ftable);
-  emit_addimm(ESP,8,ESP);
-#else
-  emit_movimm((start+(i+1)*4)+ds,ARG1_REG);
-  emit_movimm(CLOCK_DIVIDER*(adj+1),ARG2_REG);
-  emit_call((intptr_t)ftable);
-#endif
-
-  restore_regs(reglist);
-
-  if((signed int)addr_const>=(signed int)0xC0000000) {
-    // Theoretically we can have a pagefault here, if the TLB has never
-    // been enabled and the address is outside the range 80000000..BFFFFFFF
-    // Write out the registers so the pagefault can be handled.  This is
-    // a very rare case and likely represents a bug.
-    emit_cmpmem_imm((intptr_t)&g_dev.r4300.new_dynarec_hot_state.pending_exception,0);
-    intptr_t jaddr=(intptr_t)out;
-    emit_jeq(0);
-
-    if(!ds) load_all_consts(regs[i].regmap_entry,regs[i].was32,regs[i].wasdirty,regs[i].wasconst,i);
-    wb_dirtys(i_regs->regmap_entry,i_regs->was32,i_regs->wasdirty);
-
-    emit_jmp((intptr_t)&do_interrupt);
-    set_jump_target(jaddr,(intptr_t)out);
-  }
-
-  if(rt>=0) {
-    if(type==LOADB_STUB)
-      emit_movsbl((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
-    else if(type==LOADBU_STUB)
-      emit_movzbl((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
-    else if(type==LOADH_STUB)
-      emit_movswl((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
-    else if(type==LOADHU_STUB)
-      emit_movzwl((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
-    else if(type==LOADW_STUB||type==LOADWU_STUB||type==LOADWL_STUB||type==LOADWR_STUB) {
-      emit_readword((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
-      if(type==LOADWU_STUB) emit_zeroreg(rth);
-    }
-    else if(type==LOADD_STUB||type==LOADDL_STUB||type==LOADDR_STUB) {
-      emit_readword((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword,rt);
-      if(rth>=0) emit_readword((intptr_t)&g_dev.r4300.new_dynarec_hot_state.rdword+4,rth);
-    }
-  }
-}
-
-static void do_writestub(int n)
-{
-  assem_debug("do_writestub %x",start+stubs[n][3]*4);
-  literal_pool(256);
-  set_jump_target(stubs[n][1],(intptr_t)out);
-  int type=stubs[n][0];
-  int i=stubs[n][3];
-  int addr=stubs[n][4];
-  struct regstat *i_regs=(struct regstat *)stubs[n][5];
-  u_int reglist=stubs[n][7];
-  signed char *i_regmap=i_regs->regmap;
-  int rth,rt,r;
-
-  if(itype[i]==C1LS) {
-    rth=get_reg(i_regmap,FTEMP|64);
-    rt=get_reg(i_regmap,r=FTEMP);
-  }else{
-    rth=get_reg(i_regmap,rs2[i]|64);
-    rt=get_reg(i_regmap,r=rs2[i]);
-  }
-  assert(addr>=0);
-  assert(rt>=0);
-  emit_writeword(addr,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.address);
-
-  intptr_t ftable=0;
-  if(type==STOREB_STUB){
-    ftable=(intptr_t)write_byte_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
-  }
-  else if(type==STOREH_STUB){
-    ftable=(intptr_t)write_hword_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
-  }
-  else if(type==STOREW_STUB){
-    ftable=(intptr_t)write_word_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
-  }
-  else if(type==STORED_STUB){
-    ftable=(intptr_t)write_dword_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
-    emit_writeword(r?rth:rt,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
-  }
-  else if(type==STOREWL_STUB){
-    ftable=(intptr_t)SWL_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
-  }
-  else if(type==STOREWR_STUB){
-    ftable=(intptr_t)SWR_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
-  }
-  else if(type==STOREDL_STUB){
-    ftable=(intptr_t)SDL_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
-    emit_writeword(r?rth:rt,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
-  }
-  else if(type==STOREDR_STUB){
-    ftable=(intptr_t)SDR_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
-    emit_writeword(r?rth:rt,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
-  }
-
-  int cc=get_reg(i_regmap,CCREG);
-  if(cc>=0) {
-    emit_storereg(CCREG,cc);
-  }
-
-  save_regs(reglist);
-  int ds=i_regs!=&regs[i];
-
-#if NEW_DYNAREC == NEW_DYNAREC_X86
-  emit_pushimm(CLOCK_DIVIDER*(stubs[n][6]+1));
-  emit_pushimm((start+(i+1)*4)+ds);
-  emit_call((intptr_t)ftable);
-  emit_addimm(ESP,8,ESP);
-#else
-  emit_movimm((start+(i+1)*4)+ds,ARG1_REG);
-  emit_movimm(CLOCK_DIVIDER*(stubs[n][6]+1),ARG2_REG);
-  emit_call((intptr_t)ftable);
-#endif
-
-  restore_regs(reglist);
-
-  emit_cmpmem_imm((intptr_t)&g_dev.r4300.new_dynarec_hot_state.pending_exception,0);
-  intptr_t jaddr=(intptr_t)out;
-  emit_jeq(0);
-
-  if(!ds) load_all_consts(regs[i].regmap_entry,regs[i].was32,regs[i].wasdirty,regs[i].wasconst,i);
-  wb_dirtys(i_regs->regmap_entry,i_regs->was32,i_regs->wasdirty);
-
-  emit_jmp((intptr_t)&do_interrupt);
-  set_jump_target(jaddr,(intptr_t)out);
-
-  if(cc>=0) {
-    emit_loadreg(CCREG,cc);
-  }
-  emit_jmp(stubs[n][2]); // return address
-}
-
-static void inline_writestub(int type, int i, u_int addr_const, char addr, struct regstat *i_regs, int target, int adj, u_int reglist)
-{
-  assem_debug("inline_writestub");
-  int rth=get_reg(i_regs->regmap,target|64);
-  int rt=get_reg(i_regs->regmap,target);
-  assert(rt>=0);
-
-#if NEW_DYNAREC <= NEW_DYNAREC_X64
-  if(addr_const)
-    emit_writeword_imm(addr_const,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.address);
-  else
-#endif
-  {
-    assert(addr>=0);
-    emit_writeword(addr,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.address);
-  }
-
-  intptr_t ftable=0;
-  if(type==STOREB_STUB){
-    ftable=(intptr_t)write_byte_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
-  }
-  else if(type==STOREH_STUB){
-    ftable=(intptr_t)write_hword_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
-  }
-  else if(type==STOREW_STUB){
-    ftable=(intptr_t)write_word_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
-  }
-  else if(type==STORED_STUB){
-    ftable=(intptr_t)write_dword_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
-    emit_writeword(target?rth:rt,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
-  }
-  else if(type==STOREWL_STUB){
-    ftable=(intptr_t)SWL_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
-  }
-  else if(type==STOREWR_STUB){
-    ftable=(intptr_t)SWR_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wword);
-  }
-  else if(type==STOREDL_STUB){
-    ftable=(intptr_t)SDL_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
-    emit_writeword(target?rth:rt,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
-  }
-  else if(type==STOREDR_STUB){
-    ftable=(intptr_t)SDR_new;
-    emit_writeword(rt,(intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword);
-    emit_writeword(target?rth:rt,((intptr_t)&g_dev.r4300.new_dynarec_hot_state.wdword)+4);
-  }
-
-  int cc=get_reg(i_regs->regmap,CCREG);
-  if(cc>=0) {
-    emit_storereg(CCREG,cc);
-  }
-
-  save_regs(reglist);
-  int ds=i_regs!=&regs[i];
-
-#if NEW_DYNAREC == NEW_DYNAREC_X86
-  emit_pushimm(CLOCK_DIVIDER*(adj+1));
-  emit_pushimm((start+(i+1)*4)+ds);
-  emit_call((intptr_t)ftable);
-  emit_addimm(ESP,8,ESP);
-#else
-  emit_movimm((start+(i+1)*4)+ds,ARG1_REG);
-  emit_movimm(CLOCK_DIVIDER*(adj+1),ARG2_REG);
-  emit_call((intptr_t)ftable);
-#endif
-
-  restore_regs(reglist);
-
-  if(((signed int)addr_const>=(signed int)0xC0000000)||((addr_const>>16)==0xa430)||((addr_const>>16)==0x8430)) {
-    // Theoretically we can have a pagefault here, if the TLB has never
-    // been enabled and the address is outside the range 80000000..BFFFFFFF
-    // Write out the registers so the pagefault can be handled.  This is
-    // a very rare case and likely represents a bug.
-    emit_cmpmem_imm((intptr_t)&g_dev.r4300.new_dynarec_hot_state.pending_exception,0);
-    intptr_t jaddr=(intptr_t)out;
-    emit_jeq(0);
-
-    if(!ds) load_all_consts(regs[i].regmap_entry,regs[i].was32,regs[i].wasdirty,regs[i].wasconst,i);
-    wb_dirtys(i_regs->regmap_entry,i_regs->was32,i_regs->wasdirty);
-
-    emit_jmp((intptr_t)&do_interrupt);
-    set_jump_target(jaddr,(intptr_t)out);
-  }
-
-  if(cc>=0) {
-    emit_loadreg(CCREG,cc);
-  }
-}
-
-/* disassembly */
-static void disassemble_inst(int i)
-{
-    if (bt[i]) DebugMessage(M64MSG_VERBOSE, "*"); else DebugMessage(M64MSG_VERBOSE, " ");
-    switch(itype[i]) {
-      case UJUMP:
-        printf (" %x: %s %8x",start+i*4,insn[i],ba[i]);break;
-      case CJUMP:
-        printf (" %x: %s r%d,r%d,%8x",start+i*4,insn[i],rs1[i],rs2[i],i?start+i*4+4+((signed int)((unsigned int)source[i]<<16)>>14):*ba);break;
-      case SJUMP:
-        printf (" %x: %s r%d,%8x",start+i*4,insn[i],rs1[i],start+i*4+4+((signed int)((unsigned int)source[i]<<16)>>14));break;
-      case FJUMP:
-        printf (" %x: %s %8x",start+i*4,insn[i],ba[i]);break;
-      case RJUMP:
-        if ((opcode2[i]&1)&&rt1[i]!=31)
-          printf (" %x: %s r%d,r%d",start+i*4,insn[i],rt1[i],rs1[i]);
-        else
-          printf (" %x: %s r%d",start+i*4,insn[i],rs1[i]);
-        break;
-      case SPAN:
-        printf (" %x: %s (pagespan) r%d,r%d,%8x",start+i*4,insn[i],rs1[i],rs2[i],ba[i]);break;
-      case IMM16:
-        if(opcode[i]==0xf) //LUI
-          printf (" %x: %s r%d,%4x0000",start+i*4,insn[i],rt1[i],imm[i]&0xffff);
-        else
-          printf (" %x: %s r%d,r%d,%d",start+i*4,insn[i],rt1[i],rs1[i],imm[i]);
-        break;
-      case LOAD:
-      case LOADLR:
-        printf (" %x: %s r%d,r%d+%x",start+i*4,insn[i],rt1[i],rs1[i],imm[i]);
-        break;
-      case STORE:
-      case STORELR:
-        printf (" %x: %s r%d,r%d+%x",start+i*4,insn[i],rs2[i],rs1[i],imm[i]);
-        break;
-      case ALU:
-      case SHIFT:
-        printf (" %x: %s r%d,r%d,r%d",start+i*4,insn[i],rt1[i],rs1[i],rs2[i]);
-        break;
-      case MULTDIV:
-        printf (" %x: %s r%d,r%d",start+i*4,insn[i],rs1[i],rs2[i]);
-        break;
-      case SHIFTIMM:
-        printf (" %x: %s r%d,r%d,%d",start+i*4,insn[i],rt1[i],rs1[i],imm[i]);
-        break;
-      case MOV:
-        if((opcode2[i]&0x1d)==0x10)
-          printf (" %x: %s r%d",start+i*4,insn[i],rt1[i]);
-        else if((opcode2[i]&0x1d)==0x11)
-          printf (" %x: %s r%d",start+i*4,insn[i],rs1[i]);
-        else
-          printf (" %x: %s",start+i*4,insn[i]);
-        break;
-      case COP0:
-        if(opcode2[i]==0)
-          printf (" %x: %s r%d,cpr0[%d]",start+i*4,insn[i],rt1[i],(source[i]>>11)&0x1f); // MFC0
-        else if(opcode2[i]==4)
-          printf (" %x: %s r%d,cpr0[%d]",start+i*4,insn[i],rs1[i],(source[i]>>11)&0x1f); // MTC0
-        else printf (" %x: %s",start+i*4,insn[i]);
-        break;
-      case COP1:
-        if(opcode2[i]<3)
-          printf (" %x: %s r%d,cpr1[%d]",start+i*4,insn[i],rt1[i],(source[i]>>11)&0x1f); // MFC1
-        else if(opcode2[i]>3)
-          printf (" %x: %s r%d,cpr1[%d]",start+i*4,insn[i],rs1[i],(source[i]>>11)&0x1f); // MTC1
-        else printf (" %x: %s",start+i*4,insn[i]);
-        break;
-      case C1LS:
-        printf (" %x: %s cpr1[%d],r%d+%x",start+i*4,insn[i],(source[i]>>16)&0x1f,rs1[i],imm[i]);
-        break;
-      default:
-        //printf (" %s %8x",insn[i],source[i]);
-        printf (" %x: %s",start+i*4,insn[i]);
-    }
-}
-
+/**** Recompiler ****/
 void new_dynarec_init(void)
 {
   DebugMessage(M64MSG_INFO, "Init new dynarec");
-  
+
 #if defined(RECOMPILER_DEBUG) && !defined(RECOMP_DBG)
   recomp_dbg_init();
 #endif
@@ -8295,7 +8631,7 @@ void new_dynarec_init(void)
 #if !defined(RECOMP_DBG)
 #if NEW_DYNAREC == NEW_DYNAREC_ARM64
 
-#define FIXED_CACHE_ADDR 1    // Put the dynarec cache at extra_memory address 
+#define FIXED_CACHE_ADDR 1    // Put the dynarec cache at extra_memory address
 #define DYNAMIC_CACHE_ADDR 2  // Put the dynarec cache at random address
 #define DOUBLE_CACHE_ADDR 3   // Put the dynarec cache at random address with RW address != RX address
 
@@ -8357,7 +8693,7 @@ void new_dynarec_init(void)
 
   if(base_addr==(void*)-1) DebugMessage(M64MSG_ERROR, "mmap() failed");
 
-  assert(((uintptr_t)g_dev.rdram.dram&7)==0); //8 bytes aligned 
+  assert(((uintptr_t)g_dev.rdram.dram&7)==0); //8 bytes aligned
   out=(u_char *)base_addr;
 
   g_dev.r4300.new_dynarec_hot_state.pc = &g_dev.r4300.new_dynarec_hot_state.fake_pc;
@@ -8484,7 +8820,7 @@ int new_recompile_block(int addr)
   unsigned int type,op,op2;
 
   //DebugMessage(M64MSG_VERBOSE, "addr = %x source = %x %x", addr,source,source[0]);
-  
+
   /* Pass 1 disassembly */
 
   for(i=0;!done;i++) {
@@ -9022,7 +9358,7 @@ int new_recompile_block(int addr)
   /* Pass 2 - Register dependencies and branch targets */
 
   unneeded_registers(0,slen-1,0);
-  
+
   /* Pass 3 - Register allocation */
 
   struct regstat current; // Current register allocations/status
@@ -9038,9 +9374,9 @@ int new_recompile_block(int addr)
   int ds=0;
   int cc=0;
   int hr;
-  
+
   provisional_32bit();
-  
+
   if((u_int)addr&1) {
     // First instruction is delay slot
     cc=-1;
@@ -9050,7 +9386,7 @@ int new_recompile_block(int addr)
     unneeded_reg_upper[0]=1;
     current.regmap[HOST_BTREG]=BTREG;
   }
-  
+
   for(i=0;i<slen;i++)
   {
     if(bt[i])
@@ -9090,12 +9426,12 @@ int new_recompile_block(int addr)
       uint64_t temp_is32=current.is32;
       for(j=i-1;j>=0;j--)
       {
-        if(ba[j]==start+i*4) 
+        if(ba[j]==start+i*4)
           temp_is32&=branch_regs[j].is32;
       }
       for(j=i;j<slen;j++)
       {
-        if(ba[j]==start+i*4) 
+        if(ba[j]==start+i*4)
           //temp_is32=1;
           temp_is32&=p32[j];
       }
@@ -9129,12 +9465,12 @@ int new_recompile_block(int addr)
       uint64_t temp_is32=current.is32;
       for(j=i-1;j>=0;j--)
       {
-        if(ba[j]==start+i*4+4) 
+        if(ba[j]==start+i*4+4)
           temp_is32&=branch_regs[j].is32;
       }
       for(j=i;j<slen;j++)
       {
-        if(ba[j]==start+i*4+4) 
+        if(ba[j]==start+i*4+4)
           //temp_is32=1;
           temp_is32&=p32[j];
       }
@@ -9152,7 +9488,7 @@ int new_recompile_block(int addr)
                 {
                   //DebugMessage(M64MSG_VERBOSE, "dump %d/r%d",hr,r);
                   current.regmap[hr]=-1;
-                  if(get_reg(current.regmap,r|64)>=0) 
+                  if(get_reg(current.regmap,r|64)>=0)
                     current.regmap[get_reg(current.regmap,r|64)]=-1;
                 }
               }
@@ -9166,12 +9502,12 @@ int new_recompile_block(int addr)
       uint64_t temp_is32=current.is32;
       for(j=i-1;j>=0;j--)
       {
-        if(ba[j]==start+i*4+8) 
+        if(ba[j]==start+i*4+8)
           temp_is32&=branch_regs[j].is32;
       }
       for(j=i;j<slen;j++)
       {
-        if(ba[j]==start+i*4+8) 
+        if(ba[j]==start+i*4+8)
           //temp_is32=1;
           temp_is32&=p32[j];
       }
@@ -9187,7 +9523,7 @@ int new_recompile_block(int addr)
               {
                 //DebugMessage(M64MSG_VERBOSE, "dump %d/r%d",hr,r);
                 current.regmap[hr]=-1;
-                if(get_reg(current.regmap,r|64)>=0) 
+                if(get_reg(current.regmap,r|64)>=0)
                   current.regmap[get_reg(current.regmap,r|64)]=-1;
               }
             }
@@ -9281,7 +9617,7 @@ int new_recompile_block(int addr)
           }
         } else {
           // First instruction expects CCREG to be allocated
-          if(i==0&&hr==HOST_CCREG) 
+          if(i==0&&hr==HOST_CCREG)
             regs[i].regmap_entry[hr]=CCREG;
           else
             regs[i].regmap_entry[hr]=-1;
@@ -9627,7 +9963,7 @@ int new_recompile_block(int addr)
           }
         }
       }
-      
+
       // Drop the upper half of registers that have become 32-bit
       current.uu|=current.is32&((1LL<<rt1[i])|(1LL<<rt2[i]));
       if(itype[i]!=UJUMP&&itype[i]!=CJUMP&&itype[i]!=SJUMP&&itype[i]!=RJUMP&&itype[i]!=FJUMP) {
@@ -9688,7 +10024,7 @@ int new_recompile_block(int addr)
           }
         } else {
           // Branches expect CCREG to be allocated at the target
-          if(regmap_pre[i][hr]==CCREG) 
+          if(regmap_pre[i][hr]==CCREG)
             regs[i].regmap_entry[hr]=CCREG;
           else
             regs[i].regmap_entry[hr]=-1;
@@ -10011,11 +10347,11 @@ int new_recompile_block(int addr)
     }
     if(current.regmap[HOST_BTREG]==BTREG) current.regmap[HOST_BTREG]=-1;
   }
-  
+
   /* Pass 4 - Cull unused host registers */
-  
+
   uint64_t nr=0LL;
-  
+
   for (i=slen-1;i>=0;i--)
   {
     int hr=0;
@@ -10158,7 +10494,7 @@ int new_recompile_block(int addr)
     }
     // Save it
     needed_reg[i]=nr;
-    
+
     // Deallocate unneeded registers
     for(hr=0;hr<HOST_REGS;hr++)
     {
@@ -10285,13 +10621,13 @@ int new_recompile_block(int addr)
       }
     }
   }
-  
+
   /* Pass 5 - Pre-allocate registers */
-  
+
   // If a register is allocated during a loop, try to allocate it for the
   // entire loop, if possible.  This avoids loading/storing registers
   // inside of the loop.
-  
+
   signed char f_regmap[HOST_REGS];
   clear_all_regs(f_regmap);
   for(i=0;i<slen-1;i++)
@@ -10345,10 +10681,10 @@ int new_recompile_block(int addr)
             }
           }
           if(ooo[i]) {
-            if(count_free_regs(regs[i].regmap)<=minimum_free_regs[i+1]) 
+            if(count_free_regs(regs[i].regmap)<=minimum_free_regs[i+1])
               f_regmap[hr]=branch_regs[i].regmap[hr];
           }else{
-            if(count_free_regs(branch_regs[i].regmap)<=minimum_free_regs[i+1]) 
+            if(count_free_regs(branch_regs[i].regmap)<=minimum_free_regs[i+1])
               f_regmap[hr]=branch_regs[i].regmap[hr];
           }
           // Avoid dirty->clean transition
@@ -10522,10 +10858,10 @@ int new_recompile_block(int addr)
                   if(branch_regs[j].regmap[hr]>=0)
                     break;
                   if(ooo[j]) {
-                    if(count_free_regs(regs[j].regmap)<=minimum_free_regs[j+1]) 
+                    if(count_free_regs(regs[j].regmap)<=minimum_free_regs[j+1])
                       break;
                   }else{
-                    if(count_free_regs(branch_regs[j].regmap)<=minimum_free_regs[j+1]) 
+                    if(count_free_regs(branch_regs[j].regmap)<=minimum_free_regs[j+1])
                       break;
                   }
                   if(get_reg(branch_regs[j].regmap,f_regmap[hr])>=0) {
@@ -10598,7 +10934,7 @@ int new_recompile_block(int addr)
             regs[k].isconst&=~(1<<HOST_CCREG);
             k++;
           }
-          regs[j].regmap_entry[HOST_CCREG]=CCREG;          
+          regs[j].regmap_entry[HOST_CCREG]=CCREG;
         }
         // Work backwards from the branch target
         if(j>i&&f_regmap[HOST_CCREG]==CCREG)
@@ -10640,7 +10976,7 @@ int new_recompile_block(int addr)
       }
     }
   }
-  
+
   // Cache memory offset or tlb map pointer if a register is available
   #ifndef HOST_IMM_ADDR32
   #ifndef RAM_OFFSET
@@ -10820,7 +11156,7 @@ int new_recompile_block(int addr)
     }
   }
   #endif
-  
+
   // This allocates registers (if possible) one instruction prior
   // to use, which can avoid a load-use penalty on certain CPUs.
   for(i=0;i<slen-1;i++)
@@ -10868,7 +11204,7 @@ int new_recompile_block(int addr)
               }
             }
           }
-          // Load source into target register 
+          // Load source into target register
           if(lt1[i+1]&&get_reg(regs[i+1].regmap,rs1[i+1])<0) {
             if((hr=get_reg(regs[i+1].regmap,rt1[i+1]))>=0)
             {
@@ -11009,16 +11345,16 @@ int new_recompile_block(int addr)
       }
     }
   }
-  
+
   /* Pass 6 - Optimize clean/dirty state */
   clean_registers(0,slen-1,1);
-  
+
   /* Pass 7 - Identify 32-bit registers */
-  
+
   provisional_r32();
 
   u_int r32=0;
-  
+
   for (i=slen-1;i>=0;i--)
   {
     int hr;
@@ -11114,7 +11450,7 @@ int new_recompile_block(int addr)
       if((regs[i].was32>>dep2[i])&1) r32|=1LL<<dep2[i];
     }
     requires_32bit[i]=r32;
-    
+
     // Dirty registers which are 32-bit, require 32-bit input
     // as they will be written as 32-bit values
     for(hr=0;hr<HOST_REGS;hr++)
@@ -11132,210 +11468,6 @@ int new_recompile_block(int addr)
   if(itype[slen-1]==SPAN) {
     bt[slen-1]=1; // Mark as a branch target so instruction can restart after exception
   }
-  
-/* Debug/disassembly */
-#if ASSEM_DEBUG
-  for(i=0;i<slen;i++)
-  {
-    DebugMessage(M64MSG_VERBOSE, "U:");
-    int r;
-    for(r=1;r<=CCREG;r++) {
-      if((unneeded_reg[i]>>r)&1) {
-        if(r==HIREG) DebugMessage(M64MSG_VERBOSE, " HI");
-        else if(r==LOREG) DebugMessage(M64MSG_VERBOSE, " LO");
-        else DebugMessage(M64MSG_VERBOSE, " r%d",r);
-      }
-    }
-    DebugMessage(M64MSG_VERBOSE, " UU:");
-    for(r=1;r<=CCREG;r++) {
-      if(((unneeded_reg_upper[i]&~unneeded_reg[i])>>r)&1) {
-        if(r==HIREG) DebugMessage(M64MSG_VERBOSE, " HI");
-        else if(r==LOREG) DebugMessage(M64MSG_VERBOSE, " LO");
-        else DebugMessage(M64MSG_VERBOSE, " r%d",r);
-      }
-    }
-    DebugMessage(M64MSG_VERBOSE, " 32:");
-    for(r=0;r<=CCREG;r++) {
-      //if(((is32[i]>>r)&(~unneeded_reg[i]>>r))&1) {
-      if((regs[i].was32>>r)&1) {
-        if(r==CCREG) DebugMessage(M64MSG_VERBOSE, " CC");
-        else if(r==HIREG) DebugMessage(M64MSG_VERBOSE, " HI");
-        else if(r==LOREG) DebugMessage(M64MSG_VERBOSE, " LO");
-        else DebugMessage(M64MSG_VERBOSE, " r%d",r);
-      }
-    }
-    #if NEW_DYNAREC == NEW_DYNAREC_X86
-    DebugMessage(M64MSG_VERBOSE, "pre: eax=%d ecx=%d edx=%d ebx=%d ebp=%d esi=%d edi=%d",regmap_pre[i][0],regmap_pre[i][1],regmap_pre[i][2],regmap_pre[i][3],regmap_pre[i][5],regmap_pre[i][6],regmap_pre[i][7]);
-    #endif
-    #if NEW_DYNAREC == NEW_DYNAREC_ARM
-    DebugMessage(M64MSG_VERBOSE, "pre: r0=%d r1=%d r2=%d r3=%d r4=%d r5=%d r6=%d r7=%d r8=%d r9=%d r10=%d r12=%d",regmap_pre[i][0],regmap_pre[i][1],regmap_pre[i][2],regmap_pre[i][3],regmap_pre[i][4],regmap_pre[i][5],regmap_pre[i][6],regmap_pre[i][7],regmap_pre[i][8],regmap_pre[i][9],regmap_pre[i][10],regmap_pre[i][12]);
-    #endif
-    DebugMessage(M64MSG_VERBOSE, "needs: ");
-    if(needed_reg[i]&1) DebugMessage(M64MSG_VERBOSE, "eax ");
-    if((needed_reg[i]>>1)&1) DebugMessage(M64MSG_VERBOSE, "ecx ");
-    if((needed_reg[i]>>2)&1) DebugMessage(M64MSG_VERBOSE, "edx ");
-    if((needed_reg[i]>>3)&1) DebugMessage(M64MSG_VERBOSE, "ebx ");
-    if((needed_reg[i]>>5)&1) DebugMessage(M64MSG_VERBOSE, "ebp ");
-    if((needed_reg[i]>>6)&1) DebugMessage(M64MSG_VERBOSE, "esi ");
-    if((needed_reg[i]>>7)&1) DebugMessage(M64MSG_VERBOSE, "edi ");
-    DebugMessage(M64MSG_VERBOSE, "r:");
-    for(r=0;r<=CCREG;r++) {
-      //if(((requires_32bit[i]>>r)&(~unneeded_reg[i]>>r))&1) {
-      if((requires_32bit[i]>>r)&1) {
-        if(r==CCREG) DebugMessage(M64MSG_VERBOSE, " CC");
-        else if(r==HIREG) DebugMessage(M64MSG_VERBOSE, " HI");
-        else if(r==LOREG) DebugMessage(M64MSG_VERBOSE, " LO");
-        else DebugMessage(M64MSG_VERBOSE, " r%d",r);
-      }
-    }
-    /*DebugMessage(M64MSG_VERBOSE, "pr:");
-    for(r=0;r<=CCREG;r++) {
-      //if(((requires_32bit[i]>>r)&(~unneeded_reg[i]>>r))&1) {
-      if((pr32[i]>>r)&1) {
-        if(r==CCREG) DebugMessage(M64MSG_VERBOSE, " CC");
-        else if(r==HIREG) DebugMessage(M64MSG_VERBOSE, " HI");
-        else if(r==LOREG) DebugMessage(M64MSG_VERBOSE, " LO");
-        else DebugMessage(M64MSG_VERBOSE, " r%d",r);
-      }
-    }
-    if(pr32[i]!=requires_32bit[i]) DebugMessage(M64MSG_ERROR, " OOPS");*/
-    #if NEW_DYNAREC == NEW_DYNAREC_X86
-    DebugMessage(M64MSG_VERBOSE, "entry: eax=%d ecx=%d edx=%d ebx=%d ebp=%d esi=%d edi=%d",regs[i].regmap_entry[0],regs[i].regmap_entry[1],regs[i].regmap_entry[2],regs[i].regmap_entry[3],regs[i].regmap_entry[5],regs[i].regmap_entry[6],regs[i].regmap_entry[7]);
-    DebugMessage(M64MSG_VERBOSE, "dirty: ");
-    if(regs[i].wasdirty&1) DebugMessage(M64MSG_VERBOSE, "eax ");
-    if((regs[i].wasdirty>>1)&1) DebugMessage(M64MSG_VERBOSE, "ecx ");
-    if((regs[i].wasdirty>>2)&1) DebugMessage(M64MSG_VERBOSE, "edx ");
-    if((regs[i].wasdirty>>3)&1) DebugMessage(M64MSG_VERBOSE, "ebx ");
-    if((regs[i].wasdirty>>5)&1) DebugMessage(M64MSG_VERBOSE, "ebp ");
-    if((regs[i].wasdirty>>6)&1) DebugMessage(M64MSG_VERBOSE, "esi ");
-    if((regs[i].wasdirty>>7)&1) DebugMessage(M64MSG_VERBOSE, "edi ");
-    #endif
-    #if NEW_DYNAREC == NEW_DYNAREC_ARM
-    DebugMessage(M64MSG_VERBOSE, "entry: r0=%d r1=%d r2=%d r3=%d r4=%d r5=%d r6=%d r7=%d r8=%d r9=%d r10=%d r12=%d",regs[i].regmap_entry[0],regs[i].regmap_entry[1],regs[i].regmap_entry[2],regs[i].regmap_entry[3],regs[i].regmap_entry[4],regs[i].regmap_entry[5],regs[i].regmap_entry[6],regs[i].regmap_entry[7],regs[i].regmap_entry[8],regs[i].regmap_entry[9],regs[i].regmap_entry[10],regs[i].regmap_entry[12]);
-    DebugMessage(M64MSG_VERBOSE, "dirty: ");
-    if(regs[i].wasdirty&1) DebugMessage(M64MSG_VERBOSE, "r0 ");
-    if((regs[i].wasdirty>>1)&1) DebugMessage(M64MSG_VERBOSE, "r1 ");
-    if((regs[i].wasdirty>>2)&1) DebugMessage(M64MSG_VERBOSE, "r2 ");
-    if((regs[i].wasdirty>>3)&1) DebugMessage(M64MSG_VERBOSE, "r3 ");
-    if((regs[i].wasdirty>>4)&1) DebugMessage(M64MSG_VERBOSE, "r4 ");
-    if((regs[i].wasdirty>>5)&1) DebugMessage(M64MSG_VERBOSE, "r5 ");
-    if((regs[i].wasdirty>>6)&1) DebugMessage(M64MSG_VERBOSE, "r6 ");
-    if((regs[i].wasdirty>>7)&1) DebugMessage(M64MSG_VERBOSE, "r7 ");
-    if((regs[i].wasdirty>>8)&1) DebugMessage(M64MSG_VERBOSE, "r8 ");
-    if((regs[i].wasdirty>>9)&1) DebugMessage(M64MSG_VERBOSE, "r9 ");
-    if((regs[i].wasdirty>>10)&1) DebugMessage(M64MSG_VERBOSE, "r10 ");
-    if((regs[i].wasdirty>>12)&1) DebugMessage(M64MSG_VERBOSE, "r12 ");
-    #endif
-    disassemble_inst(i);
-    //printf ("ccadj[%d] = %d",i,ccadj[i]);
-    #if NEW_DYNAREC == NEW_DYNAREC_X86
-    DebugMessage(M64MSG_VERBOSE, "eax=%d ecx=%d edx=%d ebx=%d ebp=%d esi=%d edi=%d dirty: ",regs[i].regmap[0],regs[i].regmap[1],regs[i].regmap[2],regs[i].regmap[3],regs[i].regmap[5],regs[i].regmap[6],regs[i].regmap[7]);
-    if(regs[i].dirty&1) DebugMessage(M64MSG_VERBOSE, "eax ");
-    if((regs[i].dirty>>1)&1) DebugMessage(M64MSG_VERBOSE, "ecx ");
-    if((regs[i].dirty>>2)&1) DebugMessage(M64MSG_VERBOSE, "edx ");
-    if((regs[i].dirty>>3)&1) DebugMessage(M64MSG_VERBOSE, "ebx ");
-    if((regs[i].dirty>>5)&1) DebugMessage(M64MSG_VERBOSE, "ebp ");
-    if((regs[i].dirty>>6)&1) DebugMessage(M64MSG_VERBOSE, "esi ");
-    if((regs[i].dirty>>7)&1) DebugMessage(M64MSG_VERBOSE, "edi ");
-    #endif
-    #if NEW_DYNAREC == NEW_DYNAREC_ARM
-    DebugMessage(M64MSG_VERBOSE, "r0=%d r1=%d r2=%d r3=%d r4=%d r5=%d r6=%d r7=%d r8=%d r9=%d r10=%d r12=%d dirty: ",regs[i].regmap[0],regs[i].regmap[1],regs[i].regmap[2],regs[i].regmap[3],regs[i].regmap[4],regs[i].regmap[5],regs[i].regmap[6],regs[i].regmap[7],regs[i].regmap[8],regs[i].regmap[9],regs[i].regmap[10],regs[i].regmap[12]);
-    if(regs[i].dirty&1) DebugMessage(M64MSG_VERBOSE, "r0 ");
-    if((regs[i].dirty>>1)&1) DebugMessage(M64MSG_VERBOSE, "r1 ");
-    if((regs[i].dirty>>2)&1) DebugMessage(M64MSG_VERBOSE, "r2 ");
-    if((regs[i].dirty>>3)&1) DebugMessage(M64MSG_VERBOSE, "r3 ");
-    if((regs[i].dirty>>4)&1) DebugMessage(M64MSG_VERBOSE, "r4 ");
-    if((regs[i].dirty>>5)&1) DebugMessage(M64MSG_VERBOSE, "r5 ");
-    if((regs[i].dirty>>6)&1) DebugMessage(M64MSG_VERBOSE, "r6 ");
-    if((regs[i].dirty>>7)&1) DebugMessage(M64MSG_VERBOSE, "r7 ");
-    if((regs[i].dirty>>8)&1) DebugMessage(M64MSG_VERBOSE, "r8 ");
-    if((regs[i].dirty>>9)&1) DebugMessage(M64MSG_VERBOSE, "r9 ");
-    if((regs[i].dirty>>10)&1) DebugMessage(M64MSG_VERBOSE, "r10 ");
-    if((regs[i].dirty>>12)&1) DebugMessage(M64MSG_VERBOSE, "r12 ");
-    #endif
-    if(regs[i].isconst) {
-      DebugMessage(M64MSG_VERBOSE, "constants: ");
-      #if NEW_DYNAREC == NEW_DYNAREC_X86
-      if(regs[i].isconst&1) DebugMessage(M64MSG_VERBOSE, "eax=%x ",(int)constmap[i][0]);
-      if((regs[i].isconst>>1)&1) DebugMessage(M64MSG_VERBOSE, "ecx=%x ",(int)constmap[i][1]);
-      if((regs[i].isconst>>2)&1) DebugMessage(M64MSG_VERBOSE, "edx=%x ",(int)constmap[i][2]);
-      if((regs[i].isconst>>3)&1) DebugMessage(M64MSG_VERBOSE, "ebx=%x ",(int)constmap[i][3]);
-      if((regs[i].isconst>>5)&1) DebugMessage(M64MSG_VERBOSE, "ebp=%x ",(int)constmap[i][5]);
-      if((regs[i].isconst>>6)&1) DebugMessage(M64MSG_VERBOSE, "esi=%x ",(int)constmap[i][6]);
-      if((regs[i].isconst>>7)&1) DebugMessage(M64MSG_VERBOSE, "edi=%x ",(int)constmap[i][7]);
-      #endif
-      #if NEW_DYNAREC == NEW_DYNAREC_ARM
-      if(regs[i].isconst&1) DebugMessage(M64MSG_VERBOSE, "r0=%x ",(int)constmap[i][0]);
-      if((regs[i].isconst>>1)&1) DebugMessage(M64MSG_VERBOSE, "r1=%x ",(int)constmap[i][1]);
-      if((regs[i].isconst>>2)&1) DebugMessage(M64MSG_VERBOSE, "r2=%x ",(int)constmap[i][2]);
-      if((regs[i].isconst>>3)&1) DebugMessage(M64MSG_VERBOSE, "r3=%x ",(int)constmap[i][3]);
-      if((regs[i].isconst>>4)&1) DebugMessage(M64MSG_VERBOSE, "r4=%x ",(int)constmap[i][4]);
-      if((regs[i].isconst>>5)&1) DebugMessage(M64MSG_VERBOSE, "r5=%x ",(int)constmap[i][5]);
-      if((regs[i].isconst>>6)&1) DebugMessage(M64MSG_VERBOSE, "r6=%x ",(int)constmap[i][6]);
-      if((regs[i].isconst>>7)&1) DebugMessage(M64MSG_VERBOSE, "r7=%x ",(int)constmap[i][7]);
-      if((regs[i].isconst>>8)&1) DebugMessage(M64MSG_VERBOSE, "r8=%x ",(int)constmap[i][8]);
-      if((regs[i].isconst>>9)&1) DebugMessage(M64MSG_VERBOSE, "r9=%x ",(int)constmap[i][9]);
-      if((regs[i].isconst>>10)&1) DebugMessage(M64MSG_VERBOSE, "r10=%x ",(int)constmap[i][10]);
-      if((regs[i].isconst>>12)&1) DebugMessage(M64MSG_VERBOSE, "r12=%x ",(int)constmap[i][12]);
-      #endif
-    }
-    DebugMessage(M64MSG_VERBOSE, " 32:");
-    for(r=0;r<=CCREG;r++) {
-      if((regs[i].is32>>r)&1) {
-        if(r==CCREG) DebugMessage(M64MSG_VERBOSE, " CC");
-        else if(r==HIREG) DebugMessage(M64MSG_VERBOSE, " HI");
-        else if(r==LOREG) DebugMessage(M64MSG_VERBOSE, " LO");
-        else DebugMessage(M64MSG_VERBOSE, " r%d",r);
-      }
-    }
-    /*DebugMessage(M64MSG_VERBOSE, " p32:");
-    for(r=0;r<=CCREG;r++) {
-      if((p32[i]>>r)&1) {
-        if(r==CCREG) DebugMessage(M64MSG_VERBOSE, " CC");
-        else if(r==HIREG) DebugMessage(M64MSG_VERBOSE, " HI");
-        else if(r==LOREG) DebugMessage(M64MSG_VERBOSE, " LO");
-        else DebugMessage(M64MSG_VERBOSE, " r%d",r);
-      }
-    }
-    if(p32[i]!=regs[i].is32) DebugMessage(M64MSG_VERBOSE, " NO MATCH");*/
-    if(itype[i]==RJUMP||itype[i]==UJUMP||itype[i]==CJUMP||itype[i]==SJUMP||itype[i]==FJUMP) {
-      #if NEW_DYNAREC == NEW_DYNAREC_X86
-      DebugMessage(M64MSG_VERBOSE, "branch(%d): eax=%d ecx=%d edx=%d ebx=%d ebp=%d esi=%d edi=%d dirty: ",i,branch_regs[i].regmap[0],branch_regs[i].regmap[1],branch_regs[i].regmap[2],branch_regs[i].regmap[3],branch_regs[i].regmap[5],branch_regs[i].regmap[6],branch_regs[i].regmap[7]);
-      if(branch_regs[i].dirty&1) DebugMessage(M64MSG_VERBOSE, "eax ");
-      if((branch_regs[i].dirty>>1)&1) DebugMessage(M64MSG_VERBOSE, "ecx ");
-      if((branch_regs[i].dirty>>2)&1) DebugMessage(M64MSG_VERBOSE, "edx ");
-      if((branch_regs[i].dirty>>3)&1) DebugMessage(M64MSG_VERBOSE, "ebx ");
-      if((branch_regs[i].dirty>>5)&1) DebugMessage(M64MSG_VERBOSE, "ebp ");
-      if((branch_regs[i].dirty>>6)&1) DebugMessage(M64MSG_VERBOSE, "esi ");
-      if((branch_regs[i].dirty>>7)&1) DebugMessage(M64MSG_VERBOSE, "edi ");
-      #endif
-      #if NEW_DYNAREC == NEW_DYNAREC_ARM
-      DebugMessage(M64MSG_VERBOSE, "branch(%d): r0=%d r1=%d r2=%d r3=%d r4=%d r5=%d r6=%d r7=%d r8=%d r9=%d r10=%d r12=%d dirty: ",i,branch_regs[i].regmap[0],branch_regs[i].regmap[1],branch_regs[i].regmap[2],branch_regs[i].regmap[3],branch_regs[i].regmap[4],branch_regs[i].regmap[5],branch_regs[i].regmap[6],branch_regs[i].regmap[7],branch_regs[i].regmap[8],branch_regs[i].regmap[9],branch_regs[i].regmap[10],branch_regs[i].regmap[12]);
-      if(branch_regs[i].dirty&1) DebugMessage(M64MSG_VERBOSE, "r0 ");
-      if((branch_regs[i].dirty>>1)&1) DebugMessage(M64MSG_VERBOSE, "r1 ");
-      if((branch_regs[i].dirty>>2)&1) DebugMessage(M64MSG_VERBOSE, "r2 ");
-      if((branch_regs[i].dirty>>3)&1) DebugMessage(M64MSG_VERBOSE, "r3 ");
-      if((branch_regs[i].dirty>>4)&1) DebugMessage(M64MSG_VERBOSE, "r4 ");
-      if((branch_regs[i].dirty>>5)&1) DebugMessage(M64MSG_VERBOSE, "r5 ");
-      if((branch_regs[i].dirty>>6)&1) DebugMessage(M64MSG_VERBOSE, "r6 ");
-      if((branch_regs[i].dirty>>7)&1) DebugMessage(M64MSG_VERBOSE, "r7 ");
-      if((branch_regs[i].dirty>>8)&1) DebugMessage(M64MSG_VERBOSE, "r8 ");
-      if((branch_regs[i].dirty>>9)&1) DebugMessage(M64MSG_VERBOSE, "r9 ");
-      if((branch_regs[i].dirty>>10)&1) DebugMessage(M64MSG_VERBOSE, "r10 ");
-      if((branch_regs[i].dirty>>12)&1) DebugMessage(M64MSG_VERBOSE, "r12 ");
-      #endif
-      DebugMessage(M64MSG_VERBOSE, " 32:");
-      for(r=0;r<=CCREG;r++) {
-        if((branch_regs[i].is32>>r)&1) {
-          if(r==CCREG) DebugMessage(M64MSG_VERBOSE, " CC");
-          else if(r==HIREG) DebugMessage(M64MSG_VERBOSE, " HI");
-          else if(r==LOREG) DebugMessage(M64MSG_VERBOSE, " LO");
-          else DebugMessage(M64MSG_VERBOSE, " r%d",r);
-        }
-      }
-    }
-  }
-#endif
 
   /* Pass 8 - Assembly */
   linkcount=0;stubcount=0;
@@ -11676,7 +11808,7 @@ int new_recompile_block(int addr)
   // start over from the beginning. (Is 256K enough?)
   if(out > (u_char *)((u_char *)base_addr+(1<<TARGET_SIZE_2)-MAX_OUTPUT_BLOCK_SIZE-JUMP_TABLE_SIZE))
     out=(u_char *)base_addr;
-  
+
   // Trap writes to any of the pages we compiled
   for(i=start>>12;i<=(int)((start+slen*4-4)>>12);i++) {
     g_dev.r4300.cached_interp.invalid_code[i]=0;
@@ -11690,9 +11822,9 @@ int new_recompile_block(int addr)
       //DebugMessage(M64MSG_VERBOSE, "write protect physical page: %x (virtual %x)",j<<12,start);
     }
   }
-  
+
   /* Pass 10 - Free memory by expiring oldest blocks */
-  
+
   int end=((((intptr_t)out-(intptr_t)base_addr)>>(TARGET_SIZE_2-16))+16384)&65535;
   while(expirep!=end)
   {
@@ -11733,7 +11865,7 @@ int new_recompile_block(int addr)
       case 3:
         // Clear jump_out
         #if NEW_DYNAREC >= NEW_DYNAREC_ARM
-        if((expirep&2047)==0) 
+        if((expirep&2047)==0)
           do_clear_cache();
         #endif
         ll_remove_matching_addrs(jump_out+(expirep&2047),base,shift);
@@ -11743,481 +11875,4 @@ int new_recompile_block(int addr)
     expirep=(expirep+1)&65535;
   }
   return 0;
-}
-
-/* interpreted opcode */
-void* cop1_unusable(void)
-{
-    struct r4300_core* r4300 = &g_dev.r4300;
-    struct new_dynarec_hot_state* state = &r4300->new_dynarec_hot_state;
-    r4300->delay_slot = state->pcaddr & 1;
-    state->pcaddr &= ~1;
-    state->cp0_regs[CP0_CAUSE_REG] = CP0_CAUSE_EXCCODE_CPU | CP0_CAUSE_CE1;
-    exception_general(r4300);
-    return get_addr_ht(state->pcaddr);
-}
-
-void* SYSCALL_new(void)
-{
-    struct r4300_core* r4300 = &g_dev.r4300;
-    struct new_dynarec_hot_state* state = &r4300->new_dynarec_hot_state;
-    r4300->delay_slot = 0;
-    cached_interp_SYSCALL();
-    return get_addr_ht(state->pcaddr);
-}
-
-void* ERET_new(void)
-{
-    struct r4300_core* r4300 = &g_dev.r4300;
-    struct new_dynarec_hot_state* state = &r4300->new_dynarec_hot_state;
-
-    cp0_update_count(r4300);
-    if (state->cp0_regs[CP0_STATUS_REG] & CP0_STATUS_ERL)
-    {
-        DebugMessage(M64MSG_ERROR, "error in ERET");
-        state->stop = 1;
-    }
-    else
-    {
-        state->cp0_regs[CP0_STATUS_REG] &= ~CP0_STATUS_EXL;
-        state->pcaddr = state->cp0_regs[CP0_EPC_REG];
-    }
-    r4300->llbit = 0;
-    r4300->delay_slot = 0;
-    r4300_check_interrupt(r4300, CP0_CAUSE_IP2, r4300->mi->regs[MI_INTR_REG] & r4300->mi->regs[MI_INTR_MASK_REG]); // ???
-    r4300->cp0.last_addr = state->pcaddr;
-    state->pending_exception = 0;
-    if (state->cycle_count >= 0) { gen_interrupt(r4300); }
-
-    if(state->stop)
-        return NULL;
-
-    if(state->pending_exception)
-        return get_addr_ht(state->pcaddr);
-    else
-    {
-        uint32_t is64 = 0;
-        for(int i=0;i<32;i++)
-            is64 = (((int)(state->regs[i]>>32)^((int)state->regs[i]>>31))!=0)<<i;
-        
-        is64 |= (((int)(state->hi>>32)^((int)state->hi>>31))!=0);
-        is64 |= (((int)(state->lo>>32)^((int)state->lo>>31))!=0);
-        return get_addr_32(state->pcaddr, is64);
-    }
-}
-
-void dynarec_gen_interrupt(void)
-{
-    struct r4300_core* r4300 = &g_dev.r4300;
-    struct new_dynarec_hot_state* state = &r4300->new_dynarec_hot_state;
-    cp0_update_count(r4300);
-    uint32_t page = ((state->cp0_regs[CP0_COUNT_REG]>>19)&0x1fc);
-    unsigned int *candidate = (unsigned int *)&restore_candidate[page];
-    page <<= 3;
-    r4300->delay_slot = 0;
-
-    if(*candidate)
-    {
-        for(int i=0;i<32;i++)
-        {
-            if((*candidate>>i)&1)
-                clean_blocks(page+i);
-        }
-        *candidate = 0;
-    }
-
-    gen_interrupt(r4300);
-}
-
-#define UPDATE_COUNT_IN \
-  struct r4300_core* r4300 = &g_dev.r4300; \
-  struct new_dynarec_hot_state* state = &r4300->new_dynarec_hot_state; \
-  state->cycle_count += count; \
-  state->pending_exception = 0;
-
-#define UPDATE_COUNT_OUT \
-  state->cycle_count -= !state->pending_exception * count;
-
-static void TLBWI_new(int pcaddr, int count)
-{
-  unsigned int i;
-  UPDATE_COUNT_IN
-  state->pcaddr = pcaddr;
-
-  /* Remove old entries */
-  unsigned int old_start_even=r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].start_even;
-  unsigned int old_end_even=r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].end_even;
-  unsigned int old_start_odd=r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].start_odd;
-  unsigned int old_end_odd=r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].end_odd;
-  for (i=old_start_even>>12; i<=old_end_even>>12; i++)
-  {
-    if(i<0x80000||i>0xBFFFF)
-    {
-      invalidate_block(i);
-      state->memory_map[i]=(uintptr_t)-1;
-    }
-  }
-  for (i=old_start_odd>>12; i<=old_end_odd>>12; i++)
-  {
-    if(i<0x80000||i>0xBFFFF)
-    {
-      invalidate_block(i);
-      state->memory_map[i]=(uintptr_t)-1;
-    }
-  }
-  cached_interp_TLBWI();
-  //DebugMessage(M64MSG_VERBOSE, "TLBWI: index=%d",state->cp0_regs[CP0_INDEX_REG]);
-  //DebugMessage(M64MSG_VERBOSE, "TLBWI: start_even=%x end_even=%x phys_even=%x v=%d d=%d",r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].start_even,r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].end_even,r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].phys_even,r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].v_even,r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].d_even);
-  //DebugMessage(M64MSG_VERBOSE, "TLBWI: start_odd=%x end_odd=%x phys_odd=%x v=%d d=%d",r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].start_odd,r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].end_odd,r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].phys_odd,r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].v_odd,r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].d_odd);
-  /* Combine r4300->cp0.tlb.LUT_r, r4300->cp0.tlb.LUT_w, and invalid_code into a single table
-     for fast look up. */
-  for (i=r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].start_even>>12; i<=r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].end_even>>12; i++)
-  {
-    //DebugMessage(M64MSG_VERBOSE, "%x: r:%8x w:%8x",i,r4300->cp0.tlb.LUT_r[i],r4300->cp0.tlb.LUT_w[i]);
-    if(i<0x80000||i>0xBFFFF)
-    {
-      if(r4300->cp0.tlb.LUT_r[i]) {
-        state->memory_map[i]=((uintptr_t)g_dev.rdram.dram+(uintptr_t)((r4300->cp0.tlb.LUT_r[i]&0xFFFFF000)-0x80000000)-(i<<12))>>2;
-        // FIXME: should make sure the physical page is invalid too
-        if(!r4300->cp0.tlb.LUT_w[i]||!r4300->cached_interp.invalid_code[i]) {
-          state->memory_map[i]|=WRITE_PROTECT; // Write protect
-        }else{
-          assert(r4300->cp0.tlb.LUT_r[i]==r4300->cp0.tlb.LUT_w[i]);
-        }
-        if(!using_tlb) DebugMessage(M64MSG_VERBOSE, "Enabled TLB");
-        // Tell the dynamic recompiler to generate tlb lookup code
-        using_tlb=1;
-      }
-      else state->memory_map[i]=(uintptr_t)-1;
-    }
-    //DebugMessage(M64MSG_VERBOSE, "memory_map[%x]: %8x (+%8x)",i,state->memory_map[i],state->memory_map[i]<<2);
-  }
-  for (i=r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].start_odd>>12; i<=r4300->cp0.tlb.entries[state->cp0_regs[CP0_INDEX_REG]&0x3F].end_odd>>12; i++)
-  {
-    //DebugMessage(M64MSG_VERBOSE, "%x: r:%8x w:%8x",i,r4300->cp0.tlb.LUT_r[i],r4300->cp0.tlb.LUT_w[i]);
-    if(i<0x80000||i>0xBFFFF)
-    {
-      if(r4300->cp0.tlb.LUT_r[i]) {
-        state->memory_map[i]=((uintptr_t)g_dev.rdram.dram+(uintptr_t)((r4300->cp0.tlb.LUT_r[i]&0xFFFFF000)-0x80000000)-(i<<12))>>2;
-        // FIXME: should make sure the physical page is invalid too
-        if(!r4300->cp0.tlb.LUT_w[i]||!r4300->cached_interp.invalid_code[i]) {
-          state->memory_map[i]|=WRITE_PROTECT; // Write protect
-        }else{
-          assert(r4300->cp0.tlb.LUT_r[i]==r4300->cp0.tlb.LUT_w[i]);
-        }
-        if(!using_tlb) DebugMessage(M64MSG_VERBOSE, "Enabled TLB");
-        // Tell the dynamic recompiler to generate tlb lookup code
-        using_tlb=1;
-      }
-      else state->memory_map[i]=(uintptr_t)-1;
-    }
-    //DebugMessage(M64MSG_VERBOSE, "memory_map[%x]: %8x (+%8x)",i,state->memory_map[i],state->memory_map[i]<<2);
-  }
-  UPDATE_COUNT_OUT
-}
-
-static void TLBWR_new(int pcaddr, int count)
-{
-  unsigned int i;
-  UPDATE_COUNT_IN
-  state->pcaddr = pcaddr;
-  cp0_update_count(r4300);
-  state->cp0_regs[CP0_RANDOM_REG] = (state->cp0_regs[CP0_COUNT_REG]/r4300->cp0.count_per_op % (32 - state->cp0_regs[CP0_WIRED_REG])) + state->cp0_regs[CP0_WIRED_REG];
-  /* Remove old entries */
-  unsigned int old_start_even=r4300->cp0.tlb.entries[state->cp0_regs[CP0_RANDOM_REG]&0x3F].start_even;
-  unsigned int old_end_even=r4300->cp0.tlb.entries[state->cp0_regs[CP0_RANDOM_REG]&0x3F].end_even;
-  unsigned int old_start_odd=r4300->cp0.tlb.entries[state->cp0_regs[CP0_RANDOM_REG]&0x3F].start_odd;
-  unsigned int old_end_odd=r4300->cp0.tlb.entries[state->cp0_regs[CP0_RANDOM_REG]&0x3F].end_odd;
-  for (i=old_start_even>>12; i<=old_end_even>>12; i++)
-  {
-    if(i<0x80000||i>0xBFFFF)
-    {
-      invalidate_block(i);
-      state->memory_map[i]=(uintptr_t)-1;
-    }
-  }
-  for (i=old_start_odd>>12; i<=old_end_odd>>12; i++)
-  {
-    if(i<0x80000||i>0xBFFFF)
-    {
-      invalidate_block(i);
-      state->memory_map[i]=(uintptr_t)-1;
-    }
-  }
-  cached_interp_TLBWR();
-  /* Combine r4300->cp0.tlb.LUT_r, r4300->cp0.tlb.LUT_w, and invalid_code into a single table
-     for fast look up. */
-  for (i=r4300->cp0.tlb.entries[state->cp0_regs[CP0_RANDOM_REG]&0x3F].start_even>>12; i<=r4300->cp0.tlb.entries[state->cp0_regs[CP0_RANDOM_REG]&0x3F].end_even>>12; i++)
-  {
-    //DebugMessage(M64MSG_VERBOSE, "%x: r:%8x w:%8x",i,r4300->cp0.tlb.LUT_r[i],r4300->cp0.tlb.LUT_w[i]);
-    if(i<0x80000||i>0xBFFFF)
-    {
-      if(r4300->cp0.tlb.LUT_r[i]) {
-        state->memory_map[i]=((uintptr_t)g_dev.rdram.dram+(uintptr_t)((r4300->cp0.tlb.LUT_r[i]&0xFFFFF000)-0x80000000)-(i<<12))>>2;
-        // FIXME: should make sure the physical page is invalid too
-        if(!r4300->cp0.tlb.LUT_w[i]||!r4300->cached_interp.invalid_code[i]) {
-          state->memory_map[i]|=WRITE_PROTECT; // Write protect
-        }else{
-          assert(r4300->cp0.tlb.LUT_r[i]==r4300->cp0.tlb.LUT_w[i]);
-        }
-        if(!using_tlb) DebugMessage(M64MSG_VERBOSE, "Enabled TLB");
-        // Tell the dynamic recompiler to generate tlb lookup code
-        using_tlb=1;
-      }
-      else state->memory_map[i]=(uintptr_t)-1;
-    }
-    //DebugMessage(M64MSG_VERBOSE, "memory_map[%x]: %8x (+%8x)",i,state->memory_map[i],state->memory_map[i]<<2);
-  }
-  for (i=r4300->cp0.tlb.entries[state->cp0_regs[CP0_RANDOM_REG]&0x3F].start_odd>>12; i<=r4300->cp0.tlb.entries[state->cp0_regs[CP0_RANDOM_REG]&0x3F].end_odd>>12; i++)
-  {
-    //DebugMessage(M64MSG_VERBOSE, "%x: r:%8x w:%8x",i,r4300->cp0.tlb.LUT_r[i],r4300->cp0.tlb.LUT_w[i]);
-    if(i<0x80000||i>0xBFFFF)
-    {
-      if(r4300->cp0.tlb.LUT_r[i]) {
-        state->memory_map[i]=((uintptr_t)g_dev.rdram.dram+(uintptr_t)((r4300->cp0.tlb.LUT_r[i]&0xFFFFF000)-0x80000000)-(i<<12))>>2;
-        // FIXME: should make sure the physical page is invalid too
-        if(!r4300->cp0.tlb.LUT_w[i]||!r4300->cached_interp.invalid_code[i]) {
-          state->memory_map[i]|=WRITE_PROTECT; // Write protect
-        }else{
-          assert(r4300->cp0.tlb.LUT_r[i]==r4300->cp0.tlb.LUT_w[i]);
-        }
-        if(!using_tlb) DebugMessage(M64MSG_VERBOSE, "Enabled TLB");
-        // Tell the dynamic recompiler to generate tlb lookup code
-        using_tlb=1;
-      }
-      else state->memory_map[i]=(uintptr_t)-1;
-    }
-    //DebugMessage(M64MSG_VERBOSE, "memory_map[%x]: %8x (+%8x)",i,state->memory_map[i],state->memory_map[i]<<2);
-  }
-  UPDATE_COUNT_OUT
-}
-
-static void MFC0_new(int copr, int count)
-{
-  UPDATE_COUNT_IN
-  state->fake_pc.f.r.nrd = copr;
-  cached_interp_MFC0();
-  UPDATE_COUNT_OUT
-}
-
-static void MTC0_new(int copr, int count, int pcaddr)
-{
-  UPDATE_COUNT_IN
-  state->pcaddr = pcaddr;
-  r4300->delay_slot = 0;
-  state->fake_pc.f.r.nrd = copr;
-  cached_interp_MTC0();
-  UPDATE_COUNT_OUT
-}
-
-#define BITS_BELOW_MASK32(x) ((UINT32_C(1) << (x)) - 1)
-#define BITS_ABOVE_MASK32(x) (~(BITS_BELOW_MASK32((x))))
-
-#define BITS_BELOW_MASK64(x) ((UINT64_C(1) << (x)) - 1)
-#define BITS_ABOVE_MASK64(x) (~(BITS_BELOW_MASK64((x))))
-
-static unsigned int bshift(uint32_t address)
-{
-    return ((address & 3) ^ 3) << 3;
-}
-
-static unsigned int hshift(uint32_t address)
-{
-    return ((address & 2) ^ 2) << 3;
-}
-
-static void read_byte_new(int pcaddr, int count)
-{
-  uint32_t value;
-  UPDATE_COUNT_IN
-  state->pcaddr = pcaddr&~1;
-  r4300->delay_slot = pcaddr & 1;
-  unsigned int shift = bshift(state->address);
-  if (r4300_read_aligned_word(r4300, state->address, &value)) {
-    state->rdword = (uint64_t)((value >> shift) & 0xff);
-  }
-  UPDATE_COUNT_OUT
-}
-
-static void read_hword_new(int pcaddr, int count)
-{
-  uint32_t value;
-  UPDATE_COUNT_IN
-  state->pcaddr = pcaddr&~1;
-  r4300->delay_slot = pcaddr & 1;
-  unsigned int shift = hshift(state->address);
-  if (r4300_read_aligned_word(r4300, state->address, &value)) {
-    state->rdword = (uint64_t)((value >> shift) & 0xffff);
-  }
-  UPDATE_COUNT_OUT
-}
-
-static void read_word_new(int pcaddr, int count)
-{
-  uint32_t value;
-  UPDATE_COUNT_IN
-  state->pcaddr = pcaddr&~1;
-  r4300->delay_slot = pcaddr & 1;
-  if (r4300_read_aligned_word(r4300, state->address, &value)) {
-    state->rdword = (uint64_t)(value);
-  }
-  UPDATE_COUNT_OUT
-}
-
-static void read_dword_new(int pcaddr, int count)
-{
-  UPDATE_COUNT_IN
-  state->pcaddr = pcaddr&~1;
-  r4300->delay_slot = pcaddr & 1;
-  r4300_read_aligned_dword(r4300, state->address, (uint64_t*)&state->rdword);
-  UPDATE_COUNT_OUT
-}
-
-static void write_byte_new(int pcaddr, int count)
-{
-  UPDATE_COUNT_IN
-  state->pcaddr = pcaddr&~1;
-  r4300->delay_slot = pcaddr & 1;
-  unsigned int shift = bshift(state->address);
-  state->wword <<= shift;
-  r4300_write_aligned_word(r4300, state->address, state->wword, UINT32_C(0xff) << shift);
-  UPDATE_COUNT_OUT
-}
-
-static void write_hword_new(int pcaddr, int count)
-{
-  UPDATE_COUNT_IN
-  state->pcaddr = pcaddr&~1;
-  r4300->delay_slot = pcaddr & 1;
-  unsigned int shift = hshift(state->address);
-  state->wword <<= shift;
-  r4300_write_aligned_word(r4300, state->address, state->wword, UINT32_C(0xffff) << shift);
-  UPDATE_COUNT_OUT
-}
-
-static void write_word_new(int pcaddr, int count)
-{
-  UPDATE_COUNT_IN
-  state->pcaddr = pcaddr&~1;
-  r4300->delay_slot = pcaddr & 1;
-  r4300_write_aligned_word(r4300, state->address, state->wword, UINT32_C(0xffffffff));
-  UPDATE_COUNT_OUT
-}
-
-static void write_dword_new(int pcaddr, int count)
-{
-  UPDATE_COUNT_IN
-  state->pcaddr = pcaddr&~1;
-  r4300->delay_slot = pcaddr & 1;
-  /* NOTE: in dynarec, we only need an all-one mask */
-  r4300_write_aligned_dword(r4300, state->address, state->wdword, ~UINT64_C(0));
-  UPDATE_COUNT_OUT
-}
-
-static void LWL_new(int pcaddr, int count)
-{
-  UPDATE_COUNT_IN
-  state->pcaddr = pcaddr&~1;
-  r4300->delay_slot = pcaddr & 1;
-  uint32_t value;
-  unsigned int n = (state->address & 3);
-  unsigned int shift = 8 * n;
-  uint32_t mask = BITS_BELOW_MASK32(8 * n);
-  if (r4300_read_aligned_word(r4300, state->address & ~UINT32_C(0x3), &value)) {
-    state->rdword = (uint64_t)((state->wword & mask) | (value << shift));
-  }
-  UPDATE_COUNT_OUT
-}
-
-static void LWR_new(int pcaddr, int count)
-{
-  UPDATE_COUNT_IN
-  state->pcaddr = pcaddr&~1;
-  r4300->delay_slot = pcaddr & 1;
-  uint32_t value;
-  unsigned int n = (state->address & 3);
-  unsigned int shift = 8 * (3 - n);
-  uint32_t mask = (n == 3) ? UINT32_C(0) : BITS_ABOVE_MASK32(8 * (n + 1));
-  if (r4300_read_aligned_word(r4300, state->address & ~UINT32_C(0x3), &value)) {
-    state->rdword = (uint64_t)((state->wword & mask) | (value >> shift));
-  }
-  UPDATE_COUNT_OUT
-}
-
-static void LDL_new(int pcaddr, int count)
-{
-  UPDATE_COUNT_IN
-  state->pcaddr = pcaddr&~1;
-  r4300->delay_slot = pcaddr & 1;
-  uint64_t value;
-  unsigned int n = (state->address & 7);
-  unsigned int shift = 8 * n;
-  uint64_t mask = BITS_BELOW_MASK64(8 * n);
-  if(r4300_read_aligned_dword(r4300, state->address & ~UINT32_C(7), &value)) {
-      state->rdword = (state->wdword & mask) | (value << shift);
-  }
-  UPDATE_COUNT_OUT
-}
-
-static void LDR_new(int pcaddr, int count)
-{
-  UPDATE_COUNT_IN
-  state->pcaddr = pcaddr&~1;
-  r4300->delay_slot = pcaddr & 1;
-  uint64_t value;
-  unsigned int n = (state->address & 7);
-  unsigned int shift = 8 * (7 - n);
-  uint64_t mask = (n == 7) ? UINT64_C(0) : BITS_ABOVE_MASK64(8 * (n + 1));
-  if(r4300_read_aligned_dword(r4300, state->address & ~UINT32_C(7), &value)){
-      state->rdword = (state->wdword & mask) | (value >> shift);
-  }
-  UPDATE_COUNT_OUT
-}
-
-static void SWL_new(int pcaddr, int count)
-{
-  UPDATE_COUNT_IN
-  state->pcaddr = pcaddr&~1;
-  r4300->delay_slot = pcaddr & 1;
-  unsigned int n = (state->address & 3);
-  unsigned int shift = 8 * n;
-  uint32_t mask = (n == 0) ? ~UINT32_C(0) : BITS_BELOW_MASK32(8 * (4 - n));
-  r4300_write_aligned_word(r4300, state->address & ~UINT32_C(0x3), state->wword >> shift, mask);
-  UPDATE_COUNT_OUT
-}
-
-static void SWR_new(int pcaddr, int count)
-{
-  UPDATE_COUNT_IN
-  state->pcaddr = pcaddr&~1;
-  r4300->delay_slot = pcaddr & 1;
-  unsigned int n = (state->address & 3);
-  unsigned int shift = 8 * (3 - n);
-  uint32_t mask = BITS_ABOVE_MASK32(8 * (3 - n));
-  r4300_write_aligned_word(r4300, state->address & ~UINT32_C(0x3), state->wword << shift, mask);
-  UPDATE_COUNT_OUT
-}
-
-static void SDL_new(int pcaddr, int count)
-{
-  UPDATE_COUNT_IN
-  state->pcaddr = pcaddr&~1;
-  r4300->delay_slot = pcaddr & 1;
-  unsigned int n = (state->address & 7);
-  unsigned int shift = 8 * n;
-  uint64_t mask = (n == 0) ? ~UINT64_C(0) : BITS_BELOW_MASK64(8 * (8 - n));
-  r4300_write_aligned_dword(r4300, state->address & ~UINT32_C(0x7), state->wdword >> shift, mask);
-  UPDATE_COUNT_OUT
-}
-
-static void SDR_new(int pcaddr, int count)
-{
-  UPDATE_COUNT_IN
-  state->pcaddr = pcaddr&~1;
-  r4300->delay_slot = pcaddr & 1;
-  unsigned int n = (state->address & 7);
-  unsigned int shift = 8 * (7 - n);
-  uint64_t mask = BITS_ABOVE_MASK64(8 * (7 - n));
-  r4300_write_aligned_dword(r4300, state->address & ~UINT32_C(0x7), state->wdword << shift, mask);
-  UPDATE_COUNT_OUT
 }
