+++ src/device/r4300/new_dynarec/x64/assem_x64.h
@@ -62,7 +62,7 @@ The registers RBX, RBP, RDI, RSI, RSP, R12, R13, R14, and R15 are considered non
 /* amd64 calling convention:
    func(rdi, rsi, rdx, rcx, r8, r9) {return rax;}
    callee-save: %rbp %rbx %r12-%r15
-   
+
 The registers RAX, RCX, RDX, RSI, RDI, R8, R9, R10, R11 are considered volatile (caller-saved).
 The registers RBX, RBP, RSP, R12, R13, R14, and R15 are considered nonvolatile (callee-saved).*/
