+++ src/device/r4300/new_dynarec/x64/linkage_x64.asm
@@ -1,20 +1,20 @@
-;Mupen64plus - linkage_x86.asm                                           
-;Copyright (C) 2009-2011 Ari64                                         
-;                                                                      
-;This program is free software; you can redistribute it and/or modify  
-;it under the terms of the GNU General Public License as published by  
-;the Free Software Foundation; either version 2 of the License, or     
-;(at your option) any later version.                                   
-;                                                                      
-;This program is distributed in the hope that it will be useful,       
-;but WITHOUT ANY WARRANTY; without even the implied warranty of        
-;MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         
-;GNU General Public License for more details.                          
-;                                                                      
-;You should have received a copy of the GNU General Public License     
-;along with this program; if not, write to the                         
-;Free Software Foundation, Inc.,                                       
-;51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.          
+;Mupen64plus - linkage_x86.asm
+;Copyright (C) 2009-2011 Ari64
+;
+;This program is free software; you can redistribute it and/or modify
+;it under the terms of the GNU General Public License as published by
+;the Free Software Foundation; either version 2 of the License, or
+;(at your option) any later version.
+;
+;This program is distributed in the hope that it will be useful,
+;but WITHOUT ANY WARRANTY; without even the implied warranty of
+;MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+;GNU General Public License for more details.
+;
+;You should have received a copy of the GNU General Public License
+;along with this program; if not, write to the
+;Free Software Foundation, Inc.,
+;51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
 %include "asm_defines_nasm.h"
 
@@ -266,7 +266,7 @@ new_dyna_start:
     mov     CCREG,    DWORD [rel g_dev_r4300_new_dynarec_hot_state_cycle_count]
     mov     rax,    QWORD[rel base_addr]
     jmp     rax
-    
+
 invalidate_block_eax:
     mov     ARG1_REG,    eax
     jmp     invalidate_block_call
